+++ exercices/5e/5G12-1.js
@@ -1,0 +1,11 @@
+import Pavages_et_transformations from '../6e/_Pavages_et_transformations.js'
+/**
+ * Exercice en html seulement. Symétrie centrale dans un pavage.
+ * @Auteur Jean-Claude Lhote
+ * référence 5G12-1
+ */
+export default function Pavages_et_demi_tour() {
+	Pavages_et_transformations.call(this);
+	this.titre ="Trouver l'image d'une figure par une symétrie centrale dans un pavage (sortie Latex Impossible)";
+	this.sup = 2;
+}
