+++ exercices/5e/5G20.js
@@ -1,0 +1,11 @@
+import Construire_un_triangle_avec_cible from '../6e/_Construire_un_triangle_avec_cible.js';
+/**
+ * @Auteur Jean-Claude Lhote
+ * référence 5G20
+ * 
+ */
+export default function Construire_un_triangle_avec_cible_5e() {
+	Construire_un_triangle_avec_cible.call(this);
+	this.titre="Construire un triangle avec cible auto-corrective";
+	this.classe = 5;
+}
