+++ include/mathalea_exercices.js
@@ -26,40 +26,10 @@ var liste_des_exercices_disponibles = {
   "c3C10-2" : Exercice_labyrinthe_multiplesCM,
   "c3C11": Division_cycle3,
   "c3N10": Ecrire_entiers_cycle3,
-  "6C10": Additions_soustractions_multiplications_posees,
-  "6C10-4": Exercice_tables_d_additions,
-  "6C13": Vocabulaire_et_operations,
-  "6C20": Additionner_soustraires_decimaux,
   "6C21": Divisions_euclidiennes_niv2,
-  "6C23": Exercice_additionner_des_fractions_6e,
-  "6C30-2": Produit_de_decimaux_a_partir_d_un_produit_connu,
-  "6C33": Priorites,
   //"6C99" : separation6C,
   //"6C99" : separation6C,
-  "6G10": Notation_segment_droite_demi_droite,
-  "6G10-2": Utiliser_le_codage_pour_decrire_6e,
-  "6G11": Tracer_des_perpendiculaires,
-  "6G12": Tracer_des_paralleles,
-  "6G12-1": Tracer_des_perpendiculaires_et_des_paralleles,
-  "6G14": Proprietes_paralleles_perpendiculaires,
   "beta6G24": Test_main_levee,
-  "6G20": Nommer_et_coder_des_polygones,
-  "6G20-2": Vocabulaire_des_triangles_6e,
-  "6G21": Construire_un_triangle_6e,
-  "6G21-1": Construire_un_triangle_avec_cible_6e,
-  "6G23-2": Tracer_triangle_2_angles,
-  "6G24": Transformations_6e,
-  "6G24-1": Symetrie_axiale_point_6e,
-  "6G24-2": Symetrie_axiale_figure_6e,
-  "6G24-3": Construire_symetrique_point_6e,
-  "6G25": Construire_mediatrices_6e,
-  "6G25-1": Pavages_et_reflexion,
-  "6G25-3" : Pavage_et_reflexion2d,
-  "6G32-1": Symetrie_axiale_conservation1,
-  "6G33": Reconnaitre_quadrilatere_particulier,
-  "6G41": Representer_un_solide_6e,
-  "6G42": Solide_6e,
-  "6G43": Utiliser_vocabulaire_pave,
   "6M11-1": Perimetre_ou_aire_de_carres_rectangles_triangles,
   "6M11-2": Perimetre_ou_aire_de_figures_composees,
   "6M10": Reglages_6M10,
@@ -77,7 +47,6 @@ var liste_des_exercices_disponibles = {
   "6N10-3": chiffre_nombre_de,
   "6N10-4": Ecrire_nombres_entiers_formates,
   "6N10-5" : Exercice_labyrinthe_numeration,
-  "6N11": Lire_abscisse_entiere,
   "6N11-3": Encadrer_un_entier_par_deux_entiers_consecutifs,
   "6N11-4": Ranger_ordre_croissant_decroissant,
   "6N12": Multiplier_entier_par_10_100_1000,
@@ -92,7 +61,6 @@ var liste_des_exercices_disponibles = {
   "6N23-4": Nombre_decimal_oralise_de_differentes_manieres,
   "6N24": Exercice_6N24,
   "6N24-1": Exercice_multiplier_ou_diviser_un_nombre_entier_par_10_100_1000,
-  "6N30-2": Placer_points_sur_axe,
   "6N31": Comparer_decimaux,
   "6N31-1": Encadrer_un_decimal_par_deux_entiers_consecutifs,
   "6N31-2": Ordre_de_grandeur_operations_decimaux,
@@ -114,18 +82,9 @@ var liste_des_exercices_disponibles = {
   "6P11-1": Proportionnalite_par_linearite_bis,
   "6S10": Lecture_diagramme_barre,
   "6S11": Organiser_donnees_depuis_texte,
-  "5A10": Liste_des_diviseurs_5e,
   "5A11": Tableau_criteres_de_divisibilite,
-  "5A11-1" : Exercice_labyrinthe_divisibilite,
-  "5A12-1": Premier_ou_pas_5e,
-  "5A13": Exercice_decomposer_en_facteurs_premiers,
   "5C11": Traduire_une_phrase_par_une_expression,
-  "5C11-1": Traduire_une_expression_par_une_phrase,
-  "5C11-2": Ecrire_une_expression_mathador,
   "5C12": Calculer_une_expression_numerique,
-  "5C12-1": Traduire_une_phrase_par_une_expression_et_calculer,
-  "5G12" : Pavage_et_demi_tour2d,
-  "5G12-1": Pavages_et_demi_tour,
   "5G11": Transformations_5e,
   "5G10-1": Symetrie_axiale_point_5e,
   "5G10-2": Symetrie_axiale_figure_5e,
@@ -224,173 +183,7 @@ for (var i in liste_des_exercices_disponibles) {
 
 
 
-/**
- * Lire l'abscisse entière d'un point
- * @Auteur Jean-Claude Lhote et Rémi Angot
- * référence : 6N11
- */
-function Lire_abscisse_entiere() {
-  "use strict";
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Lire l'abscisse entière d'un point (grands nombres)";
-  this.consigne = "Lire l'abscisse de chacun des points suivants.";
-  this.nb_questions = 3;
-  this.nb_questions_modifiable = true;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.spacing = 1;
-  this.spacing_corr = 1;
-  this.sup = 4;
-  this.liste_packages = ["tkz-euclide"];
 
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    // numero_de_l_exercice est 0 pour l'exercice 1
-    let type_de_questions;
-    this.liste_questions = [];
-    this.liste_corrections = [];
-    this.contenu = ""; // Liste de questions
-    this.contenu_correction = ""; // Liste de questions corrigées
-    if (this.sup == 4)
-      type_de_questions = combinaison_listes([1, 2, 3], this.nb_questions);
-    else
-      type_de_questions = combinaison_listes(
-        [parseInt(this.sup)],
-        this.nb_questions
-      );
-
-    this.contenu = html_consigne(this.consigne);
-    for (
-      let i = 0,
-      abs0,
-      l1,
-      l2,
-      l3,
-      x1,
-      x2,
-      x3,
-      x11,
-      x22,
-      x33,
-      pas1,
-      pas2,
-      id_unique,
-      texte,
-      texte_corr;
-      i < this.nb_questions;
-      i++
-    ) {
-      l1 = lettre_depuis_chiffre(i * 3 + 1);
-      l2 = lettre_depuis_chiffre(i * 3 + 2);
-      l3 = lettre_depuis_chiffre(i * 3 + 3);
-      switch (type_de_questions[i]) {
-        case 1: // Placer des entiers sur un axe (milliers)
-          abs0 = randint(1, 9) * 1000;
-          pas1 = 0.001;
-          pas2 = 10;
-          break;
-
-        case 2: // Placer des entiers sur un axe (dizaines de mille)
-          abs0 = randint(5, 15) * 10000;
-          pas1 = 0.0001;
-          pas2 = 10;
-          break;
-
-        case 3: // Placer des entiers sur un axe (centaines de mille)
-          abs0 = randint(35, 85) * 100000;
-          pas1 = 0.00001;
-          pas2 = 10;
-          break;
-      }
-      x1 = randint(0, 2);
-      x2 = randint(3, 4);
-      x3 = randint(5, 6);
-      x11 = randint(1, 9);
-      x22 = randint(1, 9);
-      x33 = randint(1, 3);
-      if (sortie_html) {
-        id_unique = `${i}_${Date.now()}`;
-        this.contenu += `<div id="div_svg${numero_de_l_exercice}${id_unique}" style="width: 90%; height: 200px;  "></div>`;
-        SVG_reperage_sur_un_axe(
-          `div_svg${numero_de_l_exercice}${id_unique}`,
-          abs0,
-          6,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11],
-            [l2, x2, x22],
-            [l3, x3, x33],
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-        this.contenu_correction += `<div id="div_svg_corr${numero_de_l_exercice}${id_unique}" style="width: 90%; height: 200px;  "></div>`;
-        SVG_reperage_sur_un_axe(
-          `div_svg_corr${numero_de_l_exercice}${id_unique}`,
-          abs0,
-          6,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11, true],
-            [l2, x2, x22, true],
-            [l3, x3, x33, true],
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-      } else {
-        //sortie Latex
-        texte = Latex_reperage_sur_un_axe(
-          2,
-          abs0,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11],
-            [l2, x2, x22],
-            [l3, x3, x33],
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-        texte_corr = Latex_reperage_sur_un_axe(
-          2,
-          abs0,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11, true],
-            [l2, x2, x22, true],
-            [l3, x3, x33, true],
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-      }
-    }
-    if (!sortie_html) liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = [
-    "Niveau de difficulté",
-    4,
-    "1 : Milliers\n2 : Dizaines de mille\n3 : Centaines de mille\n4 : Mélange",
-  ];
-}
 
 
 /**
@@ -529,318 +322,6 @@ function Questions_de_cours() {
     "URL du fichier texte contenant les questions et réponses.",
   ];
 }
-
-
-
-
-
-/**
- * Placer un point d'abscisse décimale
- * @Auteur Jean-Claude Lhote et Rémi Angot
- * référence : 6N30-2
- */
-function Placer_points_sur_axe() {
-  "use strict";
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Placer un point d'abscisse décimale";
-  this.consigne = " Placer trois points sur un axe gradué.";
-  this.nb_questions = 5;
-  this.nb_questions_modifiable = true;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.spacing = 1;
-  this.spacing_corr = 1;
-  this.sup = 1;
-  this.type_exercice = "SVGJS";
-  this.liste_packages = "tkz-euclide";
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    // numero_de_l_exercice est 0 pour l'exercice 1
-    let type_de_questions;
-    this.liste_questions = [];
-    this.liste_corrections = [];
-    this.contenu = ""; // Liste de questions
-    this.contenu_correction = ""; // Liste de questions corrigées
-    if (this.sup == 4)
-      type_de_questions = combinaison_listes([1, 2, 3], this.nb_questions);
-    else
-      type_de_questions = combinaison_listes(
-        [parseInt(this.sup)],
-        this.nb_questions
-      );
-
-    this.contenu = html_consigne(this.consigne);
-    for (
-      let i = 0,
-      abs0,
-      abs1,
-      abs2,
-      abs3,
-      l1,
-      l2,
-      l3,
-      x1,
-      x2,
-      x3,
-      x11,
-      x22,
-      x33,
-      pas1,
-      pas2,
-      id_unique,
-      texte,
-      texte_corr;
-      i < this.nb_questions;
-      i++
-    ) {
-      l1 = lettre_depuis_chiffre(i * 3 + 1);
-      l2 = lettre_depuis_chiffre(i * 3 + 2);
-      l3 = lettre_depuis_chiffre(i * 3 + 3);
-
-      switch (type_de_questions[i]) {
-        case 1: // Placer un point sur un axe (1 décimale)
-          abs0 = randint(0, 9);
-          pas1 = 1;
-          pas2 = 10;
-          break;
-
-        case 2: // Placer un point sur un axe (2 décimales)
-          abs0 = randint(0, 90) / 10;
-          pas1 = 10;
-          pas2 = 10;
-          break;
-
-        case 3: // Placer un point sur un axe (3 décimales)
-          abs0 = randint(0, 990) / 100;
-          pas1 = 100;
-          pas2 = 10;
-          break;
-      }
-      x1 = randint(0, 2);
-      x2 = randint(3, 4);
-      x3 = randint(5, 6);
-      x11 = randint(1, 9);
-      x22 = randint(1, 9);
-      x33 = randint(1, 3);
-      abs1 = arrondi(
-        abs0 + x1 / pas1 + x11 / pas1 / pas2,
-        type_de_questions[i]
-      ); // le type de questions est égal au nombre de décimales.
-      abs2 = arrondi(
-        abs0 + x2 / pas1 + x22 / pas1 / pas2,
-        type_de_questions[i]
-      );
-      abs3 = arrondi(
-        abs0 + x3 / pas1 + x33 / pas1 / pas2,
-        type_de_questions[i]
-      );
-
-      if (sortie_html) {
-        texte_corr = "";
-        this.contenu += `<h3>Placer les points : ${l1}(${tex_nombrec(
-          abs1
-        )}), ${l2}(${tex_nombrec(abs2)}), ${l3}(${tex_nombrec(abs3)})</h3>`;
-        id_unique = `${i}_${Date.now()}`;
-        this.contenu += `<div id="div_svg${numero_de_l_exercice}${id_unique}" style="width: 90%; height: 200px;  "></div>`;
-        SVG_reperage_sur_un_axe(
-          `div_svg${numero_de_l_exercice}${id_unique}`,
-          abs0,
-          6,
-          pas1,
-          pas2,
-          [],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-        this.contenu_correction += `<div id="div_svg_corr${numero_de_l_exercice}${id_unique}" style="width: 90%; height: 200px;  "></div>`;
-        SVG_reperage_sur_un_axe(
-          `div_svg_corr${numero_de_l_exercice}${id_unique}`,
-          abs0,
-          6,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11, true],
-            [l2, x2, x22, true],
-            [l3, x3, x33, true],
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-      } else {
-        //sortie Latex
-        texte = `{\\small Placer les points : $${l1}$(${tex_nombrec(
-          abs1
-        )}), $${l2}$(${tex_nombrec(abs2)}), $${l3}$(${tex_nombrec(abs3)})}<br>`;
-        texte += Latex_reperage_sur_un_axe(
-          2.4,
-          abs0,
-          pas1,
-          pas2,
-          [],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-        texte_corr = `{\\small Les points $${l1}$(${tex_nombrec(
-          abs1
-        )}), $${l2}$(${tex_nombrec(abs2)}), $${l3}$(${tex_nombrec(
-          abs3
-        )}) sont placés ci dessus}<br>`;
-        texte_corr += Latex_reperage_sur_un_axe(
-          2.4,
-          abs0,
-          pas1,
-          pas2,
-          [
-            [l1, x1, x11, true],
-            [l2, x2, x22, true],
-            [l3, x3, x33, true],
-          ],
-          [
-            [calcul(abs0, 0), 0, 0],
-            [calcul(abs0 + 1 / pas1, 0), 1, 0],
-          ],
-          false
-        );
-
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-      }
-    }
-    if (!sortie_html) liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = [
-    "Niveau de difficulté",
-    4,
-    "1 : Un chiffre après la virgule\n2 : Deux chiffres après la virgule \n3 : Trois chiffres après la virgule\n4 : Mélange",
-  ];
-}
-
-
-
-
-
-
-
-
-
-/**
- * Additions et/ou soustractions classique et/ou à trou.
- *
- * Par défaut c'est un mélange d'additions, soustractions avec et sans trou avec des nombres jusqu'à 20.
- * @Auteur Rémi Angot
- * Référence 6D12
- */
-function Tables_additions_soustractions() {
-  "use strict";
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.sup = 20;
-  this.sup2 = 6; // additions|additions à trous|soustractions|soustractions à trous|mélange sans trou|mélange avec trou
-  this.titre = "Additions et de soustractions";
-  this.consigne = "Calculer";
-  this.spacing = 2;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let liste_type_de_questions = [];
-    if (this.sup2 == 1) {
-      liste_type_de_questions = combinaison_listes(
-        ["addition"],
-        this.nb_questions
-      );
-    }
-    if (this.sup2 == 2) {
-      liste_type_de_questions = combinaison_listes(
-        ["addition_a_trou"],
-        this.nb_questions
-      );
-    }
-    if (this.sup2 == 3) {
-      liste_type_de_questions = combinaison_listes(
-        ["soustraction"],
-        this.nb_questions
-      );
-    }
-    if (this.sup2 == 4) {
-      liste_type_de_questions = combinaison_listes(
-        ["soustraction_a_trou"],
-        this.nb_questions
-      );
-    }
-    if (this.sup2 == 5) {
-      liste_type_de_questions = combinaison_listes(
-        ["addition", "soustraction"],
-        this.nb_questions
-      );
-    }
-    if (this.sup2 == 6) {
-      liste_type_de_questions = combinaison_listes(
-        ["addition", "addition_a_trou", "soustraction", "soustraction_a_trou"],
-        this.nb_questions
-      );
-    }
-    for (let i = 0, a, b, texte, texte_corr; i < this.nb_questions; i++) {
-      a = randint(2, this.sup);
-      b = randint(2, this.sup);
-
-      switch (liste_type_de_questions[i]) {
-        case "addition":
-          texte = `$${a} + ${b} = \\dotfill$`;
-          texte_corr = `$${a} + ${b} = ${a + b}$`;
-          break;
-        case "addition_a_trou":
-          texte = `$${a} + \\ldots\\ldots = ${a + b}$`;
-          texte_corr = `$${a} + ${mise_en_evidence(b)} = ${a + b}$`;
-          break;
-        case "soustraction":
-          if (a == b) {
-            a = randint(2, this.sup, b);
-          }
-          if (a < b) {
-            b = [a, (a = b)][0]; //échange les variables a et b
-          }
-          texte = `$${a} - ${b} = \\dotfill$`;
-          texte_corr = `$${a} - ${b} = ${a - b}$`;
-          break;
-        case "soustraction_a_trou":
-          if (a == b) {
-            a = randint(2, this.sup, b);
-          }
-          if (a < b) {
-            b = [a, (a = b)][0]; //échange les variables a et b
-          }
-          texte = `$${a} - \\ldots\\ldots = ${a - b}$`;
-          texte_corr = `$${a} - ${mise_en_evidence(b)} = ${a - b}$`;
-          break;
-      }
-
-      if (est_diaporama) {
-        texte = texte.replace("= \\dotfill", "");
-      }
-      this.liste_questions.push(texte);
-      this.liste_corrections.push(texte_corr);
-    }
-    liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = ["Valeur maximale", 9999]; // Texte, tooltip
-  this.besoin_formulaire2_numerique = [
-    "Style de questions",
-    6,
-    "1 : Additions\n2: Additions à trous\n3: Soustractions\n4 : Soustractions à trous\n5 : Additions et soustractions \n6 : Additions et soustractions avec ou sans trous",
-  ];
-}
-
-
 
 /**
  * Tables de divisions classiques, à trou ou un mélange des deux.
@@ -1501,95 +982,7 @@ function Somme_de_deux_nombres_maries_et_un_entier() {
   //this.besoin_formulaire_numerique = ['Niveau de difficulté',3];
 }
 
-/**
- * Mettre en relation un calcul, une traduction en français, une expression, un résultat, pour les décliner dans différents exercices.
- * Exercice sur le vocabulaire : somme,différence, produit, quotient...
- * @Auteur Jean-Claude Lhote
- * Référence 6C13
- */
-function Vocabulaire_et_operations() {
-  "use strict";
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Traduire des phrases en calculs et réciproquement";
-  this.consigne = "";
-  this.nb_questions = 5;
-  this.nb_cols = 2;
-  this.nb_cols_corr = 2;
-  this.sup = 4;
-  this.sup2 = false;
 
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let decimal;
-    let expf, expn, expc, resultats;
-    let type_de_questions_disponibles;
-    if (this.sup < 4) type_de_questions_disponibles = [parseInt(this.sup)];
-    else type_de_questions_disponibles = [1, 2, 3];
-    let liste_type_de_questions = combinaison_listes(
-      type_de_questions_disponibles,
-      this.nb_questions
-    );
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    if (this.sup2) decimal = 10 ** randint(1, 2);
-    else decimal = 1;
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      resultats = Choisir_expression_numerique(1, decimal);
-      expf = resultats[0];
-      expn = resultats[1];
-      expc = resultats[2];
-      texte = ``;
-      texte_corr = ``;
-      switch (liste_type_de_questions[i]) {
-        case 1:
-          texte +=
-            num_alpha(i) +
-            `Traduire la phrase par un calcul (il n’est pas demandé d’effectuer ce calcul) : `;
-          expf = `l` + expf.substring(1);
-          texte += `${expf}.`;
-          texte_corr += num_alpha(i) + `${expf} s'écrit ${expn}.`;
-          break;
-        case 2:
-          if (expn.indexOf("ou") > 0)
-            expn = expn.substring(0, expn.indexOf("ou")); // on supprime la deuxième expression fractionnaire
-          texte +=
-            num_alpha(i) + `Traduire le calcul par une phrase en français : `;
-          texte += `${expn}`;
-          expf = `l` + expf.substring(1);
-          texte_corr += num_alpha(i) + `${expn} est ${expf}.`;
-          break;
-        case 3:
-          texte +=
-            num_alpha(i) +
-            `Traduire la phrase par un calcul et effectuer ce calcul : `;
-          expf = `l` + expf.substring(1);
-          texte += `${expf}.`;
-          expf = `L` + expf.substring(1);
-          texte_corr += num_alpha(i) + `${expf} s'écrit ${expn}.<br>`;
-          texte_corr += `${expc}.`;
-          break;
-      }
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu_sans_numero(this);
-  };
-  this.besoin_formulaire_numerique = [
-    "Type de questions",
-    4,
-    "1 : Phrase -> Calcul\n 2 : Calcul -> Phrase\n 3 : Phrase -> Calcul + résultat\n 4 : Mélange",
-  ];
-  this.besoin_formulaire2_case_a_cocher = ["Décimaux", false];
-}
 
 /**
  * Un "Le compte est bon" avec des solutions "formatées" pour travailler certains incontournables du calcul mental
@@ -2997,139 +2390,6 @@ function Double_moitie_tiers_triple() {
   //this.besoin_formulaire_numerique = ['Niveau de difficulté',3];
 }
 
-
-
-
-/**
- * Additionner deux entiers
- * @Auteur Rémi Angot
- * Référence 6C10-4
- */
-function Exercice_tables_d_additions(max = 20) {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Addition de deux entiers";
-  this.consigne = "Calculer";
-  this.sup = max; // Le paramètre accessible à l'utilisateur sera la valeur maximale
-  this.spacing = 2;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    for (
-      let i = 0, a, b, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      a = randint(2, this.sup);
-      b = randint(2, this.sup);
-      texte = "$ " + tex_nombre(a) + " + " + tex_nombre(b) + " = \\dotfill $";
-      texte_corr =
-        "$ " +
-        tex_nombre(a) +
-        " + " +
-        tex_nombre(b) +
-        " = " +
-        tex_nombre(a + b) +
-        " $";
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = ["Valeur maximale", 99999];
-}
-
-/**
- * Simplifier une fraction, le facteur commun est inférieur à une valeur donnée en paramètre qui est 11 par défaut
- * @Auteur Rémi Angot
- *  5N13
- */
-function Exercice_fractions_simplifier(max = 11) {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.sup = max; // Correspond au facteur commun
-  this.titre = "Simplification de fractions";
-  this.consigne = "Simplifier les fractions suivantes.";
-  this.spacing = 2;
-  this.spacing_corr = 2;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    liste_fractions = [
-      [1, 2],
-      [1, 3],
-      [2, 3],
-      [1, 4],
-      [3, 4],
-      [1, 5],
-      [2, 5],
-      [3, 5],
-      [4, 5],
-      [1, 6],
-      [5, 6],
-      [1, 7],
-      [2, 7],
-      [3, 7],
-      [4, 7],
-      [5, 7],
-      [6, 7],
-      [1, 8],
-      [3, 8],
-      [5, 8],
-      [7, 8],
-      [1, 9],
-      [2, 9],
-      [4, 9],
-      [5, 9],
-      [7, 9],
-      [8, 9],
-      [1, 10],
-      [3, 10],
-      [7, 10],
-      [9, 10],
-    ]; // Couples de nombres premiers entre eux
-    for (
-      let i = 0, fraction, a, b, texte, texte_corr, cpt = 0;
-      i < this.nb_questions;
-      i++
-    ) {
-      fraction = choice(liste_fractions); //
-      a = fraction[0];
-      b = fraction[1];
-      k = randint(2, this.sup);
-      enleve_element(liste_fractions, fraction); // Il n'y aura pas 2 fois la même réponse
-      texte =
-        "$ " +
-        tex_fraction(k * a, k * b) +
-        " = " +
-        tex_fraction("\\phantom{00000000000000}", "") +
-        " = " +
-        tex_fraction("\\phantom{0000}", "") +
-        " $";
-      texte_corr =
-        "$ " +
-        tex_fraction(k * a, k * b) +
-        " = " +
-        tex_fraction(k + " \\times " + a, k + " \\times " + b) +
-        " = " +
-        tex_fraction(a, b) +
-        " $";
-      this.liste_questions.push(texte);
-      this.liste_corrections.push(texte_corr);
-    }
-    liste_de_question_to_contenu(this); //Espacement de 2 em entre chaque questions.
-  };
-  this.besoin_formulaire_numerique = [
-    "Valeur maximale du facteur commun",
-    99999,
-  ];
-}
-
 /**
  * Écrire une fraction avec un nouveau dénominateur qui est un multiple de son dénominateur (ce multiple est inférieur à une valeur maximale de 11 par défaut)
  * @Auteur Rémi Angot
@@ -4309,9 +3569,6 @@ function Exercice_fractions_decomposer() {
     liste_de_question_to_contenu(this); //Espacement de 2 em entre chaque questions.
   };
 }
-
-
-
 
 
 /**
@@ -7422,621 +6679,12 @@ function Nombre_decimal_oralise_de_differentes_manieres() {
 }
 
 
-/**
- * Additions, soustractions et multiplications posées de nombres entiers
- *
- * * abcd +efg
- * * abc0-efg
- * * 1abc-def
- * * abc*d0e tables de 2 à 5
- * * abc*de tables de 5 à 9
- * @Auteur Rémi Angot
- * Référence 6C10 
- */
-function Additions_soustractions_multiplications_posees() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre =
-    "Additions, soustractions et multiplications posées de nombres entiers";
-  this.consigne = "Poser et effectuer les calculs suivants.";
-  this.spacing = 2;
-  sortie_html ? (this.spacing_corr = 2) : (this.spacing_corr = 1); //Important sinon les opérations posées ne sont pas jolies
-  this.nb_questions = 5;
-  // this.pas_de_version_HMTL=true;
-  this.liste_packages = "xlop";
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    let type_de_questions_disponibles = [1, 2, 3, 4, 5];
-    let liste_type_de_questions = combinaison_listes_sans_changer_ordre(
-      type_de_questions_disponibles,
-      this.nb_questions
-    ); // Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-    if (this.nb_questions == 3) {
-      liste_type_de_questions = [1, 2, 5];
-    }
-    if (this.nb_questions == 4) {
-      liste_type_de_questions = [1, 2, 4, 5];
-    }
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0, a, b, c, d, e, f, g, x, y;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      type_de_questions = liste_type_de_questions[i];
-      switch (type_de_questions) {
-        case 1: // abcd +efg
-          a =
-            randint(1, 9) * 10000 +
-            randint(5, 9) * 1000 +
-            randint(5, 9) * 100 +
-            randint(7, 9) * 10 +
-            randint(1, 9);
-          b = randint(5, 9) * 100 + randint(7, 9) * 10 + randint(1, 9);
-          texte = `$${tex_nombre(a)}+${b}$`;
-          !sortie_html
-            ? (texte_corr = `$\\opadd{${a}}{${b}}$`)
-            : (texte_corr = `$${tex_nombre(a)}+${b}=${tex_nombre(a + b)}$`);
-          break;
-        case 2: // abc0-efg
-          a = randint(1, 9);
-          b = randint(1, 9);
-          c = randint(1, 9);
-          e = randint(b, 9);
-          f = randint(c, 9);
-          g = randint(2, 9);
-          x = a * 1000 + b * 100 + c * 10;
-          y = e * 100 + f * 10 + g;
-          texte = `$${tex_nombre(x)}-${y}$`;
-          !sortie_html
-            ? (texte_corr = `$\\opsub{${x}}{${y}}$`)
-            : (texte_corr = `$${tex_nombre(x)}+${y}=${tex_nombre(x + y)}$`);
-          break;
-        case 3: // 1abc-def
-          a = randint(1, 9);
-          b = randint(1, 9);
-          c = randint(1, 9);
-          d = randint(a, 9);
-          e = randint(1, 9);
-          f = randint(c, 9);
-          x = 1000 + a * 100 + b * 10 + c;
-          y = d * 100 + e * 10 + f;
-          texte = `$${tex_nombre(x)}-${y}$`;
-          !sortie_html
-            ? (texte_corr = `$\\opsub{${x}}{${y}}$`)
-            : (texte_corr = `$${tex_nombre(x)}-${y}=${tex_nombre(x - y)}$`);
-          break;
-        case 4: // abc*d0e tables de 2 à 5
-          a = randint(2, 5);
-          b = randint(2, 5);
-          c = randint(2, 5);
-          d = randint(2, 5);
-          e = randint(2, 5);
-          x = 100 * a + 10 * b + c;
-          y = d * 100 + e;
-          texte = `$${tex_nombre(x)}\\times${y}$`;
-          !sortie_html
-            ? (texte_corr = `$\\opmul{${x}}{${y}}$`)
-            : (texte_corr = `$${tex_nombre(x)}\\times${y}=${tex_nombre(
-              x * y
-            )}$`);
-          break;
-        case 5: // abc*de tables de 5 à 9
-          a = randint(5, 9);
-          b = randint(5, 9);
-          c = randint(5, 9);
-          d = randint(5, 9);
-          e = randint(5, 9);
-          x = 100 * a + 10 * b + c;
-          y = 10 * d + e;
-          texte = `$${x}\\times${y}$`;
-          !sortie_html
-            ? (texte_corr = `$\\opmul{${x}}{${y}}$`)
-            : (texte_corr = `$${tex_nombre(x)}\\times${y}=${tex_nombre(
-              x * y
-            )}$`);
-          break;
-      }
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        if (!sortie_html && i == 0) {
-          texte_corr = `\\setlength\\itemsep{2em}` + texte_corr;
-        } // espacement entre les questions
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-}
-
-/**
- * Additions et soustractions de nombres décimaux
- * * xxx-xx,x
- * * xxx-xx,xx
- * * xxx,x-xxx
- * * x0x-xx9,x
- * * xxx+xx,x
- * * xxx+xx,xx
- * * xxx,x+xxx
- * * x0x+xx9,x
- * @Auteur Rémi Angot
- * Référence 6C20
- */
-function Additionner_soustraires_decimaux() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Additions et soustractions de nombres décimaux";
-  this.consigne = "Poser et effectuer les calculs suivants.";
-  this.spacing = 2;
-  sortie_html ? (this.spacing_corr = 2) : (this.spacing_corr = 1); //Important sinon les opérations posées ne sont pas jolies
-  this.nb_questions = 4;
-  this.sup = 3;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    let liste_de_type_d_additions = combinaison_listes(
-      [5, 6, 7, 8],
-      this.nb_questions
-    );
-    let liste_de_type_de_soustractions = combinaison_listes(
-      [1, 2, 3, 4],
-      this.nb_questions
-    );
-    let liste_type_de_questions = [];
-    if (this.sup == 1) {
-      liste_type_de_questions = combinaison_listes([5, 6, 7, 8], this.nb_questions)
-    } else if (this.sup == 2) {
-      liste_type_de_questions = combinaison_listes([1, 2, 3, 4], this.nb_questions)
-    } else {
-      for (let i = 0; i < this.nb_questions; i++) {
-        if (i + 1 <= this.nb_questions / 2) {
-          // première moitié sont des additions mais si c'est impair on prendra plus de soustractions
-          liste_type_de_questions.push(liste_de_type_d_additions[i]);
-        } else {
-          liste_type_de_questions.push(liste_de_type_de_soustractions[i]);
-        }
-      }
-    }
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0, a, b, c, d, e, f, g, x, y;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      type_de_questions = liste_type_de_questions[i];
-      switch (type_de_questions) {
-        case 1: // xxx-xx,x
-          a = randint(1, 4) * 100 + randint(2, 5) * 10 + randint(1, 9);
-          b = calcul(randint(5, 9) * 10 + randint(6, 9) + randint(1, 9) / 10);
-          texte = `$${tex_nombre(a)}-${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}-${tex_nombre(b)}=${tex_nombrec(
-              a - b
-            )}$`)
-            : (texte_corr = `$\\opsub[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 2: // xxx-xx,xx
-          a = randint(1, 4) * 100 + randint(2, 5) * 10 + randint(1, 9);
-          b = calcul(
-            randint(5, 9) * 10 +
-            randint(6, 9) +
-            randint(1, 9) / 10 +
-            randint(1, 9) / 100
-          );
-          texte = `$${tex_nombre(a)}-${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}-${tex_nombre(b)}=${tex_nombrec(
-              a - b
-            )}$`)
-            : (texte_corr = `$\\opsub[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 3: // xxx,x-xxx
-          a = calcul(
-            randint(5, 9) * 100 +
-            randint(2, 5) * 10 +
-            randint(1, 9) +
-            randint(1, 9) / 10
-          );
-          b = randint(1, 4) * 100 + randint(6, 9) * 10 + randint(1, 9);
-          texte = `$${tex_nombre(a)}-${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}-${tex_nombre(b)}=${tex_nombrec(
-              a - b
-            )}$`)
-            : (texte_corr = `$\\opsub[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 4: // x0x-xx9,x
-          a = calcul(randint(5, 9) * 100 + randint(1, 5));
-          b = calcul(
-            randint(1, 4) * 100 + randint(1, 9) * 10 + 9 + randint(1, 9) / 10
-          );
-          texte = `$${tex_nombre(a)}-${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}-${tex_nombre(b)}=${tex_nombrec(
-              a - b
-            )}$`)
-            : (texte_corr = `$\\opsub[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 5: // xxx+xx,x
-          a = randint(1, 4) * 100 + randint(2, 5) * 10 + randint(1, 9);
-          b = calcul(randint(5, 9) * 10 + randint(6, 9) + randint(1, 9) / 10);
-          texte = `$${tex_nombre(a)}+${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}+${tex_nombre(b)}=${tex_nombrec(
-              a + b
-            )}$`)
-            : (texte_corr = `$\\opadd[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 6: // xxx+xx,xx
-          a = randint(1, 4) * 100 + randint(2, 5) * 10 + randint(1, 9);
-          b = calcul(
-            randint(5, 9) * 10 +
-            randint(6, 9) +
-            randint(1, 9) / 10 +
-            randint(1, 9) / 100
-          );
-          texte = `$${tex_nombre(a)}+${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}+${tex_nombre(b)}=${tex_nombrec(
-              a + b
-            )}$`)
-            : (texte_corr = `$\\opadd[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 7: // xxx,x+xxx
-          a = calcul(
-            randint(5, 9) * 100 +
-            randint(2, 5) * 10 +
-            randint(1, 9) +
-            randint(1, 9) / 10
-          );
-          b = randint(1, 4) * 100 + randint(6, 9) * 10 + randint(1, 9);
-          texte = `$${tex_nombre(a)}+${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}+${tex_nombre(b)}=${tex_nombrec(
-              a + b
-            )}$`)
-            : (texte_corr = `$\\opadd[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-        case 8: // x0x+xx9,x
-          a = calcul(randint(5, 9) * 100 + randint(1, 5));
-          b = calcul(
-            randint(1, 4) * 100 + randint(1, 9) * 10 + 9 + randint(1, 9) / 10
-          );
-          texte = `$${tex_nombre(a)}+${tex_nombre(b)}$`;
-          sortie_html
-            ? (texte_corr = `$${tex_nombre(a)}+${tex_nombre(b)}=${tex_nombrec(
-              a + b
-            )}$`)
-            : (texte_corr = `$\\opadd[decimalsepsymbol={,}]{${a}}{${b}}$`);
-          break;
-      }
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = ["Niveau de difficulté", 3, "1 : Additions de décimaux\n2: Soustraction de décimaux\n3 : Additions et soustraction de décimaux"];
-}
 
 
-/**
- * Plusieurs type de calcul avec des entiers.
- *
- * Sans parenthèses :
- * * a+b*c
- * * a+b÷c
- * * a÷b*c
- * * a-b*c
- * * a*b÷c
- * * a*b+c
- * * a-b+c
- * * a+b+c*d
- * * a*b+c*d
- * * a*b*c-d
- * * a*b-c÷d
- * * a*b+c÷d
- *
- * Avec parenthèses :
- * * a*(b-c)
- * * (a-b)*c
- * * (a-b)÷c
- * * a÷(b+c)
- * * (a-b)÷c
- * * a*(b-c)*d
- * * a*b*(c-d)
- * * a*(b-c*d)
- * * (a+b*c)÷d
- * * a*(b-c*d)
- * * a*b÷(c+d)
- * * a*(b÷c+d)
- * @Auteur Rémi Angot
- * Référence 6C33
- */
-function Priorites() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Calculer en utilisant les priorités opératoires";
-  this.consigne = "Calculer";
-  this.nb_questions = 5;
-  this.nb_cols = 2;
-  this.nb_cols_corr = 1;
-  this.sup = 3;
 
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let liste_questions_disponibles = []; //
-    if (this.sup == 1) {
-      liste_questions_disponibles = range1(12);
-    } else if (this.sup == 2) {
-      liste_questions_disponibles = range1(22, range1(12));
-    } else {
-      liste_questions_disponibles = range1(22);
-    }
-    let liste_type_de_questions = combinaison_listes(
-      liste_questions_disponibles,
-      this.nb_questions
-    );
-    for (
-      let i = 0, texte, texte_corr, a, b, c, d, cpt = 0;
-      i < this.nb_questions && cpt < 50;
 
-    ) {
-      switch (liste_type_de_questions[i]) {
-        case 1:
-          a = randint(2, 11);
-          b = randint(2, 11);
-          c = randint(2, 11);
-          texte = `$${a}+${b}\\times${c}$`;
-          texte_corr = `$${a}+${mise_en_evidence(b + "\\times" + c)}=${a}+${b * c
-            }=${a + b * c}$`;
-          break;
-        case 2:
-          a = randint(2, 11);
-          c = randint(2, 11);
-          b = c * randint(2, 11);
-          texte = `$${a}+${b}\\div${c}$`;
-          texte_corr = `$${a}+${mise_en_evidence(b + "\\div" + c)}=${a}+${b / c
-            }=${a + b / c}$`;
-          break;
-        case 3:
-          b = randint(2, 11);
-          c = randint(2, 11);
-          a = b * randint(2, 11);
-          texte = `$${a}\\div${b}\\times${c}$`;
-          texte_corr = `$${mise_en_evidence(a + "\\div" + b)}\\times${c}=${a / b
-            }\\times${c}=${(a / b) * c}$`;
-          break;
-        case 4:
-          b = randint(2, 11);
-          c = randint(2, 11);
-          a = b * c + randint(2, 11);
-          texte = `$${a}-${b}\\times${c}$`;
-          texte_corr = `$${a}-${mise_en_evidence(b + "\\times" + c)}=${a}-${b * c
-            }=${a - b * c}$`;
-          break;
-        case 5:
-          if (choice([true, false])) {
-            //a est un multiple de c
-            c = randint(2, 6);
-            a = c * randint(2, 5);
-            b = randint(2, 6);
-          } else {
-            // b est un multiple de c
-            c = randint(2, 6);
-            b = c * randint(2, 5);
-            a = randint(2, 6);
-          }
-          texte = `$${a}\\times${b}\\div${c}$`;
-          texte_corr = `$${mise_en_evidence(a + "\\times" + b)}\\div${c}=${a * b
-            }\\div${c}=${(a * b) / c}$`;
-          break;
-        case 6:
-          a = randint(2, 11);
-          b = randint(2, 11);
-          c = randint(2, 11);
-          texte = `$${a}\\times${b}+${c}$`;
-          texte_corr = `$${mise_en_evidence(a + "\\times" + b)}+${c}=${a * b
-            }+${c}=${a * b + c}$`;
-          break;
-        case 7:
-          b = randint(20, 59);
-          a = b + randint(11, 29);
-          c = randint(11, 29);
-          texte = `$${a}-${b}+${c}$`;
-          texte_corr = `$${mise_en_evidence(a + "-" + b)}+${c}=${a - b}+${c}=${a - b + c
-            }$`;
-          break;
-        case 8:
-          a = randint(2, 20);
-          b = randint(2, 20);
-          c = randint(2, 11);
-          d = randint(2, 11);
-          texte = `$${a}+${b}+${c}\\times${d}$`;
-          texte_corr = `$${a}+${b}+${mise_en_evidence(
-            c + "\\times" + d
-          )}=${a}+${b}+${c * d}=${a + b + c * d}$`;
-          break;
-        case 9:
-          a = randint(2, 11);
-          b = randint(2, 11);
-          c = randint(2, 11);
-          d = randint(2, 11);
-          texte = `$${a}\\times${b}+${c}\\times${d}$`;
-          texte_corr = `$${mise_en_evidence(
-            a + "\\times" + b
-          )}+${mise_en_evidence(c + "\\times" + d)}=${a * b}+${c * d}=${a * b + c * d
-            }$`;
-          break;
-        case 10:
-          a = randint(2, 5);
-          b = randint(2, 5);
-          c = randint(2, 5);
-          d = randint(2, a * b * c - 1);
-          texte = `$${a}\\times${b}\\times${c}-${d}$`;
-          texte_corr = `$${mise_en_evidence(
-            a + "\\times" + b
-          )}\\times${c}-${d}=${mise_en_evidence(a * b + "\\times" + c)}-${d}=${a * b * c - d
-            }$`;
-          break;
-        case 11:
-          a = randint(3, 11);
-          b = randint(3, 11);
-          d = randint(2, 11);
-          c = d * randint(2, 8);
-          texte = `$${a}\\times${b}-${c}\\div${d}$`;
-          texte_corr = `$${mise_en_evidence(
-            a + "\\times" + b
-          )}-${mise_en_evidence(c + "\\div" + d)}=${a * b}-${c / d}=${a * b - c / d
-            }$`;
-          break;
-        case 12:
-          a = randint(2, 11);
-          b = randint(2, 11);
-          d = randint(2, 11);
-          c = d * randint(2, 8);
-          texte = `$${a}\\times${b}+${c}\\div${d}$`;
-          texte_corr = `$${mise_en_evidence(
-            a + "\\times" + b
-          )}+${mise_en_evidence(c + "\\div" + d)}=${a * b}+${c / d}=${a * b + c / d
-            }$`;
-          break;
-        case 13:
-          a = randint(2, 11);
-          c = randint(2, 11);
-          b = c + randint(2, 11);
-          texte = `$${a}\\times(${b}-${c})$`;
-          texte_corr = `$${a}\\times(${mise_en_evidence(
-            b + `-` + c
-          )})=${a}\\times${b - c}=${a * (b - c)}$`;
-          break;
-        case 14:
-          b = randint(11, 39);
-          a = b + randint(2, 11);
-          c = randint(2, 11);
-          texte = `$(${a}-${b})\\times${c}$`;
-          texte_corr = `$(${mise_en_evidence(a + `-` + b)})\\times${c}=${a - b
-            }\\times${c}=${(a - b) * c}$`;
-          break;
-        case 15:
-          c = randint(2, 11);
-          b = randint(11, 39);
-          a = b + c * randint(2, 9);
-          texte = `$(${a}-${b})\\div${c}$`;
-          texte_corr = `$(${mise_en_evidence(a + `-` + b)})\\div${c}=${a - b
-            }\\div${c}=${(a - b) / c}$`;
-          break;
-        case 16:
-          b = randint(2, 5);
-          c = randint(2, 6);
-          a = (b + c) * randint(2, 9);
-          texte = `$${a}\\div(${b}+${c})$`;
-          texte_corr = `$${a}\\div(${mise_en_evidence(b + `+` + c)})=${a}\\div${b + c
-            }=${a / (b + c)}$`;
-          break;
-        case 17:
-          c = randint(2, 11);
-          b = randint(11, 39);
-          a = b + c * randint(2, 9);
-          texte = `$(${a}-${b})\\div${c}$`;
-          texte_corr = `$(${mise_en_evidence(a + `-` + b)})\\div${c}=${a - b
-            }\\div${c}=${(a - b) / c}$`;
-          break;
-        case 18:
-          c = randint(11, 39);
-          b = c + randint(2, 5);
-          a = randint(2, 5);
-          d = randint(2, 5);
-          texte = `$${a}\\times(${b}-${c})\\times${d}$`;
-          texte_corr = `$${a}\\times(${mise_en_evidence(
-            b + `-` + c
-          )})\\times${d}=${a}\\times${b - c}\\times${d}=${a * (b - c) * d}$`;
-          break;
-        case 19:
-          d = randint(11, 39);
-          c = d + randint(2, 5);
-          a = randint(2, 5);
-          b = randint(2, 5);
-          texte = `$${a}\\times${b}\\times(${c}-${d})$`;
-          texte_corr = `$${a}\\times${b}\\times(${mise_en_evidence(
-            c + `-` + d
-          )})=${a}\\times${b}\\times${c - d}=${a * b * (c - d)}$`;
-          break;
-        case 20:
-          a = randint(2, 11);
-          c = randint(2, 11);
-          d = randint(2, 11);
-          b = c * d + randint(2, 11);
-          texte = `$${a}\\times(${b}-${c}\\times${d})$`;
-          texte_corr = `$${a}\\times(${b}-${mise_en_evidence(
-            c + `\\times` + d
-          )})=${a}\\times(${mise_en_evidence(b + `-` + c * d)})=${a}\\times${b - c * d
-            }=${a * (b - c * d)}$`;
-          break;
-        case 21:
-          a = randint(2, 11);
-          b = randint(2, 11);
-          if (liste_des_diviseurs(a * b).length <= 2) {
-          }
-          let liste = liste_des_diviseurs(a * b);
-          if (liste.length > 2) {
-            liste.pop(); //on supprime le plus grand diviseur qui est le produit
-          }
-          if (liste.length > 2) {
-            enleve_element(liste, liste[1]); //on supprime le plus petit diviseur (autre que 1)
-          }
 
-          let somme = choice(liste, [1]); // la somme doit être un diviseur différent de 1
-          c = randint(1, somme - 1);
-          d = somme - c;
-          texte = `$${a}\\times${b}\\div(${c}+${d})$`;
-          texte_corr = `$${a}\\times${b}\\div(${mise_en_evidence(
-            c + `+` + d
-          )})=${mise_en_evidence(a + "\\times" + b)}\\div${c + d}=${a * b
-            }\\div${c + d}=${(a * b) / (c + d)}$`;
-          break;
-        case 22:
-          a = randint(2, 11);
-          c = randint(2, 11);
-          b = c * randint(2, 5);
-          d = randint(2, 6);
-          texte = `$${a}\\times(${b}\\div${c}+${d})$`;
-          texte_corr = `$${a}\\times(${mise_en_evidence(
-            b + `\\div` + c
-          )}+${d})=${a}\\times(${mise_en_evidence(
-            b / c + `+` + d
-          )})=${a}\\times${b / c + d}=${a * (b / c + d)}$`;
-          break;
-      }
 
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = [
-    "Type de calculs",
-    3,
-    "1 : Sans parenthèses\n2: Avec parenthèses\n3: Avec ou sans parenthèses",
-  ];
-}
 
 /**
  * @Auteur Rémi Angot
@@ -8124,1122 +6772,6 @@ function HTML_personnalise() {
     this.contenu_correction = this.sup2;
   };
 }
-/**
- * @Auteur Jean-Claude Lhote
- * publié le 14/12/2020
- * Réf : 6G25-3
- * Trouver une figure symétrique dans un pavage. Symétrie axiale. 6 pavages différents.
- */
-function Pavage_et_reflexion2d() {
-  "use strict";
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre =
-    "Trouver l\'image d'une figure par une symétrie axiale dans un pavage";
-  this.consigne = "";
-  this.nb_questions = 3;
-  this.nb_questions_modifiable = true;
-  this.correction_detaillee=true;
-  this.correction_detaillee_disponible=true;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 1; // 1 pour des pavages modestes, 2 pour des plus grand.
-  this.sup2=false // On cache les centres par défaut.
-  this.sup3=7;
-  sortie_html ? (this.spacing_corr = 2.5) : (this.spacing_corr = 1.5);
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let videcouples=function(tableau){
-      for (let k=0;k<tableau.length;k++){
-        for (let j=k+1;j<tableau.length;j++){
-          if (tableau[k][1]==tableau[j][0]) {
-            tableau.splice(j,1)
-          }
-        }
-      }
-      return tableau
-    }
-    let compare2polys=function(poly1,poly2){
-      if (comparenbsommets(poly1,poly2)) {
-        if (comparesommets(poly1,poly2)) 
-          return true
-        else
-          return false
-      }
-      else 
-        return false 
-      }
-      let comparenbsommets = function(poly1,poly2){
-        if (poly1.listePoints.length==poly2.listePoints.length){
-          return true
-        }
-        else return false
-      }
-      
-      let compare2sommets=function(sommet1,sommet2){
-        if (egal(sommet1.x,sommet2.x,0.1)&&egal(sommet1.y,sommet2.y,0.1)) {
-          return true
-        }
-        else return false
-      }
-      let comparesommets = function(poly1,poly2){
-        let trouve=false,trouves=0
-        if (comparenbsommets(poly1,poly2))
-        for (let P of poly1.listePoints) {
-          for (let M of poly2.listePoints) {
-            if (compare2sommets(M,P)) {
-              trouve=true
-            }
-            if (trouve) break
-          }
-          if (trouve) {
-            trouves++
-            trouve=false
-          }
-          else {
-            trouves-=100
-          }
-          if (trouves<0)
-          break
-        }
-        if (trouves==poly1.listePoints.length)
-          return true
-        else return false
-      }
- /*     let associesommets=function(poly1,poly2,d){ //Pour chercher les indices des symétriques dans leur polygone respectif
-        let binomes=[],P,M   
-        for (let k=0;k<poly1.listePoints.length;k++) { // afin éventuellement de faire clignoter ces paires de points lors de la correction
-          P=symetrieAxiale(poly1.listePoints[k],d)
-          for (let l=0;l<poly2.listePoints.length;l++) {
-            M=poly2.listePoints[l]
-            if (compare2sommets(M,P)) {
-              binomes.push([k,l])
-              break
-            }
-          }
-        }
-        return binomes
-      }
-  */  
-    let refleccion = function (pavage, d, numero) { // retourne le numero du polygone symétrique ou -1 si il n'existe pas
-      let poly=pavage.polygones[numero-1],pol
-      let result=-1
-      let sympoly=symetrieAxiale(poly,d)
-      for (let k= 0;k<pavage.polygones.length;k++) {
-        pol=pavage.polygones[k]
-        if (compare2polys(sympoly,pol)) {
-          return k+1
-        }
-      }
-      return result
-    } 
-
-    let objets=[],objets_correction=[],symetriques=[],P1,P2,P3,t
-    let codes=['/','//','///','o','w','X','U','*']
-    let taillePavage=parseInt(this.sup)
-    if (taillePavage<1||taillePavage>2) {
-      taillePavage=1
-    }
-    if (this.nb_questions>5) {
-      taillePavage=2
-    }
-    this.liste_corrections = []
-    this.liste_questions = []
-    let Nx,Ny,index1,index2,A,B,d,image,couples=[],tailles=[],monpavage,fenetre
-    let texte = "", texte_corr = "", type_de_pavage = parseInt(this.sup)
-    let nombreTentatives,nombrePavageTestes=1
-    if (this.sup3==8) {
-      type_de_pavage =  randint(1,7)
-    }
-    else {
-      type_de_pavage=parseInt(this.sup3)
-    }
-    while (couples.length<this.nb_questions&&nombrePavageTestes<7){
-      nombreTentatives=0
-    monpavage = pavage() // On crée l'objet Pavage qui va s'appeler monpavage
-    tailles = [[[3, 2], [3, 2], [2, 2], [2, 2], [2, 2], [2, 2],[3,2]], [[4, 3], [4, 3], [3, 3], [3, 3], [3, 3], [3, 2],[5,3]]]
-    Nx = tailles[taillePavage-1][type_de_pavage-1][0]
-    Ny = tailles[taillePavage-1][type_de_pavage-1][1]
-    monpavage.construit(type_de_pavage, Nx, Ny, 3) // On initialise toutes les propriétés de l'objet.
-    fenetre=monpavage.fenetre
-    fenetreMathalea2d=[fenetre.xmin,fenetre.ymin,fenetre.xmax,fenetre.ymax]
-    while (couples.length<this.nb_questions+2&&nombreTentatives<3) { // On cherche d pour avoir suffisamment de couples
-    couples=[] // On vide la liste des couples pour une nouvelle recherche
-    index1=randint(Math.floor(monpavage.nb_polygones/3),Math.ceil(monpavage.nb_polygones*2/3)) // On choisit 2 points dans 2 polygones distincts.
-    index2=randint(Math.floor(monpavage.nb_polygones/3),Math.ceil(monpavage.nb_polygones*2/3),index1) 
-    A=monpavage.polygones[index1].listePoints[randint(0,2)] // On les choisit dans les trois premiers
-    B=monpavage.polygones[index2].listePoints[randint(0,2)] // points pour éviter un point qui n'éxiste pas
-    while (compare2sommets(A,B)){ // On vérifie qu'ils sont bien distincts sinon, on change.
-      index1=randint(0,monpavage.nb_polygones-1) 
-      index2=randint(0,monpavage.nb_polygones-1,index1)
-      A=monpavage.polygones[index1].listePoints[randint(0,2)] // idem ci-dessus
-      B=monpavage.polygones[index2].listePoints[randint(0,2)] // mais à la sortie du While A!=B
-    }
-    d=mediatrice(A,B,'(d)','red') // l'axe sera la droite passant par ces deux points si ça fonctionne
-    d.epaisseur=3
-    for (let i=1;i<= monpavage.nb_polygones; i++){ //on crée une liste des couples (antécédents, images)
-      image=refleccion(monpavage,d,i)
-      if (image!=-1){ // si l'image du polygone i existe, on ajoute le couple à la liste
-        couples.push([i,image])
-      }
-    }
-    couples=videcouples(couples) //supprime tous les couples en double (x,y)=(y,x)
-    nombreTentatives++ 
-    }
-    if (couples.length<this.nb_questions){
-    if (this.sup3==7) {
-      type_de_pavage=(type_de_pavage+1)%5+1
-    }
-    nombrePavageTestes++
-    }
-  }
-  if (couples.length<this.nb_questions){
-    console.log('trop de questions, augmentez la taille du pavage')
-    return
-  }
-
-    objets.push(d) // la droite d est trouvée
-    couples=shuffle(couples) // on mélange les couples
-    for (let i = 0; i < monpavage.nb_polygones; i++) {
-      objets.push(texteParPosition(nombre_avec_espace(i + 1), monpavage.barycentres[i].x + 0.5, monpavage.barycentres[i].y, 'milieu', 'gray', 1, 0, true))
-    }
-    if (this.sup2) { // Doit-on montrer les centres des figures ?
-      for (let i = 0; i < monpavage.nb_polygones; i++) {
-        objets.push(monpavage.tracesCentres[i])
-      }
-    }
-    for (let i = 0; i < monpavage.nb_polygones; i++) { // il faut afficher tous les polygones du pavage
-      objets.push(monpavage.polygones[i])
-    }
-    texte = mathalea2d(fenetre, objets) // monpavage.fenetre est calibrée pour faire entrer le pavage dans une feuille A4
-    texte+=`<br>`
-    for (let i=0;i<this.nb_questions;i++){  
-      texte+=`Quel est l'image de la figure $${couples[i][0]}$ dans la symétrie d'axe $(d)$ ?<br>`
-      texte_corr+=`L'image de la figure $${couples[i][0]}$ dans la symétrie d'axe $(d)$ est la figure ${couples[i][1]}<br>`
-//      symetriques=associesommets(monpavage.polygones[couples[i][0]-1],monpavage.polygones[couples[i][1]-1],d)
-      if (this.correction_detaillee){
-        t=this.nb_questions*3;
-        A=monpavage.barycentres[couples[i][0]-1]
-        B=monpavage.barycentres[couples[i][1]-1]
-        P1=monpavage.polygones[couples[i][0]-1]
-        P1.color=texcolors(i)
-        P1.couleurDeRemplissage=texcolors(i)
-        P1.opaciteDeRemplissage=0.5
-        P1.epaisseur=2
-        P2=monpavage.polygones[couples[i][1]-1]
-        P2.color=texcolors(i)
-        P2.couleurDeRemplissage=texcolors(i)
-        P2.opaciteDeRemplissage=0.5
-        P2.epaisseur=2
-        P3=symetrieAnimee(P1,d,`begin="${i*3}s;${i*3+t}s;${i*3+t*2}s" end="${i*3+2}s;${i*3+t+2}s;${i*3+t*2+2}s" dur="2s" repeatCount="indefinite" repeatDur="${9*this.nb_questions}s" id="poly-${i}-anim"`)
-        P3.color=texcolors(i)
-        P3.epaisseur=2
-        objets_correction.push(tracePoint(A,B),segment(A,B,texcolors(i)),codageMediatrice(A,B,texcolors(i),codes[i]),P1,P2,P3)
-      }
-    }
-    if (this.correction_detaillee){
-      texte_corr+=mathalea2d(fenetre, objets,objets_correction)
-    }
-    this.liste_questions.push(texte);
-    this.liste_corrections.push(texte_corr);
-    liste_de_question_to_contenu(this)
-  }
-	this.besoin_formulaire_numerique = ['Taille du pavage (la grande est automatique au-delà de 5 questions)', 2, '1 : Taille modeste\n 2 : Grande taille'];
-  this.besoin_formulaire2_case_a_cocher=["Montrer les centres"]
-  this.besoin_formulaire3_numerique=['Choix du pavage',8,'1 : Pavage de triangles équilatéraux\n2 : Pavage de carrés\n3 : Pavage d\'hexagones réguliers\n4 : Pavage 3².4.3.4\n5 : Pavage 8².4\n 6 : Pavage de losanges (hexagonal d\'écolier)\n7 : Pavage 6.3.6.3\n8 : Un des sept pavages au hasard']
-}
-/**
- * Pavages et symétrie axiale.
- * Pas de version LaTeX
- * @Auteur Jean-Claude Lhote
- * Publié en 02/2020
- * référence 6G25-1
- */
-function Pavages_et_reflexion() {
-  "use strict";
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre =
-    "Trouver l'image d'une figure par une symétrie axiale dans un pavage triangulaire (sortie Latex Impossible)";
-  this.pas_de_version_LaTeX = true;
-  this.consigne = "";
-  this.nb_questions = 1;
-  this.nb_questions_modifiable = false;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  // this.sup = 1; // 1 pour les 6ème, 2 pour les 5èmes, 3 pour les 4èmes, et 4 pour les 3èmes.
-  sortie_html ? (this.spacing_corr = 2.5) : (this.spacing_corr = 1.5);
-  this.liste_packages = "tkz-euclide";
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.type_exercice = "MG32";
-    this.MG32editable = false;
-    this.taille_div_MG32 = [700, 700];
-    this.MG32codeBase64 =
-      "TWF0aEdyYXBoSmF2YTEuMAAAABI+TMzNAAJmcv###wEA#wEAAAAAAAAAAAUcAAAC0gAAAQEAAAAAAAAAAQAAAIP#####AAAAAQAKQ0NhbGNDb25zdAD#####AAJwaQAWMy4xNDE1OTI2NTM1ODk3OTMyMzg0Nv####8AAAABAApDQ29uc3RhbnRlQAkh+1RELRj#####AAAAAQAKQ1BvaW50QmFzZQD#####AAAAAAEQAAFPAAAAAAAAAAAAQAgAAAAAAAAFAAFATMAAAAAAAEBTYUeuFHri#####wAAAAEAFENEcm9pdGVEaXJlY3Rpb25GaXhlAP####8BAAAAARAAAAEAAQAAAAEBP#AAAAAAAAD#####AAAAAQAPQ1BvaW50TGllRHJvaXRlAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQABQF6AAAAAAAAAAAAC#####wAAAAEAB0NDYWxjdWwA#####wAEem9vbQABOAAAAAFAIAAAAAAAAP####8AAAABAAtDSG9tb3RoZXRpZQD#####AAAAAf####8AAAABAApDT3BlcmF0aW9uAwAAAAE#8AAAAAAAAAAAAAcBAAAAAUA0AAAAAAAA#####wAAAAEAD0NSZXN1bHRhdFZhbGV1cgAAAAT#####AAAAAQALQ1BvaW50SW1hZ2UA#####wEAAAABEAABSQAAAAAAAAAAAEAIAAAAAAAABQAAAAADAAAABf####8AAAABAAlDUm90YXRpb24A#####wAAAAH#####AAAAAQAMQ01vaW5zVW5haXJlAAAAAUBOAAAAAAAAAAAACQD#####AQAAAAEQAAFKAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAYAAAAH#####wAAAAIAB0NSZXBlcmUA#####wDm5uYAAQAAAAEAAAAGAAAACAAAAAAAAAEAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAE#8AAAAAAAAAAAAAE#8AAAAAAAAAAAAAUA#####wACeEIAATYAAAABQBgAAAAAAAAAAAAFAP####8AAnlCAAEwAAAAAQAAAAAAAAAA#####wAAAAEAEENQb2ludERhbnNSZXBlcmUA#####wEAAAABEAABQgAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAACAAAAAoAAAAIAAAACwAAAAUA#####wACeEMAATAAAAABAAAAAAAAAAAAAAAFAP####8AAnlDAAE2AAAAAUAYAAAAAAAAAAAADQD#####AQAAAAEQAAFDAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAkAAAAIAAAADQAAAAgAAAAOAAAABQD#####AAN4TTEAAjQwAAAAAUBEAAAAAAAAAAAABQD#####AAN5TTEAATAAAAABAAAAAAAAAAAAAAANAP####8BAAAAABAAAk0xAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAkAAAAIAAAAEAAAAAgAAAARAAAADQD#####Af8A#wAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAACQAAAAFAAAAAAAAAAAAAAAFAAAAAAAAAAP####8AAAABAAxDVHJhbnNsYXRpb24A#####wAAABMAAAABAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAEgAAABQAAAAOAP####8AAAABAAAAFQAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAEAAAAWAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAADAAAABYAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAPAAAAFv####8AAAABAAlDUG9seWdvbmUA#####wAAfwAAAQAAAAQAAAAXAAAAGAAAABkAAAAX#####wAAAAEAEENTdXJmYWNlUG9seWdvbmUA#####wAAfwAAAAAFAAAAGgAAAAUA#####wACbngAATcAAAABQBwAAAAAAAAAAAAFAP####8AAm55AAE3AAAAAUAcAAAAAAAAAAAADAD#####AObm5gABAAAAAQAAAAwAAAAPAAAAAAAAAQAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAT#wAAAAAAAAAAAAAT#wAAAAAAAAAAAADQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAHgAAAAgAAAAcAAAACAAAAB3#####AAAAAQAHQ01pbGlldQD#####Af8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAAQAAAB8AAAAFAP####8AA3hOMwABOAAAAAFAIAAAAAAAAAAAAAUA#####wADeU4zAAEyAAAAAUAAAAAAAAAAAAAABQD#####AAN4TTMAATQAAAABQBAAAAAAAAAAAAAFAP####8AA3lNMwACMTAAAAABQCQAAAAAAAAAAAANAP####8BAAD#ABAAAk4zAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAkAAAAIAAAAIQAAAAgAAAAiAAAADQD#####AQAA#wAQAAJNMwAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAACAAAACMAAAAIAAAAJP####8AAAABAAtDTWVkaWF0cmljZQD#####AAAA#wAQAAABAAIAAAAlAAAAJgAAAAUA#####wADeE4yAAEwAAAAAQAAAAAAAAAAAAAABQD#####AAN5TjIAATYAAAABQBgAAAAAAAAAAAAFAP####8AA3lNMgABNgAAAAFAGAAAAAAAAAAAAAUA#####wADeE0yAAEwAAAAAQAAAAAAAAAAAAAADQD#####AQB#AAAQAAJOMgAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAACAAAACgAAAAIAAAAKQAAAA0A#####wEAfwAAEAACTTIAAAAAAAAAAABACAAAAAAAAAUAAAAACQAAAAgAAAArAAAACAAAACoAAAAFAP####8AA3hOMQACNDIAAAABQEUAAAAAAAAAAAAFAP####8AA3lOMQABMgAAAAFAAAAAAAAAAAAAAA0A#####wEAAAAAEAACTjEAAAAAAAAAAABACAAAAAAAAAUAAAAACQAAAAgAAAAuAAAACAAAAC8AAAASAP####8AAH8AARAABChkMSkBAAI#7MzMzMzMzQAAADAAAAAS#####wAAAAEAEUNQb2ludFBhckFic2Npc3NlAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAABAAAADAAAAAcBAAAACAAAABwAAAABP#AAAAAAAAD#####AAAAAQAIQ1NlZ21lbnQA#####wAAAAAAEAAAAQABAAAAAQAAADIAAAATAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAABAAAADwAAAAcBAAAACAAAAB0AAAABP#AAAAAAAAAAAAAEAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQABAAAAAAAAAAAAAAAzAAAADgD#####AAAAAQAAADQAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAA1AAAANgAAABQA#####wEAAAAAEAAAAQABAAAANQAAADcAAAAEAP####8BAAAAARAAAVUAAAAAAAAAAABACAAAAAAAAAUAAT#Q3iTV#EVNAAAAOP####8AAAACAA1DTWVzdXJlQWZmaXhlAP####8AAXoAAAAeAAAAOf####8AAAABAA1DRm9uY0NvbXBsZXhlAP####8AA251bQASMipyZSh6KSsyKm54KmltKHopAAAABwAAAAAHAgAAAAFAAAAAAAAAAP####8AAAACAAlDRm9uY3Rpb25M#####wAAAAIAEUNWYXJpYWJsZUZvcm1lbGxlAAAAAAAAAAcCAAAABwIAAAABQAAAAAAAAAD#####AAAAAQAXQ1Jlc3VsdGF0VmFsZXVyQ29tcGxleGUAAAAcAAAAF00AAAAYAAAAAAABegAAAA8A#####wH#AP8AAQAAAAQAAAABAAAADAAAAA8AAAAB#####wAAAAIACENNZXN1cmVZAP####8AAnlVAAAACQAAADn#####AAAAAgAIQ01lc3VyZVgA#####wACeFUAAAAJAAAAOQAAAA4A#####wAAAAEAAAA5AAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAAQAAAD8AAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAMAAAAPwAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAA8AAAA#AAAADwD#####AQAAAAABAAAABAAAAEAAAABBAAAAQgAAAED#####AAAAAgASQ0xpZXVPYmpldFBhclB0TGllAP####8AAAAAAAAAQwAAAAgAAAAdAAAAOQAAAAYAAAA5AAAAPwAAAEAAAABBAAAAQgAAAEMAAAAcAP####8AAAAAAAAARAAAAAgAAAAcAAAANQAAAAoAAAA1AAAANwAAADgAAAA5AAAAPwAAAEAAAABBAAAAQgAAAEMAAABEAAAACgD#####AAAADAAAAAFATgAAAAAAAAAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAEAAABGAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAADAAAAEYAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAPAAAARgAAAA8A#####wH#AP8ABAAAAAQAAABHAAAASAAAAEkAAABHAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAARwAAAD8AAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABIAAAAPwAAAAkA#####wEAAAAAEAABQQAAAAAAAAAAAEAIAAAAAAAABQAAAABJAAAAPwAAAA8A#####wEAAAAAAQAAAAQAAABLAAAATAAAAE0AAABLAAAAHAD#####AQAAAAAAAE4AAAAIAAAAHQAAADkAAAAGAAAAOQAAAD8AAABLAAAATAAAAE0AAABOAAAAHAD#####AAAAAAAAAE8AAAAIAAAAHAAAADUAAAAKAAAANQAAADcAAAA4AAAAOQAAAD8AAABLAAAATAAAAE0AAABOAAAATwAAABMA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAADkAAABNAAAAAT#VVUdaMaS+AAAAEwD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAOQAAAE0AAAABP+VVR1oxpL4AAAAFAP####8AA3hVJwAbaW50KHhVKjEwMDAwMDArMC41KS8xMDAwMDAwAAAABwMAAAAXAgAAAAcAAAAABwIAAAAIAAAAPgAAAAFBLoSAAAAAAAAAAAE#4AAAAAAAAAAAAAFBLoSAAAAAAAAAAAUA#####wADeVUnABtpbnQoeVUqMTAwMDAwMCswLjUpLzEwMDAwMDAAAAAHAwAAABcCAAAABwAAAAAHAgAAAAgAAAA9AAAAAUEuhIAAAAAAAAAAAT#gAAAAAAAAAAAAAUEuhIAAAAAA#####wAAAAEAD0NDYWxjdWxDb21wbGV4ZQD#####AAJ6VQAJeFUnK2kqeVUnAAAABwAAAAAZAAAAUwAAAAcC#####wAAAAEAC0NDb25zdGFudGVpAAAAGQAAAFQAAAAdAP####8ABG51bVUABm51bSh6Kf####8AAAABABZDQXBwZWxGb25jdGlvbkNvbXBsZXhlAAAAOwAAABkAAAA6#####wAAAAEAD0NWYWxldXJBZmZpY2hlZQD#####AQAAAADAMQAAAAAAAMAkAAAAAAAAAAAAURAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAIAAABWAAAAHQD#####AAVudW1VMQAGbnVtVSsxAAAABwAAAAAZAAAAVgAAAAE#8AAAAAAAAAAAACAA#####wEAAAAAwC4AAAAAAADAIAAAAAAAAAAAAFIQAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAACAAAAWAAAABwA#####wAAAAAAAABXAAAACAAAAB0AAAA5AAAABwAAADkAAAA6AAAAPwAAAE0AAABRAAAAVgAAAFcAAAAcAP####8AAAAAAAAAWgAAAAgAAAAcAAAANQAAAAsAAAA1AAAANwAAADgAAAA5AAAAOgAAAD8AAABNAAAAUQAAAFYAAABXAAAAWgAAABwA#####wAAAAAAAABZAAAACAAAAB0AAAA5AAAACAAAADkAAAA6AAAAPwAAAE0AAABSAAAAVgAAAFgAAABZAAAAHAD#####AAAAAAAAAFwAAAAIAAAAHAAAADUAAAAMAAAANQAAADcAAAA4AAAAOQAAADoAAAA#AAAATQAAAFIAAABWAAAAWAAAAFkAAABcAAAACQD#####Af8A#wAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAALQAAABQAAAAOAP####8AAAABAAAAXgAAAAkA#####wH#AP8AEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAEAAABfAAAACQD#####Af8A#wAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAADAAAAF8AAAAJAP####8B#wD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAPAAAAXwAAAA8A#####wD#AAAAAgAAAAQAAABgAAAAYQAAAGIAAABgAAAAEAD#####AP8AAAAAAAUAAABjAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAJgAAABQAAAAOAP####8AAAABAAAAZQAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAEAAABmAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAADAAAAGYAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAPAAAAZgAAAAkA#####wEAfwAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAACYAAAAUAAAADgD#####AAAAAQAAABMAAAAJAP####8B#wD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAATAAAAawAAAA4A#####wAAAGwAAAAmAAAACQD#####Af8A#wAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAARwAAAG0AAAAJAP####8B#wD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABIAAAAbQAAAAkA#####wH#AP8AEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAEkAAABtAAAADwD#####AAAA#wABAAAABAAAAG4AAABvAAAAcAAAAG4AAAAQAP####8AAAD#AAAABQAAAHEAAAAOAP####8AAAABAAAADwAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAADQAAABzAAAAFAD#####AAAAAAAQAAABAAEAAAB0AAAAH#####8AAAABABBDSW50RHJvaXRlRHJvaXRlAP####8AAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAxAAAAM#####8AAAACAAxDQ29tbWVudGFpcmUA#####wAAfwAAQDUAAAAAAADARQAAAAAAAQAAAHYQAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAIKGQjTCgxKSkAAAAUAP####8AAAD#ABAAAAEAAQAAAAEAAAA0AAAAIQD#####AQAA#wAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAeAAAACcAAAAiAP####8AAAD#AMBAAAAAAAAAwAgAAAAAAAAAAAB5EAAAAAAAAAAAAAAAAAABAAAAAAAAAAAACChkI0woMykpAAAAIQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAMQAAAHUAAAAiAP####8AAH8AAMBEgAAAAAAAP#AAAAAAAAAAAAB7EAAAAAAAAAAAAAAAAAABAAAAAAAAAAAACChkI0woMSkpAAAAEQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAALQAAACz#####AAAAAQAJQ0Ryb2l0ZU9tAP####8A#wAAABAAAAEAAgAAAAkAAAB9AAAAAT#wAAAAAAAAAAAAIQD#####AP8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAfgAAADMAAAAiAP####8A#wAAAMA1AAAAAAAAwEKAAAAAAAAAAAB#EAAAAAAAAAAAAAAAAAABAAAAAAAAAAAACChkI0woMikpAAAAIQD#####AP8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAfgAAAHUAAAAiAP####8A#wAAAMAUAAAAAAAAQCYAAAAAAAAAAACBEAAAAAAAAAAAAAAAAAABAAAAAAAAAAAACChkI0woMikp################";
-    this.MG32codeBase64corr =
-      "TWF0aEdyYXBoSmF2YTEuMAAAABI+TMzNAAJmcv###wEA#wEAAAAAAAAAAAUcAAAC0gAAAQEAAAAAAAAAAQAAAMz#####AAAAAQAKQ0NhbGNDb25zdAD#####AAJwaQAWMy4xNDE1OTI2NTM1ODk3OTMyMzg0Nv####8AAAABAApDQ29uc3RhbnRlQAkh+1RELRj#####AAAAAQAKQ1BvaW50QmFzZQD#####AAAAAAEQAAFPAAAAAAAAAAAAQAgAAAAAAAAFAAFATMAAAAAAAEBTYUeuFHri#####wAAAAEAFENEcm9pdGVEaXJlY3Rpb25GaXhlAP####8BAAAAARAAAAEAAQAAAAEBP#AAAAAAAAD#####AAAAAQAPQ1BvaW50TGllRHJvaXRlAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQABQF6AAAAAAAAAAAAC#####wAAAAEAB0NDYWxjdWwA#####wAEem9vbQABOAAAAAFAIAAAAAAAAP####8AAAABAAtDSG9tb3RoZXRpZQD#####AAAAAf####8AAAABAApDT3BlcmF0aW9uAwAAAAE#8AAAAAAAAAAAAAcBAAAAAUA0AAAAAAAA#####wAAAAEAD0NSZXN1bHRhdFZhbGV1cgAAAAT#####AAAAAQALQ1BvaW50SW1hZ2UA#####wEAAAABEAABSQAAAAAAAAAAAEAIAAAAAAAABQAAAAADAAAABf####8AAAABAAlDUm90YXRpb24A#####wAAAAH#####AAAAAQAMQ01vaW5zVW5haXJlAAAAAUBOAAAAAAAAAAAACQD#####AQAAAAEQAAFKAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAYAAAAH#####wAAAAIAB0NSZXBlcmUA#####wDm5uYAAQAAAAEAAAAGAAAACAAAAAAAAAEAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAE#8AAAAAAAAAAAAAE#8AAAAAAAAAAAAAUA#####wACeEIAATYAAAABQBgAAAAAAAAAAAAFAP####8AAnlCAAEwAAAAAQAAAAAAAAAA#####wAAAAEAEENQb2ludERhbnNSZXBlcmUA#####wEAAAABEAABQgAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAACAAAAAoAAAAIAAAACwAAAAUA#####wACeEMAATAAAAABAAAAAAAAAAAAAAAFAP####8AAnlDAAE2AAAAAUAYAAAAAAAAAAAADQD#####AQAAAAEQAAFDAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAkAAAAIAAAADQAAAAgAAAAOAAAABQD#####AAN4TTEAATgAAAABQCAAAAAAAAAAAAAFAP####8AA3lNMQABMgAAAAFAAAAAAAAAAAAAAA0A#####wEAAAAAEAACTTEAAAAAAAAAAABACAAAAAAAAAUAAAAACQAAAAgAAAAQAAAACAAAABEAAAANAP####8B#wD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAAAUAAAAAAAAAAAAAAAUAAAAAAAAAA#####wAAAAEADENUcmFuc2xhdGlvbgD#####AAAAEwAAAAEAAAAJAP####8BAAAAABAAAVcAAAAAAAAAAABACAAAAAAAAAUAAAAAEgAAABQAAAAOAP####8AAAABAAAAFQAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAEAAAAWAAAACQD#####AQAAAAAQAAFYAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAwAAAAWAAAACQD#####AQAAAAAQAAFZAAAAAAAAAAAAQAgAAAAAAAAFAAAAAA8AAAAW#####wAAAAEACUNQb2x5Z29uZQD#####AQB#AAABAAAABAAAABcAAAAYAAAAGQAAABf#####AAAAAQAQQ1N1cmZhY2VQb2x5Z29uZQD#####AQB#AAAAAAUAAAAaAAAABQD#####AAJueAABNwAAAAFAHAAAAAAAAAAAAAUA#####wACbnkAATcAAAABQBwAAAAAAAAAAAAMAP####8A5ubmAAEAAAABAAAADAAAAA8AAAAAAAABAAAAAAAAAAAAAAABAAAAAAAAAAAAAAABP#AAAAAAAAAAAAABP#AAAAAAAAAAAAANAP####8BAH8AABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAeAAAACAAAABwAAAAIAAAAHf####8AAAABAAdDTWlsaWV1AP####8B#wAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAABAAAAHwAAAAUA#####wADeE4zAAIyNgAAAAFAOgAAAAAAAAAAAAUA#####wADeU4zAAEyAAAAAUAAAAAAAAAAAAAABQD#####AAN4TTMAAjE2AAAAAUAwAAAAAAAAAAAABQD#####AAN5TTMAAjIyAAAAAUA2AAAAAAAAAAAADQD#####AQAA#wAQAAJOMwAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAACAAAACEAAAAIAAAAIgAAAA0A#####wEAAP8AEAACTTMAAAAAAAAAAABACAAAAAAAAAUAAAAACQAAAAgAAAAjAAAACAAAACT#####AAAAAQALQ01lZGlhdHJpY2UA#####wEAAP8AEAAAAQACAAAAJQAAACYAAAAFAP####8AA3hOMgACMTQAAAABQCwAAAAAAAAAAAAFAP####8AA3lOMgABMgAAAAFAAAAAAAAAAAAAAAUA#####wADeU0yAAIxNAAAAAFALAAAAAAAAAAAAAUA#####wADeE0yAAEyAAAAAUAAAAAAAAAAAAAADQD#####AQB#AAAQAAJOMgAAAAAAAAAAAEAIAAAAAAAABQAAAAAJAAAACAAAACgAAAAIAAAAKQAAAA0A#####wEAfwAAEAACTTIAAAAAAAAAAABACAAAAAAAAAUAAAAACQAAAAgAAAArAAAACAAAACoAAAAFAP####8AA3hOMQACMTYAAAABQDAAAAAAAAAAAAAFAP####8AA3lOMQACMTAAAAABQCQAAAAAAAAAAAANAP####8BAAAAABAAAk4xAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAkAAAAIAAAALgAAAAgAAAAvAAAAEgD#####AQB#AAEQAAQoZDEpAQACP+zMzMzMzM0AAAAwAAAAEv####8AAAABABFDUG9pbnRQYXJBYnNjaXNzZQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAAQAAAAwAAAAHAQAAAAgAAAAcAAAAAT#wAAAAAAAA#####wAAAAEACENTZWdtZW50AP####8AAAAAABAAAAEAAQAAAAEAAAAyAAAAEwD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAAQAAAA8AAAAHAQAAAAgAAAAdAAAAAT#wAAAAAAAAAAAABAD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAQAAAAAAAAAAAAAAMwAAAA4A#####wAAAAEAAAA0AAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAANQAAADYAAAAUAP####8BAAAAABAAAAEAAQAAADUAAAA3AAAABAD#####AQAAAAEQAAFVAAAAAAAAAAAAQAgAAAAAAAAFAAE#0N4k1fxFTQAAADj#####AAAAAgANQ01lc3VyZUFmZml4ZQD#####AAF6AAAAHgAAADn#####AAAAAQANQ0ZvbmNDb21wbGV4ZQD#####AANudW0AEjIqcmUoeikrMipueCppbSh6KQAAAAcAAAAABwIAAAABQAAAAAAAAAD#####AAAAAgAJQ0ZvbmN0aW9uTP####8AAAACABFDVmFyaWFibGVGb3JtZWxsZQAAAAAAAAAHAgAAAAcCAAAAAUAAAAAAAAAA#####wAAAAEAF0NSZXN1bHRhdFZhbGV1ckNvbXBsZXhlAAAAHAAAABdNAAAAGAAAAAAAAXoAAAAPAP####8B#wD#AAEAAAAEAAAAAQAAAAwAAAAPAAAAAf####8AAAACAAhDTWVzdXJlWQD#####AAJ5VQAAAAkAAAA5#####wAAAAIACENNZXN1cmVYAP####8AAnhVAAAACQAAADkAAAAOAP####8AAAABAAAAOQAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAEAAAA#AAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAADAAAAD8AAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAPAAAAPwAAAA8A#####wEAAAAAAQAAAAQAAABAAAAAQQAAAEIAAABA#####wAAAAIAEkNMaWV1T2JqZXRQYXJQdExpZQD#####AAAAAAAAAEMAAAAIAAAAHQAAADkAAAAGAAAAOQAAAD8AAABAAAAAQQAAAEIAAABDAAAAHAD#####AAAAAAAAAEQAAAAIAAAAHAAAADUAAAAKAAAANQAAADcAAAA4AAAAOQAAAD8AAABAAAAAQQAAAEIAAABDAAAARAAAAAoA#####wAAAAwAAAABQE4AAAAAAAAAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAABAAAARgAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAAwAAABGAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAADwAAAEYAAAAPAP####8B#wD#AAQAAAAEAAAARwAAAEgAAABJAAAARwAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAEcAAAA#AAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAASAAAAD8AAAAJAP####8BAAAAABAAAUEAAAAAAAAAAABACAAAAAAAAAUAAAAASQAAAD8AAAAPAP####8BAAAAAAEAAAAEAAAASwAAAEwAAABNAAAASwAAABwA#####wEAAAAAAABOAAAACAAAAB0AAAA5AAAABgAAADkAAAA#AAAASwAAAEwAAABNAAAATgAAABwA#####wAAAAAAAABPAAAACAAAABwAAAA1AAAACgAAADUAAAA3AAAAOAAAADkAAAA#AAAASwAAAEwAAABNAAAATgAAAE8AAAATAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAA5AAAATQAAAAE#1VVHWjGkvgAAABMA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAADkAAABNAAAAAT#lVUdaMaS+AAAABQD#####AAN4VScAG2ludCh4VSoxMDAwMDAwKzAuNSkvMTAwMDAwMAAAAAcDAAAAFwIAAAAHAAAAAAcCAAAACAAAAD4AAAABQS6EgAAAAAAAAAABP+AAAAAAAAAAAAABQS6EgAAAAAAAAAAFAP####8AA3lVJwAbaW50KHlVKjEwMDAwMDArMC41KS8xMDAwMDAwAAAABwMAAAAXAgAAAAcAAAAABwIAAAAIAAAAPQAAAAFBLoSAAAAAAAAAAAE#4AAAAAAAAAAAAAFBLoSAAAAAAP####8AAAABAA9DQ2FsY3VsQ29tcGxleGUA#####wACelUACXhVJytpKnlVJwAAAAcAAAAAGQAAAFMAAAAHAv####8AAAABAAtDQ29uc3RhbnRlaQAAABkAAABUAAAAHQD#####AARudW1VAAZudW0oein#####AAAAAQAWQ0FwcGVsRm9uY3Rpb25Db21wbGV4ZQAAADsAAAAZAAAAOv####8AAAABAA9DVmFsZXVyQWZmaWNoZWUA#####wEAAAAAwDEAAAAAAADAJAAAAAAAAAAAAFEQAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAACAAAAVgAAAB0A#####wAFbnVtVTEABm51bVUrMQAAAAcAAAAAGQAAAFYAAAABP#AAAAAAAAAAAAAgAP####8BAAAAAMAuAAAAAAAAwCAAAAAAAAAAAABSEAAAAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAAgAAAFgAAAAcAP####8AAAAAAAAAVwAAAAgAAAAdAAAAOQAAAAcAAAA5AAAAOgAAAD8AAABNAAAAUQAAAFYAAABXAAAAHAD#####AAAAAAAAAFoAAAAIAAAAHAAAADUAAAALAAAANQAAADcAAAA4AAAAOQAAADoAAAA#AAAATQAAAFEAAABWAAAAVwAAAFoAAAAcAP####8AAAAAAAAAWQAAAAgAAAAdAAAAOQAAAAgAAAA5AAAAOgAAAD8AAABNAAAAUgAAAFYAAABYAAAAWQAAABwA#####wAAAAAAAABcAAAACAAAABwAAAA1AAAADAAAADUAAAA3AAAAOAAAADkAAAA6AAAAPwAAAE0AAABSAAAAVgAAAFgAAABZAAAAXAAAAAkA#####wH#AP8AEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAC0AAAAUAAAADgD#####AAAAAQAAAF4AAAAJAP####8B#wD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAABAAAAXwAAAAkA#####wH#AP8AEAABUQAAAAAAAAAAAEAIAAAAAAAABQAAAAAMAAAAXwAAAAkA#####wH#AP8AEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAA8AAABfAAAADwD#####Af8AAAACAAAABAAAAGAAAABhAAAAYgAAAGAAAAAQAP####8B#wAAAAAABQAAAGMAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAmAAAAFAAAAA4A#####wAAAAEAAABlAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAAQAAAGYAAAAJAP####8BAAAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAMAAAAZgAAAAkA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAA8AAABmAAAACQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAJgAAABQAAAAOAP####8AAAABAAAAEwAAAAkA#####wH#AP8AEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAABMAAABrAAAADgD#####AAAAbAAAACYAAAAJAP####8B#wD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABHAAAAbQAAAAkA#####wH#AP8AEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAEgAAABtAAAACQD#####Af8A#wAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAASQAAAG0AAAAPAP####8BAAD#AAEAAAAEAAAAbgAAAG8AAABwAAAAbgAAABAA#####wEAAP8AAAAFAAAAcQAAAA4A#####wAAAAEAAAAPAAAACQD#####AQAAAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAANAAAAHMAAAAUAP####8AAAAAABAAAAEAAQAAAHQAAAAf#####wAAAAEAEENJbnREcm9pdGVEcm9pdGUA#####wEAAAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAADEAAAAz#####wAAAAIADENDb21tZW50YWlyZQD#####AQB#AABANQAAAAAAAMBFAAAAAAABAAAAdhAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAgoZCNMKDEpKQAAABQA#####wAAAP8AEAAAAQABAAAAAQAAADQAAAAhAP####8BAAD#ABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAB4AAAAJwAAACIA#####wEAAP8AwEAAAAAAAADACAAAAAAAAAAAAHkQAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAIKGQjTCgzKSkAAAAhAP####8BAH8AABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAxAAAAdQAAACIA#####wAAfwAAwESAAAAAAAA#8AAAAAAAAAAAAHsQAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAIKGQjTCgxKSkAAAARAP####8BAH8AABAAAkgyAAAAAAAAAAAAQAgAAAAAAAAFAAAAAC0AAAAs#####wAAAAEACUNEcm9pdGVPbQD#####Af8AAAAQAAABAAIAAAAJAAAAfQAAAAE#8AAAAAAAAAAAACEA#####wH#AAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAH4AAAAzAAAAIgD#####Af8AAADANQAAAAAAAMBCgAAAAAAAAAAAfxAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAgoZCNMKDIpKQAAACEA#####wH#AAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAH4AAAB1AAAAIgD#####Af8AAADAFAAAAAAAAEAmAAAAAAAAAAAAgRAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAgoZCNMKDIpKQAAABEA#####wEAfwAAEAACSDEAAAAAAAAAAABACAAAAAAAAAUAAAAAEgAAADAAAAAUAP####8BAH8AABAAAAEAAgAAABIAAAAwAAAABAD#####AQB#AAEQAANNJzEAAAAAAAAAAABACAAAAAAAAAUAAT#YiIhSPtDkAAAAhP####8AAAABAAlDTG9uZ3VldXIA#####wAAAAEAAAAM#####wAAAAIAD0NNZXN1cmVBYnNjaXNzZQD#####AAJrMQAAAIMAAAASAAAAhQAAAAYA#####wAAAIMAAAAIAAAAhwAAAAkA#####wEAfwAAEAACVycAAAAAAAAAAABACAAAAAAAAAUAAAAAFQAAAIj#####AAAAAQANQ1BvaW50UHJvamV0ZQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAGAAAADEAAAAGAP####8AAACKAAAACAAAAIcAAAAJAP####8BAH8AABAAAlgnAAAAAAAAAAAAQAgAAAAAAAAFAAAAABgAAACLAAAAJgD#####AQB#AAAQAAJZJwAAAAAAAAAAAEAIAAAAAAAABQAAAAAZAAAAMQAAAAYA#####wAAAI0AAAAIAAAAhwAAAAkA#####wEAfwAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAABkAAACOAAAADwD#####AQB#AAACAAAABAAAAIkAAACMAAAAjwAAAIkAAAAQAP####8BAH8AAAAABQAAAJD#####AAAAAQAPQ1N5bWV0cmllQXhpYWxlAP####8AAAAxAAAACQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAFwAAAJIAAAAJAP####8BAH8AABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAAAYAAAAkgAAAAkA#####wEAfwAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAABkAAACSAAAADwD#####AQB#AAACAAAABAAAAJMAAACUAAAAlQAAAJMAAAAQAP####8BAH8AAAAABQAAAJb#####AAAAAgAXQ01hY3JvQW5pbWF0aW9uUG9pbnRMaWUA#####wAAfwAB#####wpAi1QAAAAAAEA#hR64UeuGAgAAAAAAAAAAAAAAAAEAAAAAAAAAAAAFYW5pbTEAAAAAAAAUAAAAZAAAADwAAACFAAEA#####wAAAAEAEUNNYWNyb0Rpc3Bhcml0aW9uAP####8AAH8AAf####8KQIs8AAAAAABAT8KPXCj1wwIAAAAAAAAAAAAAAAABAAAAAAAAAAAAB21hc3F1ZTEAAAAAAAgAAACRAAAAkAAAABsAAAAaAAAAMQAAAHcAAACXAAAAlv####8AAAABABBDTWFjcm9BcHBhcml0aW9uAP####8AAH8AAf####8KQItEAAAAAABAV+FHrhR64gIAAAAAAAAAAAAAAAABAAAAAAAAAAAACWFwcGFyYWl0MQAAAAAACAAAAJEAAACQAAAAGwAAABoAAAAxAAAAdwAAAJcAAACWAP####8AAAABAAtDTWFjcm9QYXVzZQD#####AAB#AAH#####EECN1AAAAAAAQEJCj1wo9cMCAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAVQYXVzZQAAAAAAAf####8AAAABABFDTWFjcm9TdWl0ZU1hY3JvcwD#####AAB#AAH#####EEBTYAAAAAAAQCMKPXCj1wwCAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAxDb3JyZWN0aW9uIGEAAAAAAAQAAACaAAAAmAAAAJkAAACbAAAAFAD#####AQB#AAAQAAABAAIAAAAtAAAALAAAAAQA#####wEAfwAAEAADTScyAAAAAAAAAAAAQAgAAAAAAAAFAAE#089zR#b9RQAAAJ0AAAAlAP####8AAmsyAAAAfQAAAC0AAACeAAAAJgD#####AAB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAYAAAAH4AAAAGAP####8AAACgAAAACAAAAJ8AAAAJAP####8BAH8AABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABgAAAAoQAAACYA#####wEAfwAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAGIAAAB+AAAABgD#####AAAAowAAAAgAAACfAAAACQD#####AQB#AAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAYgAAAKQAAAAJAP####8BAH8AABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABhAAAApAAAAA8A#####wH#AAAAAgAAAAQAAACiAAAApQAAAKYAAACiAAAAEAD#####Af8AAAAAAAUAAACnAAAAKAD#####AP8AAAH#####CkCLlAAAAAAAQGKQo9cKPXECAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAVhbmltMgAAAAAAABQAAABkAAAAPAAAAJ4AAQAAAAAnAP####8AAAB+AAAACQD#####Af8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAYAAAAKoAAAAJAP####8B#wAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABhAAAAqgAAAAkA#####wH#AAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAGIAAACqAAAADwD#####Af8AAAACAAAABAAAAKsAAACsAAAArQAAAKsAAAAQAP####8B#wAAAAAABQAAAK4AAAApAP####8A#wAAAf####8KQIu0AAAAAABAZpCj1wo9cQIAAAAAAAAAAAAAAAABAAAAAAAAAAAAB21hc3F1ZTIAAAAAAAkAAABkAAAAYwAAAKgAAACnAAAArwAAAK4AAAB+AAAAggAAAIAAAAAqAP####8A#wAAAf####8KQIvMAAAAAABAahCj1wo9cQIAAAAAAAAAAAAAAAABAAAAAAAAAAAACWFwcGFyYWl0MgAAAAAACQAAAGQAAABjAAAAqAAAAKcAAACvAAAArgAAAH4AAACCAAAAgAAAAAAsAP####8A#wAAAf####8QQGhwAAAAAABAIwo9cKPXDAIAAAAAAAAAAAAAAAABAAAAAAAAAAAADENvcnJlY3Rpb24gYgAAAAAABAAAALEAAACpAAAAsAAAAJsAAAARAP####8B#wAAABAAAkgzAAAAAAAAAAAAQAgAAAAAAAAFAAAAACYAAAAlAAAAFAD#####Af8AAAAQAAABAAIAAAAmAAAAJQAAAAQA#####wH#AAAAEAADTSczAAAAAAAAAAAAQAgAAAAAAAAFAAE#0spidDyBMgAAALQAAAAlAP####8AAmszAAAAswAAACYAAAC1AAAAJgD#####Af8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAbgAAACcAAAAmAP####8B#wAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABwAAAAJwAAAAYA#####wAAALcAAAAIAAAAtgAAAAkA#####wH#AAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAG4AAAC5AAAABgD#####AAAAuAAAAAgAAAC2AAAACQD#####Af8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAcAAAALsAAAAGAP####8AAACzAAAACAAAALYAAAAJAP####8B#wAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABvAAAAvQAAAA8A#####wEAAP8AAgAAAAQAAAC6AAAAvgAAALwAAAC6AAAAJwD#####AAAAJwAAAAkA#####wH#AAAAEAAAAAAAAAAAAAAAQAgAAAAAAAAFAAAAAG4AAADAAAAACQD#####Af8AAAAQAAAAAAAAAAAAAABACAAAAAAAAAUAAAAAbwAAAMAAAAAJAP####8B#wAAABAAAAAAAAAAAAAAAEAIAAAAAAAABQAAAABwAAAAwAAAAA8A#####wEAAP8AAgAAAAQAAADBAAAAwgAAAMMAAADBAAAAEAD#####AQAA#wAAAAUAAADEAAAAEAD#####AQAA#wAAAAUAAAC#AAAAKAD#####AAAA#wH#####CkCL9AAAAAAAQHAoUeuFHrgCAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAVhbmltMwAAAAAAABQAAABkAAAAPAAAALUAAQAAAAApAP####8AAAD#Af####8KQIwEAAAAAABAcjhR64UeuAIAAAAAAAAAAAAAAAABAAAAAAAAAAAAB21hc3F1ZTMAAAAAAAgAAAByAAAAcQAAAMYAAAC#AAAAxQAAAMQAAAAnAAAAegAAACoA#####wAAAP8B#####wpAjAwAAAAAAEB0CFHrhR64AgAAAAAAAAAAAAAAAAEAAAAAAAAAAAAJYXBwYXJhaXQzAAAAAAAIAAAAcgAAAHEAAADGAAAAvwAAAMUAAADEAAAAJwAAAHoAAAAALAD#####AAAA#wH#####EEBzWAAAAAAAQCUKPXCj1wwCAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAxDb3JyZWN0aW9uIGMAAAAAAAQAAADJAAAAxwAAAJsAAADIAAAALAD#####AAAA#wH#####EECMdAAAAAAAQHb4UeuFHrgCAAAAAAAAAAAAAAAAAQAAAAAAAAAAAApDb3JyZWN0aW9uAAAAAAADAAAAnAAAALIAAADKAAAAhv##########";
-    this.liste_questions = [];
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    let texte = ``,
-      texte_corr = ``;
-    let tabfigA = [],
-      tabfigB = [];
-    let nx = 7,
-      ny = 7,
-      xB = 6,
-      yB = 0,
-      xC = 0,
-      yC = 6,
-      zoom = 8;
-    let xM1, yM1, xN1, yN1, xM2, yM2, xN2, yN2, xM3, yM3, xN3, yN3;
-    let trouver = false,
-      indexa,
-      numa,
-      anum,
-      indexaxe1,
-      xmil1,
-      ymil1,
-      indexb,
-      numb,
-      bnum,
-      indexaxe2,
-      xmil2,
-      ymil2,
-      indexc,
-      numc,
-      cnum,
-      indexaxe3,
-      xmil3,
-      ymil3;
-    let point = [0, 0, 0];
-
-    for (let y = 0; y < ny; y++) {
-      // On initialise les tableaux avec les coordonnées des points de référence (A,B,C et D) de chaque translaté et son numéro dans le pavage.
-      for (let x = 0; x < nx; x++) {
-        tabfigA.push([2 + x * 6, 2 + y * 6, 2 * x + 2 * y * nx]);
-        tabfigB.push([4 + x * 6, 4 + y * 6, 2 * x + 2 * y * nx + 1]);
-      }
-    }
-
-    // Première question : une figure dans tabfigA, une symétrie par rapport à une droite parallèle à y=-x, l'image est dans tabfigB	 !
-    // L'axe de symétrie est la médiatrice du segment d'extrémités les centres de gravités des figures A et B.
-
-    indexa = randint(0, nx * ny - 1);
-    numa = tabfigA[indexa][2];
-    indexaxe1 = randint(0, nx * ny - 1); // l'index aléatoire dans le tableau des centres de gravité des figA sert à choisir un point de passage pour l'axe (xG-2,yG-2).
-    //on calcule les coordonnées du milieu de [BC] on ajoute aux coordonnées du milieu de [BC] celles du vecteur BB'. (j'aurais pu réduire mais cela aurait rendu le calcul plus opaque)
-    xmil1 = tabfigA[indexaxe1][0] - 2;
-    ymil1 = tabfigA[indexaxe1][1] - 2;
-    point = [tabfigA[indexa][0], tabfigA[indexa][1]];
-    point = image_point_par_transformation(2, point, [xmil1, ymil1]);
-
-    trouver = false;
-    while (trouver == false) {
-      for (let j = 0; j < nx * ny; j++) {
-        if (point[0] == tabfigB[j][0] && point[1] == tabfigB[j][1]) {
-          trouver = true;
-          anum = tabfigB[j][2];
-          break;
-        }
-      }
-      if (trouver == false) {
-        indexa = randint(0, nx * ny - 1);
-        numa = tabfigA[indexa][2];
-        indexaxe1 = randint(0, nx * ny - 1);
-        xmil1 = tabfigA[indexaxe1][0] - 2;
-        ymil1 = tabfigA[indexaxe1][1] - 2;
-        point = [tabfigA[indexa][0], tabfigA[indexa][1]];
-        point = image_point_par_transformation(2, point, [xmil1, ymil1]);
-      }
-    }
-    xM1 = tabfigA[indexa][0];
-    yM1 = tabfigA[indexa][1];
-    xN1 = point[0];
-    yN1 = point[1];
-
-    texte = `Le point O peut être déplacé pour recadrer éventuellement le pavage.<br>`;
-    texte +=
-      num_alpha(0) +
-      texte_en_couleur_et_gras(
-        ` Quel est le numéro de la figure symétrique de la figure ${numa} dans la symétrie par rapport à la droite $(d_1)$ ?<br>`,
-        `green`
-      );
-    texte_corr =
-      num_alpha(0) +
-      texte_en_couleur_et_gras(
-        ` La figure symétrique de la figure ${numa} dans la symétrie par rapport à la droite $(d_1)$ porte le numéro ${anum}.<br>`,
-        `green`
-      );
-    // Deuxième question : une figure dans tabfigA, une symétrie par rapport à une parallèle à y=x ql'image est une figure dans tabfigA
-    // On tracera la médiatrice du segment d'extrémité les 2 centres de gravités symétriques.
-    indexb = randint(0, nx * ny - 1);
-    numb = tabfigA[indexb][2];
-    indexaxe2 = randint(0, nx * ny - 1, [indexb]); // l'index aléatoire dans le tableau des centres de gravité des figA sert à choisir un point de passage pour l'axe (xG-2,yG-2).
-    xmil2 = tabfigA[indexaxe2][0] - 2;
-    ymil2 = tabfigA[indexaxe2][1] - 2;
-    point = image_point_par_transformation(
-      1,
-      [tabfigA[indexb][0], tabfigA[indexb][1]],
-      [xmil2, ymil2]
-    );
-    trouver = false;
-    while (trouver == false) {
-      for (let j = 0; j < nx * ny; j++) {
-        if (point[0] == tabfigA[j][0] && point[1] == tabfigA[j][1]) {
-          trouver = true;
-          bnum = tabfigA[j][2];
-          break;
-        }
-      }
-      if (trouver == false) {
-        indexb = randint(0, nx * ny - 1);
-        numb = tabfigA[indexb][2];
-        indexaxe2 = randint(0, nx * ny - 1);
-        xmil2 = tabfigA[indexaxe2][0] - 2;
-        ymil2 = tabfigA[indexaxe2][1] - 2;
-        point = image_point_par_transformation(
-          1,
-          [tabfigA[indexb][0], tabfigA[indexb][1]],
-          [xmil2, ymil2]
-        );
-      }
-    }
-    xM2 = tabfigA[indexb][0];
-    yM2 = tabfigA[indexb][1];
-    xN2 = point[0];
-    yN2 = point[1];
-
-    texte +=
-      num_alpha(1) +
-      texte_en_couleur_et_gras(
-        ` Quel est le numéro de la figure symétrique de la figure ${numb} dans la symétrie par rapport à la droite $(d_2)$ ?<br>`,
-        `red`
-      );
-    texte_corr +=
-      num_alpha(1) +
-      texte_en_couleur_et_gras(
-        ` La figure symétrique de la figure ${numb} dans la symétrie par rapport à la droite $(2)$ porte le numéro ${bnum}.<br>`,
-        `red`
-      );
-
-    // troisième question : une figure dans tabfigB, une symétrie par rapport à une parallèle à y=0, l'image est dans tabfigA
-    indexc = 3; //randint(0,nx*ny-1)
-    numc = tabfigB[indexc][2];
-    indexaxe3 = 14; //randint(0,nx*ny-1) // l'index aléatoire dans le tableau des centres de gravité des figA sert à choisir un point de passage pour l'axe (xG-2,yG-2).
-    xmil3 = tabfigA[indexaxe3][0] - 2;
-    ymil3 = tabfigA[indexaxe3][1] - 2;
-    point = [tabfigB[indexc][0], tabfigB[indexc][1]];
-    point = changement_de_base_ortho_tri(
-      image_point_par_transformation(
-        3,
-        changement_de_base_tri_ortho(point),
-        changement_de_base_tri_ortho([xmil3, ymil3])
-      )
-    );
-    point[0] = arrondi(point[0], 1); // Les coordonnées sont censées être entières mais les calculs JS laissent toujours de cent-milliardièmes indésirables.
-    point[1] = arrondi(point[1], 1);
-    trouver = false;
-    while (trouver == false) {
-      for (let j = 0; j < nx * ny; j++) {
-        if (point[0] == tabfigA[j][0] && point[1] == tabfigA[j][1]) {
-          trouver = true;
-          cnum = tabfigA[j][2];
-          break;
-        }
-      }
-      if (trouver == false) {
-        indexc = randint(0, nx * ny - 1);
-        numc = tabfigB[indexc][2];
-        indexaxe3 = randint(0, nx * ny - 1);
-        xmil3 = tabfigA[indexaxe3][0] - 2;
-        ymil3 = tabfigA[indexaxe3][1] - 2;
-        point = [tabfigA[indexa][0], tabfigA[indexa][1]];
-        point = changement_de_base_ortho_tri(
-          image_point_par_transformation(
-            3,
-            changement_de_base_tri_ortho(point),
-            [xmil3, ymil3]
-          )
-        );
-        point[0] = arrondi(point[0], 1);
-        point[1] = arrondi(point[1], 1);
-      }
-    }
-
-    xM3 = tabfigB[indexc][0];
-    yM3 = tabfigB[indexc][1];
-    xN3 = point[0];
-    yN3 = point[1];
-
-    texte +=
-      num_alpha(2) +
-      texte_en_couleur_et_gras(
-        ` Quel est le numéro de la figure symétrique de la figure ${numc} dans la symétrie par rapport à la droite $(d_3)$ ?<br>`,
-        `blue`
-      );
-    texte_corr +=
-      num_alpha(2) +
-      texte_en_couleur_et_gras(
-        ` La figure symétrique de la figure ${numc} dans la symétrie par rapport à la droite $(d_3)$ porte le numéro ${cnum}.<br>`,
-        `blue`
-      );
-
-    if (sortie_html) {
-      this.MG32code_pour_modifier_la_figure = `
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xB", "${xB}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yB", "${yB}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xC", "${xC}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yC", "${yC}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xM1", "${xM1}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yM1", "${yM1}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xN1", "${xN1}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yN1", "${yN1}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xM2", "${xM2}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yM2", "${yM2}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xN2", "${xN2}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yN2", "${yN2}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xM3", "${xM3}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yM3", "${yM3}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "xN3", "${xN3}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "yN3", "${yN3}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "nx", "${nx}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "ny", "${ny}");
-			mtg32App.giveFormula2("MG32svg${numero_de_l_exercice}", "zoom", "${zoom}");
-
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xB", "${xB}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yB", "${yB}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xC", "${xC}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yC", "${yC}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xM1", "${xM1}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yM1", "${yM1}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xN1", "${xN1}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yN1", "${yN1}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xM2", "${xM2}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yM2", "${yM2}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xN2", "${xN2}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yN2", "${yN2}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xM3", "${xM3}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yM3", "${yM3}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "xN3", "${xN3}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "yN3", "${yN3}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "nx", "${nx}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "ny", "${ny}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "nx", "${nx}");
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "ny", "${ny}");	
-			mtg32App.giveFormula2("MG32svgcorr${numero_de_l_exercice}", "zoom", "${zoom}");
-
-			mtg32App.calculate("MG32svg${numero_de_l_exercice}");
-			mtg32App.display("MG32svg${numero_de_l_exercice}");
-			mtg32App.calculate("MG32svgcorr${numero_de_l_exercice}");
-			mtg32App.executeMacro("MG32svgcorr${numero_de_l_exercice}","Correction")
-			mtg32App.display("MG32svgcorr${numero_de_l_exercice}");
-			`;
-
-      this.liste_questions.push(texte);
-      this.liste_corrections.push(texte_corr);
-      liste_de_question_to_contenu_sans_numero(this);
-    } else {
-      texte = ``;
-      texte_cor = ``;
-      this.liste_questions.push(texte); // on envoie la question
-      this.liste_corrections.push(texte_corr);
-      liste_de_question_to_contenu_sans_numero(this);
-    }
-  };
-  // this.besoin_formulaire_numerique = ['Transformations',5, '1 : Symétries axiales\n 2 : Symétries centrales\n 3 : Rotations\n 4 : Translations\n 5 : Homothéties\n'];
-}
-
-/**
- * Construction de symétrique avec dispositif d'auto-correction aléatoire
- * Ref 6G24-3 
- * @Auteur Jean-Claude Lhote
- * Publié le 30/11/2020
- */
-function Construire_symetrique_point_6e() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Construire le symétrique d'un point avec cible auto-corrective";
-  this.consigne = "";
-  this.nb_questions = 1;
-  this.nb_questions_modifiable = false
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 3;
-  this.nouvelle_version = function () {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let result = [0, 0], texte_corr = "", nbpoints = parseInt(this.sup)
-    let celluleAlea = function (rang) {
-      let lettre = lettre_depuis_chiffre(randint(1, rang))
-      let chiffre = Number(randint(1, rang)).toString()
-      return lettre + chiffre
-    }
-    // On prépare la figure...
-    let a = randint(-10, 10), b = randint(-10, 10, a)
-    let d = droite(a, b, 0, '(d)')
-    let A = translation(point(0, 0), homothetie(d.directeur, point(0, 0), -0.5))
-    let B = translation(point(0, 0), homothetie(d.directeur, point(0, 0), 0.5))
-    let marks = ['/', '//', '///', 'x', 'o', 'S', 'V']
-    let noms = choisit_lettres_differentes(nbpoints, 'Q', majuscule = true)
-    this.consigne = `Construire le symétrique des points $${noms[0]}$`
-    for (let i = 1; i < nbpoints - 1; i++) {
-      this.consigne += `, $${noms[i]}$`
-    }
-    this.consigne += ` et $${noms[nbpoints - 1]}$ par rapport à $(d)$.`;
-    let cibles = [], M = [], N = [], objets_enonce = [], objets_correction = []  //cibles, M point marqués, N symétrique de M
-    let cellules = []
-    let xMin, yMin, xMax, yMax
-    [xMin, yMin, xMax, yMax] = [0, 0, 0, 0]
-    for (let i = 0; i < nbpoints; i++) { //On place les cibles.
-      N.push(point(calcul(randint(-80, 80, 0) / 10), calcul(randint(-80, 80, 0) / 10), noms[i] + "\'"))
-      nontrouve = true
-      while (distancePointDroite(N[i], d) < 3 || nontrouve) {
-        nontrouve = true
-        if (distancePointDroite(N[i], d) < 3) {
-          N[i].x = calcul(randint(-80, 80, 0) / 10)
-          N[i].y = calcul(randint(-80, 80, 0) / 10)
-        }
-        else {
-          assezloin = true
-          for (let j = 0; j < i; j++) {
-            if (longueur(N[i], N[j]) < 4.5) assezloin = false
-          }
-          if (assezloin == false) {//éloigner les points donc les grilles
-            N[i].x = calcul(randint(-80, 80, 0) / 10)
-            N[i].y = calcul(randint(-80, 80, 0) / 10)
-          }
-          else nontrouve = false
-        }
-      }
-    }
-
-    objets_enonce.push(d)
-    objets_correction.push(d, tracePoint(A, B))
-
-    for (let i = 0; i < nbpoints; i++) {
-      cellules.push(celluleAlea(4))
-      result = dansLaCibleCarree(N[i].x, N[i].y, 4, 0.6, cellules[i])
-      cible = cibleCarree({ x: result[0], y: result[1], rang: 4, num: i + 1, taille: 0.6 })
-      cible.taille = 0.6
-      cible.color = 'orange'
-      cible.opacite = 0.7
-      cibles.push(cible)
-    }
-    for (let i = 0; i < nbpoints; i++) {
-      M.push(symetrieAxiale(N[i], d, noms[i]))
-      objets_enonce.push(tracePoint(M[i]), labelPoint(M[i]), cibles[i])
-      objets_correction.push(tracePoint(M[i], N[i]), labelPoint(M[i], N[i]), cibles[i])
-      objets_correction.push(segment(M[i], N[i], arcenciel(i)), codageMediatrice(M[i], N[i], arcenciel(i + 5), marks[i]))
-      objets_correction.push(traceCompas(A, N[i], 20), traceCompas(B, N[i], 20))
-      texte_corr += `$${noms[i]}\'$, le symétrique du point $${noms[i]}$ est dans la case ${cellules[i]} de la grille ${i + 1}.<br>`
-    }
-
-    for (let i = 0; i < nbpoints; i++) {
-      xMin = Math.min(xMin, N[i].x - 3, M[i].x - 3)
-      yMin = Math.min(yMin, N[i].y - 3, M[i].y - 3)
-      xMax = Math.max(xMax, N[i].x + 3, M[i].x + 3)
-      yMax = Math.max(yMax, N[i].y + 3, M[i].y + 3)
-    }
-
-    fenetreMathalea2d = [xMin, yMin, xMax, yMax]
-
-    this.liste_questions.push(mathalea2d({ xmin: xMin, ymin: yMin, xmax: xMax, ymax: yMax, pixelsParCm: 20, scale: 0.7 }, objets_enonce))
-    this.liste_corrections.push(texte_corr + mathalea2d({ xmin: xMin, ymin: yMin, xmax: xMax, ymax: yMax, pixelsParCm: 20, scale: 0.7 }, objets_correction))
-    liste_de_question_to_contenu(this)
-
-    //  let nonchoisi,coords=[],x,y,objets_enonce=[],objets_correction=[],nomd,label_pos
-
-  }
-  this.besoin_formulaire_numerique = ['Nombre de points (1 à 5)', 5, "1\n2\n3\n4\n5"];
-  // this.besoin_formulaire2_case_a_cocher = ["Avec des points de part et d'autre"];	
-}
-/**
- * Construction de médiatrices avec dispositif d'auto-correction aléatoire
- * Ref 6G25
- * @Auteur Jean-Claude Lhote
- * Publié le 30/11/2020
- */
-function Construire_mediatrices_6e() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Construire des médiatrices avec cible auto-corrective";
-  this.consigne = "";
-  this.nb_questions = 1;
-  this.nb_questions_modifiable = false
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 3;
-  this.nouvelle_version = function () {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let result = [0, 0], texte_corr = "", texte = "", num1, num2
-    if (sortie_html) {
-      num1 = `<tspan dy="5" style="font-size:70%">`
-      num2 = `</tspan><tspan dy="-5">)</tspan>`
-    }
-    else {
-      num1 = `_`
-      num2 = `)`
-    }
-    let celluleAlea = function (rang) {
-      let lettre = lettre_depuis_chiffre(randint(1, rang))
-      let chiffre = Number(randint(1, rang)).toString()
-      return lettre + chiffre
-    }
-    // On prépare la figure...
-    let noms = choisit_lettres_differentes(4, 'QI', majuscule = true)
-    texte = `Construire la médiatrice $(d_1)$ du segment $[${noms[0]}${noms[1]}]$ et la médiatrice $(d_2)$ du segment $[${noms[2]}${noms[3]}]$.<br>`
-    texte += `Prolonger les droites $(d_1)$ et $(d_2)$ pour obtenir leur point d'intersection.<br>`
-    let marks = ['/', '//', '///', 'x', 'o', 'S', 'V']
-    let I = point(0, 0, 'I')
-    let A = pointAdistance(I, randint(3, 6))
-    let B = similitude(A, I, randint(65, 150), randint(8, 15) / 10)
-    let medA = droite(I, A, `(d${num1}1${num2}`), medB = droite(I, B, `(d${num1}2${num2}`)
-
-    let dA = droiteParPointEtPerpendiculaire(A, medA)
-    let dB = droiteParPointEtPerpendiculaire(B, medB)
-    medA.color = 'blue'
-    medB.color = 'green'
-    let cA = cercle(A, calcul(randint(25, 40) / 20))
-    let cB = cercle(B, calcul(randint(45, 60) / 20))
-    let A1 = pointIntersectionLC(dA, cA, noms[0], 1)
-    let A2 = pointIntersectionLC(dA, cA, noms[1], 2)
-    let B1 = pointIntersectionLC(dB, cB, noms[2], 1)
-    let B2 = pointIntersectionLC(dB, cB, noms[3], 2)
-    let sA = segmentAvecExtremites(A1, A2)
-    let sB = segmentAvecExtremites(B1, B2)
-    sA.color = 'black'
-    sB.color = 'black'
-
-
-    let cible, objets_enonce = [], objets_correction = [], cellule
-    let xMin, yMin, xMax, yMax
-    let nomA1 = texteParPoint(noms[0], homothetie(A1, A2, 1.1), 'milieu', 'black', 1, '', true)
-    let nomA2 = texteParPoint(noms[1], homothetie(A2, A1, 1.1), 'milieu', 'black', 1, '', true)
-    let nomB1 = texteParPoint(noms[2], homothetie(B1, B2, 1.1), 'milieu', 'black', 1, '', true)
-    let nomB2 = texteParPoint(noms[3], homothetie(B2, B1, 1.1), 'milieu', 'black', 1, '', true)
-
-    cellule = celluleAlea(6)
-    result = dansLaCibleCarree(I.x, I.y, 6, 0.6, cellule)
-    cible = cibleCarree({ x: result[0], y: result[1], rang: 6, taille: 0.6 })
-    cible.taille = 0.6
-    cible.color = 'orange'
-    cible.opacite = 0.7
-
-    objets_enonce.push(cible, sA, sB, nomA1, nomA2, nomB1, nomB2)
-    objets_correction.push(cible, sA, sB, tracePoint(I), labelPoint(I), nomA1, nomA2, nomB1, nomB2)
-    objets_correction.push(medA, medB, codageMediatrice(A1, A2, 'blue', marks[1]), codageMediatrice(B1, B2, 'green', marks[2]))
-
-    //      objets_correction.push(segment(M[i],N[i],arcenciel(i)),codageMediatrice(M[i],N[i],arcenciel(i+5),marks[i])) 
-    //      objets_correction.push(traceCompas(A1,N[i],20),traceCompas(B,N[i],20))
-    texte_corr += `Le point $I$ d'intersection des deux médiatrices est dans la case ${cellule} de la grille.<br>`
-    xMin = Math.min(A1.x - 1, A2.x - 1, B1.x - 1, B2.x - 1, I.x - 4)
-    yMin = Math.min(A1.y - 1, A2.y - 1, B1.y - 1, B2.y - 1, I.y - 4)
-    xMax = Math.max(A1.x + 1, A2.x + 1, B1.x + 1, B2.x + 1, I.x + 4)
-    yMax = Math.max(A1.y + 1, A2.y + 1, B1.y + 1, B2.y + 1, I.y + 4)
-
-    fenetreMathalea2d = [xMin, yMin, xMax, yMax]
-
-    this.liste_questions.push(texte + mathalea2d({ xmin: xMin, ymin: yMin, xmax: xMax, ymax: yMax, pixelsParCm: 20, scale: 0.7 }, objets_enonce))
-    this.liste_corrections.push(texte_corr + mathalea2d({ xmin: xMin, ymin: yMin, xmax: xMax, ymax: yMax, pixelsParCm: 20, scale: 0.7 }, objets_correction))
-    liste_de_question_to_contenu(this)
-
-    //  let nonchoisi,coords=[],x,y,objets_enonce=[],objets_correction=[],nomd,label_pos
-
-  }
-}
-
-/**
- * Ref 6G32-1
- * Publié le 26/10/2020
- * @Auteur Jean-Claude Lhote
- */
-function Symetrie_axiale_conservation1() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Propriétés de conservation de la symétrie axiale";
-  this.consigne = "";
-  this.nb_questions = 4;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 1;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let type_de_questions_disponibles = ["Segment", "Droite", "1/2droite", "Triangle", "Angle"];
-    let points = [], traces = [], nom = [], alternance
-    for (let i = 0; i < 25; i++) nom.push(lettre_depuis_chiffre(i + 1))
-    let noms = shuffle(nom)
-
-    let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions);
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    this.liste_questions.push(`${texte_gras('Dans la symétrie d\'axe (d)...')}`);
-    this.liste_corrections.push(`${texte_gras('Dans la symétrie d\'axe (d)...')}`);
-    // On prépare la figure...
-    let axe = parseInt(this.sup)
-    let d, nonchoisi, coords = [], x, y, objets_enonce = [], objets_correction = [], nomd, label_pos
-    if (axe == 5) axe = randint(1, 4) //choix de l'axe et des coordonnées
-    switch (axe) {
-      case 1: d = droite(1, 0, 0);
-        nomd = texteParPosition('(d)', 0.3, 5.6)
-        label_pos = 'above left'
-        for (let i = 0; i < 12; i++) {
-          nonchoisi = false
-          while (!nonchoisi) { // Le nouveau point est-il déjà dans la liste ?
-            [x, y] = [randint(-5, 0), randint(-5, 5)]
-            nonchoisi = true
-            for (let j = 0; j < i; j++)
-              if (coords[j][0] == x && coords[j][1] == y) nonchoisi = false
-          }
-          coords.push([x, y]) //on stocke les 12 points
-        }
-        for (let j = 0; j < 12; j++) coords.push([-coords[j][0], coords[j][1]]) // on stocke les 12 images
-        break;
-      case 2: d = droite(0, 1, 0);
-        label_pos = 'above'
-        nomd = texteParPosition('(d)', 5.6, 0.3)
-        for (let i = 0; i < 12; i++) {
-          nonchoisi = false
-          while (!nonchoisi) { // Le nouveau point est-il déjà dans la liste ?
-            [x, y] = [randint(-5, 5), randint(-5, 0)]
-            nonchoisi = true
-            for (let j = 0; j < i; j++)
-              if (coords[j][0] == x && coords[j][1] == y) nonchoisi = false
-          }
-          coords.push([x, y]) //on stocke les 12 points
-        }
-        for (let j = 0; j < 12; j++) coords.push([coords[j][0], -coords[j][1]]) // on stocke les 12 images
-        break;
-      case 3: d = droite(1, -1, 0);
-        label_pos = 'above'
-        nomd = texteParPosition('(d)', -5.8, -5.4)
-        for (let i = 0; i < 12; i++) {
-          nonchoisi = false
-          while (!nonchoisi) { // Le nouveau point est-il déjà dans la liste ?
-            x = randint(-5, 5)
-            y = randint(x, 5)
-            nonchoisi = true
-            for (let j = 0; j < i; j++)
-              if (coords[j][0] == x && coords[j][1] == y) nonchoisi = false
-          }
-          coords.push([x, y]) //on stocke les 12 points
-        }
-        for (let j = 0; j < 12; j++) coords.push([coords[j][1], coords[j][0]]) // on stocke les 12 images
-        break;
-      case 4: d = droite(1, 1, 0);
-        label_pos = 'above'
-        nomd = texteParPosition('(d)', -5.8, 5.4)
-        for (let i = 0; i < 12; i++) {
-          nonchoisi = false
-          while (!nonchoisi) { // Le nouveau point est-il déjà dans la liste ? Si oui, on recommence.
-            x = randint(-5, 5)
-            y = randint(-5, -x)
-            nonchoisi = true
-            for (let j = 0; j < i; j++)
-              if (coords[j][0] == x && coords[j][1] == y)
-                nonchoisi = false;
-          }
-          coords.push([x, y]) //on stocke les 12 points
-        }
-        for (let j = 0; j < 12; j++)
-          coords.push([-coords[j][1], -coords[j][0]]); // on stocke les 12 images
-        break;
-    }
-    for (let i = 0; i < 24; i++) {
-      if (i < 12) points.push(point(coords[i][0], coords[i][1], noms[i], label_pos))
-      else if (coords[i][0] == coords[i - 12][0] && coords[i][1] == coords[i - 12][1]) {
-        points.push(point(coords[i][0], coords[i][1], noms[i - 12], label_pos))
-        noms[i] = noms[i - 12]
-      }
-      else points.push(point(coords[i][0], coords[i][1], noms[i], label_pos))
-      traces.push(tracePoint(points[i]));
-    }
-    // On rédige les questions et les réponses
-    if (this.sup2 == true) alternance = 2
-    else alternance = 1
-    function index(i) {
-      return (i + 12 * (i % alternance)) % 24
-    }
-    objets_enonce.length = 0
-    objets_correction.lenght = 0
-    for (let i = 0, texte, texte_corr, choix, cpt = 0; i < this.nb_questions && cpt < 50;) {
-
-      switch (liste_type_de_questions[i]) {
-        case "Segment":
-          choix = randint(0, 10) + randint(0, 1) * 12
-          texte = `Quel est le symétrique du segment $[${noms[index(choix)]}${noms[index(choix + 1)]}]$ ?`
-          texte_corr = `Le symétrique du segment $[${noms[index(choix)]}${noms[index(choix + 1)]}]$ est le segment $[${noms[index(choix + 12)]}${noms[index(choix + 13)]}]$.`
-          s1 = segment(points[index(choix)], points[index(choix + 1)], texcolors(i * 3 + 2))
-          s2 = segment(points[index(choix + 12)], points[index(choix + 13)], texcolors(i * 3 + 2))
-          s1.epaisseur = 2
-          s2.epaisseur = 2
-          objets_correction.push(s1, s2)
-          break;
-        case "Droite":
-          choix = randint(0, 10) + randint(0, 1) * 12
-          texte = `Quel est la symétrique de la droite $(${noms[index(choix)]}${noms[index(choix + 1)]})$ ?`
-          texte_corr = `La symétrique de la droite $(${noms[index(choix)]}${noms[index(choix + 1)]})$ est la droite $(${noms[index(choix + 12)]}${noms[index(choix + 13)]})$.`
-          objets_correction.push(droite(points[index(choix)], points[index(choix + 1)], "", texcolors(i * 3 + 2)))
-          objets_correction.push(droite(points[index(choix + 12)], points[index(choix + 13)], "", texcolors(i * 3 + 2)))
-          break;
-        case "1/2droite":
-          choix = randint(0, 10) + randint(0, 1) * 12
-          texte = `Quel est la symétrique de la demi-droite $[${noms[index(choix)]}${noms[index(choix + 1)]})$ ?`
-          texte_corr = `La symétrique de la demi-droite $[${noms[index(choix)]}${noms[index(choix + 1)]})$ est la demi-droite $[${noms[index(choix + 12)]}${noms[index(choix + 13)]})$`
-          objets_correction.push(demiDroite(points[index(choix)], points[index(choix + 1)], texcolors(i * 3 + 2)))
-          objets_correction.push(demiDroite(points[index(choix + 12)], points[index(choix + 13)], texcolors(i * 3 + 2)))
-          break;
-        case "Triangle":
-          choix = randint(0, 9) + randint(0, 1) * 12
-          texte = `Quel est le symétrique du triangle $${noms[index(choix)]}${noms[index(choix + 1)]}${noms[index(choix + 2)]}$ ?`
-          texte_corr = `Le symétrique du triangle $${noms[index(choix)]}${noms[index(choix + 1)]}${noms[index(choix + 2)]}$ est le triangle $${noms[index(choix + 12)]}${noms[index(choix + 13)]}${noms[index(choix + 14)]}$.`
-          objets_correction.push(polygone([points[index(choix)], points[index(choix + 1)], points[index(choix + 2)]], texcolors(i * 3 + 2)))
-          objets_correction.push(polygone([points[index(choix + 12)], points[index(choix + 13)], points[index(choix + 14)]], texcolors(i * 3 + 2)))
-          break;
-        case "Angle":
-          choix = randint(0, 9) + randint(0, 1) * 12
-          texte = `Quel est le symétrique de l'angle $\\widehat{${noms[index(choix)]}${noms[index(choix + 1)]}${noms[index(choix + 2)]}}$ ?`
-          texte_corr = `Le symétrique de l'angle $\\widehat{${noms[index(choix)]}${noms[index(choix + 1)]}${noms[index(choix + 2)]}}$ est l'angle $\\widehat{${noms[index(choix + 12)]}${noms[index(choix + 13)]}${noms[index(choix + 14)]}}$.`
-          objets_correction.push(codeAngle(points[index(choix)], points[index(choix + 1)], points[index(choix + 2)], 2, '', texcolors(i * 3 + 2), 2, 0.5, texcolors(i * 3 + 2), 0.2))
-          objets_correction.push(codeAngle(points[index(choix + 12)], points[index(choix + 13)], points[index(choix + 14)], 2, '', texcolors(i * 3 + 2), 2, 0.5, texcolors(i * 3 + 2), 0.2))
-          objets_correction.push(segment(points[index(choix)], points[index(choix + 1)], texcolors(i * 3 + 2)))
-          objets_correction.push(segment(points[index(choix + 1)], points[index(choix + 2)], texcolors(i * 3 + 2)))
-          objets_correction.push(segment(points[index(choix + 12)], points[index(choix + 13)], texcolors(i * 3 + 2)))
-          objets_correction.push(segment(points[index(choix + 13)], points[index(choix + 14)], texcolors(i * 3 + 2)))
-
-          break;
-      }
-
-      if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    d.isVisible = true;
-    objets_enonce.push(nomd, d);
-    objets_correction.push(nomd, d);
-    for (let i = 0; i < 24; i++) {
-      objets_enonce.push(labelPoint(points[i]), tracePoint(points[i], 'blue'))
-      objets_correction.push(labelPoint(points[i]), tracePoint(points[i], 'blue'))
-
-    }
-    this.liste_questions.push(mathalea2d({ xmin: -6, ymin: -6, xmax: 6, ymax: 6, pixelsParCm: 40, scale: 1 }, objets_enonce))
-    this.liste_corrections.push(mathalea2d({ xmin: -6, ymin: -6, xmax: 6, ymax: 6, pixelsParCm: 40, scale: 1 }, objets_correction))
-    liste_de_question_to_contenu_sans_numero(this);
-
-  }
-  this.besoin_formulaire_numerique = ['Type d\'axe', 5, "1 : Axe vertical\n2 : Axe horizontal\n3 : Axe oblique 1\n4 : Axe oblique 2\n5 : Axe aléatoire"];
-  this.besoin_formulaire2_case_a_cocher = ["Avec des points de part et d'autre"];
-}
-
-
-/**
- * Reconnaitre un quadrilatère particulier à partir de ses propriétés
- * @Auteur Rémi Angot
- * Référence 6G33
-*/
-function Reconnaitre_quadrilatere_particulier() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Reconnaitre un quadrilatère particulier à partir de ses propriétés";
-  this.consigne = "";
-  this.nb_questions = 3;
-  this.nb_questions_modifiable = false;
-  this.nb_cols = 2; // Nombre de colonnes pour la sortie LaTeX
-  this.nb_cols_corr = 2; // Nombre de colonnes dans la correction pour la sortie LaTeX
-  this.correction_detaillee_disponible = true;
-  sortie_html ? this.correction_detaillee = true : this.correction_detaillee = false
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    let liste_de_questions = shuffle([choice(['losange1','losange2']),choice(['rectangle1','rectangle2']),choice(['carre1','carre2','carre3'])])
-    for (let i = 0, texte, texte_corr,cpt = 0; i < this.nb_questions && cpt < 50;)
-     {
-      texte = '';
-      texte_corr = '';
-      let A,B,C,D,O,ABCD,codage,codage1,codage2,codage3,codage4,sAC,sBD,sOA,sOB,sOC,sOD,marquesDemiDiagonales,marquesDemiDiagonales1,marquesDemiDiagonales2;
-      switch (liste_de_questions[i]) {
-          case 'losange1':
-              texte = "Quelle est la nature d'un quadrilatère ayant 4 côtés de même longueur ?";
-              A = point(0, 0);
-              B = point(2, 3);
-              C = point(0, 6);
-              D = point(-2, 3);
-              O = point(0, 3);
-              ABCD = polygone(A, B, C, D);
-              //codage = codageAngleDroit(C, O, B);
-              marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              // sAC = segment(A, C);
-              // sBD = segment(B, D);
-              // sOA = segment(O, A);
-              // sOB = segment(O, B);
-              // sOC = segment(O, C);
-              // sOD = segment(O, D);
-              // sAC.pointilles = true;
-              // sBD.pointilles = true;
-              // marquesDemiDiagonales = codeSegments("|", "blue", O, A, O, B, O, C, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-3,xmax:3,ymin:-1,ymax:7},ABCD,marquesCotes)+"<br>"}
-              texte_corr += "C'est un losange."
-              break;
-          case 'losange2':
-              texte = "Quelle est la nature d'un quadrilatère ayant ses diagonales perpendiculaires et sécantes en leur milieu ?";
-              A = point(0, 0);
-              B = point(2, 3);
-              C = point(0, 6);
-              D = point(-2, 3);
-              O = point(0, 3);
-              ABCD = polygone(A, B, C, D);
-              codage = codageAngleDroit(C, O, B);
-              //marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              sAC = segment(A, C);
-              sBD = segment(B, D);
-              sOA = segment(O, A);
-              sOB = segment(O, B);
-              sOC = segment(O, C);
-              sOD = segment(O, D);
-              sAC.pointilles = true;
-              sBD.pointilles = true;
-              marquesDemiDiagonales1 = codeSegments("|", "blue", O, A, O, C);
-              marquesDemiDiagonales2 = codeSegments("|||", "blue", O, B, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-3,xmax:3,ymin:-1,ymax:7},ABCD,codage,sAC,sBD,marquesDemiDiagonales1,marquesDemiDiagonales2)+"<br>"}
-              texte_corr += "C'est un losange."
-              break;
-          case 'rectangle1':
-              texte = "Quelle est la nature d'un quadrilatère ayant 3 angles droits ?";
-              A = point(0, 0);
-              B = point(5, 0);
-              C = point(5, 3);
-              D = point(0, 3);
-              O = point(2.5, 1.5);
-              ABCD = polygone(A, B, C, D);
-              codage1 = codageAngleDroit(A,B,C);
-              codage2 = codageAngleDroit(B,C,D);
-              codage3 = codageAngleDroit(C,D,A);
-              //marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              // sAC = segment(A, C);
-              // sBD = segment(B, D);
-              // sOA = segment(O, A);
-              // sOB = segment(O, B);
-              // sOC = segment(O, C);
-              // sOD = segment(O, D);
-              // sAC.pointilles = true;
-              // sBD.pointilles = true;
-              // marquesDemiDiagonales = codeSegments("||", "blue", O, A, O, B, O, C, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-1,xmax:6,ymin:-1,ymax:4},ABCD,codage1,codage2,codage3)+"<br>"}
-              texte_corr += "C'est un rectangle."
-              break;
-          case 'rectangle2':
-              texte = "Quelle est la nature d'un quadrilatère ayant ses diagonales de même longueur et sécantes en leur milieu ?";
-              A = point(0, 0);
-              B = point(5, 0);
-              C = point(5, 3);
-              D = point(0, 3);
-              O = point(2.5, 1.5);
-              ABCD = polygone(A, B, C, D);
-              // codage1 = codageAngleDroit(A,B,C);
-              // codage2 = codageAngleDroit(B,C,D);
-              // codage3 = codageAngleDroit(C,D,A);
-              //marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              sAC = segment(A, C);
-              sBD = segment(B, D);
-              // sOA = segment(O, A);
-              // sOB = segment(O, B);
-              // sOC = segment(O, C);
-              // sOD = segment(O, D);
-              // sAC.pointilles = true;
-              // sBD.pointilles = true;
-              marquesDemiDiagonales = codeSegments("||", "blue", O, A, O, B, O, C, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-1,xmax:6,ymin:-1,ymax:4},ABCD,marquesDemiDiagonales,sAC,sBD)+"<br>"}
-              texte_corr += "C'est un rectangle."
-              break;
-          case 'carre1':
-              texte = "Quelle est la nature d'un quadrilatère ayant ses 4 côtés de même longueur et 4 angles droits ?";
-              A = point(0, 0);
-              B = point(3, 0);
-              C = point(3, 3);
-              D = point(0, 3);
-              O = point(1.5, 1.5);
-              ABCD = polygone(A, B, C, D);
-              codage1 = codageAngleDroit(A,B,C);
-              codage2 = codageAngleDroit(B,C,D);
-              codage3 = codageAngleDroit(C,D,A);
-              codage4 = codageAngleDroit(D,A,B);
-              marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              // sAC = segment(A, C);
-              // sBD = segment(B, D);
-              // sOA = segment(O, A);
-              // sOB = segment(O, B);
-              // sOC = segment(O, C);
-              // sOD = segment(O, D);
-              // sAC.pointilles = true;
-              // sBD.pointilles = true;
-              // marquesDemiDiagonales = codeSegments("||", "blue", O, A, O, B, O, C, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-1,xmax:4,ymin:-1,ymax:4},ABCD,codage1,codage2,codage3,marquesCotes)+"<br>"}
-              texte_corr += "C'est un carré."
-              break;
-          case 'carre2':
-              texte = "Quelle est la nature d'un quadrilatère ayant ses ses diagonales perpendiculaires, de même longueur et sécantes en leur milieu ?";
-              A = point(0, 0);
-              B = point(3, 0);
-              C = point(3, 3);
-              D = point(0, 3);
-              O = point(1.5, 1.5);
-              ABCD = polygone(A, B, C, D);
-              codage = codageAngleDroit(C,O,D);
-              // codage2 = codageAngleDroit(B,C,D);
-              // codage3 = codageAngleDroit(C,D,A);
-              // codage4 = codageAngleDroit(D,A,B);
-              // marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              sAC = segment(A, C);
-              sBD = segment(B, D);
-              sOA = segment(O, A);
-              sOB = segment(O, B);
-              sOC = segment(O, C);
-              sOD = segment(O, D);
-              sAC.pointilles = true;
-              sBD.pointilles = true;
-              marquesDemiDiagonales = codeSegments("||", "blue", O, A, O, B, O, C, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-1,xmax:4,ymin:-1,ymax:4},ABCD,codage,marquesDemiDiagonales,sAC,sBD)+"<br>"}
-              texte_corr += "C'est un carré."
-              break;
-          case 'carre3':
-              texte = "Quelle est la nature d'un quadrilatère ayant ses 4 côtés de même longueur et un angle droit ?";
-              A = point(0, 0);
-              B = point(3, 0);
-              C = point(3, 3);
-              D = point(0, 3);
-              O = point(1.5, 1.5);
-              ABCD = polygone(A, B, C, D);
-              codage = codageAngleDroit(A,B,C);
-              // codage2 = codageAngleDroit(B,C,D);
-              // codage3 = codageAngleDroit(C,D,A);
-              // codage4 = codageAngleDroit(D,A,B);
-              marquesCotes = codeSegments("||", "blue", A, B, B, C, C, D, D, A);
-              // sAC = segment(A, C);
-              // sBD = segment(B, D);
-              // sOA = segment(O, A);
-              // sOB = segment(O, B);
-              // sOC = segment(O, C);
-              // sOD = segment(O, D);
-              // sAC.pointilles = true;
-              // sBD.pointilles = true;
-              // marquesDemiDiagonales = codeSegments("||", "blue", O, A, O, B, O, C, O, D);
-              if (this.correction_detaillee) {texte_corr = mathalea2d({xmin:-1,xmax:4,ymin:-1,ymax:4},ABCD,codage,marquesCotes)+"<br>"}
-              texte_corr += "C'est un carré."
-              break;
-      }
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  //this.besoin_formulaire_numerique = ['Niveau de difficulté',3,'1 : ....\n2 : .....,\n3 : .....];
-}
-
 
 // Exercices paramétrés pour correspondre au référentiel
 // Référence 5P10
@@ -9247,67 +6779,6 @@ function Reconnaitre_quadrilatere_particulier() {
 //  Proportionnalite_pas_proportionnalite.call(this)
 // Pas de paramètres Sup
 //}
-
-// Référence 6C23
-function Exercice_additionner_des_fractions_6e() {
-  Exercice_additionner_ou_soustraire_des_fractions_5e.call(this);
-  this.sup2 = 1;
-  this.level = 6;
-  this.titre = "Additionner des fractions de même dénominateur";
-  this.besoin_formulaire_numerique2 = false;
-}
-/**
- * @Auteur Jean-Claude Lhote
- * référence 6G11
- */
-function Tracer_des_perpendiculaires() {
-  Parallele_et_Perpendiculaires.call(this);
-  this.titre = "Tracer des perpendiculaires";
-  this.sup = 1;
-  this.besoin_formulaire_numerique = false;
-}
-/**
- * @Auteur Jean-Claude Lhote
- * référence 6G12
- */
-function Tracer_des_paralleles() {
-  Parallele_et_Perpendiculaires.call(this);
-  this.titre = "Tracer des parallèles";
-  this.sup = 2;
-  this.besoin_formulaire_numerique = false;
-}
-/**
- * @Auteur Jean-Claude Lhote
- * référence 6G12-1
- */
-function Tracer_des_perpendiculaires_et_des_paralleles() {
-  Parallele_et_Perpendiculaires.call(this);
-  this.titre = "Tracer des perpendiculaires et des parallèles";
-  this.sup = 3;
-  this.besoin_formulaire_numerique = false;
-}
-/**
- * Exercice en html seulement. Symétrie centrale dans un pavage.
- * @Auteur Jean-Claude Lhote
- * référence 5G12
- */
-function Pavages_et_demi_tour() {
-  Pavages_et_transformations.call(this);
-  this.titre =
-    "Trouver l'image d'une figure par une symétrie centrale dans un pavage (sortie Latex Impossible)";
-  this.sup = 2;
-}
-/**
- * Exercice en html seulement. Rotationss dans un pavage.
- * @Auteur Jean-Claude Lhote
- * référence 3G12-1
- */
-function Pavages_et_rotation() {
-  Pavages_et_transformations.call(this);
-  this.titre =
-    "Trouver l'image d'une figure par une rotation de 90 degrés dans un pavage (sortie Latex Impossible)";
-  this.sup = 4;
-}
 
 /**
  * Lire les coordonnées d'un point du quart de plan positif avec une précision allant de l'unité à 0,25.
@@ -9428,213 +6899,7 @@ function Reglages_6N34() {
 }
 
 
-/**
- * @Auteur Jean-Claude Lhote
- * Placer les sommets et les égalités de longueur...
- * Référence 6G20
- */
-function Nommer_et_coder_des_polygones() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Nommer et coder des polygones";
-  this.consigne =
-    "Nommer les figures en fonction de l'énoncé puis ajouter le codage.";
-  this.nb_questions = 4;
-  this.nb_cols = 2;
-  this.nb_cols_corr = 2;
-  this.sup = 3;
 
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let Xmin, Xmax, Ymin, Ymax, ppc, sc, g, k, carreaux
-    ppc = 20
-    if (sortie_html) {
-      sc = 0.5
-    } else {
-      sc = 0.4
-    }
-
-    let params
-
-    let liste = combinaison_listes([1, 2, 3, 4, 5, 6, 7, 8], this.nb_questions);
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      pixelsParCm = 40;
-      let pol, polcode, polsom, polnom
-      function choisir_polygone(n) { // n compris entre 1 et 8 (1 à 4 pour un triangle, 5 à 8 pour une quadrilatère)
-        let A, B, C, D
-        let nom = creerNomDePolygone(4, "PQ"), pnom, q, p, pcode, enonce
-        switch (n) {
-          case 1: // triangle isocèle
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(0, 10) / 10, nom[1])
-            C = rotation(B, A, randint(25, 80), nom[2])
-            q = polygone(A, B, C)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2])
-            pcode = [codeSegments('||', 'blue', A, B, A, C), codeAngle(B, C, A, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2), codeAngle(C, B, A, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2)]
-            enonce = `le triangle $${nom[0] + nom[1] + nom[2]}$ est isocèle en $${nom[0]}$.<br>`
-            break
-          case 2: // triangle équilatéral
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(0, 10) / 10, nom[1])
-            C = rotation(B, A, 60, nom[2])
-            q = polygone(A, B, C)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2])
-            pcode = [codeSegments('||', 'blue', A, B, A, C, B, C), codeAngle(B, C, A, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2), codeAngle(C, B, A, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2), codeAngle(C, A, B, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2)]
-            enonce = `le triangle $${nom[0] + nom[1] + nom[2]}$ est équilatéral.<br>$\\phantom{et sa longueur est AB}$`
-            break
-          case 3: // triangle rectangle
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(0, 10) / 10, nom[1])
-            C = similitude(B, A, 90, randint(30, 100) / 100, nom[2])
-            q = polygone(A, B, C)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2])
-            pcode = codageAngleDroit(B, A, C)
-            enonce = `le triangle $${nom[0] + nom[1] + nom[2]}$ est rectangle en $${nom[0]}$.<br>$\\phantom{et sa longueur est AB}$`
-            break
-          case 4: // triangle rectangle isocèle
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(0, 10) / 10, nom[1])
-            C = rotation(B, A, 90, nom[2])
-            q = polygone(A, B, C)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2])
-            pcode = [codeSegments('||', 'blue', A, B, A, C), codageAngleDroit(B, A, C), codeAngle(B, C, A, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2), codeAngle(C, B, A, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2)]
-            enonce = `le triangle $${nom[0] + nom[1] + nom[2]}$ est rectangle et isocèle en $${nom[0]}$.`
-            break
-          //on choisit un quadrilatère 
-          case 5: // carré
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(10, 30) / 10, nom[1])
-            q = carre(A, B)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            D = p.listePoints[3]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2] + nom[3])
-            pcode = [codeSegments('||', 'blue', A, B, B, C, C, D, D, A), codageAngleDroit(B, A, D), codageAngleDroit(A, B, C), codageAngleDroit(B, C, D), codageAngleDroit(A, D, C)]
-            enonce = `le quadrilatère $${nom[0] + nom[1] + nom[2] + nom[3]}$ est un carré.<br>$\\phantom{et sa longueur est AB}$`
-            break
-          case 6: // rectangle
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(10, 30) / 10, nom[1])
-            C = similitude(A, B, -90, randint(30, 80) / 100, nom[2])
-            D = translation(C, vecteur(B, A), nom[3])
-            q = polygone(A, B, C, D)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            D = p.listePoints[3]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2] + nom[3])
-            pcode = [codeSegments('||', 'blue', A, B, C, D), codeSegments('|', 'red', C, B, A, D), codageAngleDroit(B, A, C), codageAngleDroit(A, B, C), codageAngleDroit(B, C, D), codageAngleDroit(A, D, C)]
-            enonce = `le quadrilatère $${nom[0] + nom[1] + nom[2] + nom[3]}$ est un rectangle et $${nom[0] + nom[1]}$ est sa longueur.`
-            break
-          case 7: // losange
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(10, 30) / 10, nom[1])
-            C = rotation(A, B, randint(100, 150), nom[2])
-            D = translation(C, vecteur(B, A), nom[3])
-            q = polygone(A, B, C, D)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            D = p.listePoints[3]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2] + nom[3])
-            pcode = [codeSegments('O', 'blue', A, B, B, C, C, D, D, A), codeAngle(C, D, A, 0.8, '||', 'red', 2, 0.8, 'red', 0.2), codeAngle(C, B, A, 0.8, '||', 'red', 2, 0.8, 'red', 0.2), codeAngle(B, C, D, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2), codeAngle(D, A, B, 0.8, '|', 'blue', 2, 0.8, 'blue', 0.2)]
-            enonce = `le quadrilatère $${nom[0] + nom[1] + nom[2] + nom[3]}$ est un losange et [$${nom[0] + nom[2]}$] est sa plus grande diagonale.`
-            break
-          case 8: // trapèze rectangle
-            A = point(3, randint(0, 20) / 10, nom[0])
-            B = point(randint(7, 8), randint(10, 30) / 10, nom[1])
-            D = similitude(B, A, 90, randint(30, 80) / 100, nom[3])
-            C = translation(D, homothetie(vecteur(A, B), A, randint(30, 80) / 100), nom[2])
-            q = polygone(A, B, C, D)
-            p = rotation(q, barycentre(q), randint(0, 360))
-            A = p.listePoints[0]
-            B = p.listePoints[1]
-            C = p.listePoints[2]
-            D = p.listePoints[3]
-            pnom = nommePolygone(p, nom[0] + nom[1] + nom[2] + nom[3])
-            pcode = [codageAngleDroit(B, A, D), codageAngleDroit(C, D, A)]
-            enonce = `le quadrilatère $${nom[0] + nom[1] + nom[2] + nom[3]}$ est un trapèze rectangle de grande base $${nom[0] + nom[1]}$ de hauteur $${nom[0] + nom[3]}$.`
-            break
-        }
-        return [p, nom, pcode, pnom, enonce]
-      }
-      [pol, polnom, polcode, polsom, texte] = choisir_polygone(liste[i]);
-      if (pol.listePoints.length == 4) {
-        Xmin = Math.floor(Math.min(pol.listePoints[0].x, pol.listePoints[1].x, pol.listePoints[2].x, pol.listePoints[3].x) - 1)
-        Ymin = Math.floor(Math.min(pol.listePoints[0].y, pol.listePoints[1].y, pol.listePoints[2].y, pol.listePoints[3].y) - 1)
-        Xmax = Math.ceil(Math.max(pol.listePoints[0].x, pol.listePoints[1].x, pol.listePoints[2].x, pol.listePoints[3].x) + 1)
-        Ymax = Math.ceil(Math.max(pol.listePoints[0].y, pol.listePoints[1].y, pol.listePoints[2].y, pol.listePoints[3].y) + 1)
-      }
-      else {
-        Xmin = Math.floor(Math.min(pol.listePoints[0].x, pol.listePoints[1].x, pol.listePoints[2].x) - 1)
-        Ymin = Math.floor(Math.min(pol.listePoints[0].y, pol.listePoints[1].y, pol.listePoints[2].y) - 1)
-        Xmax = Math.ceil(Math.max(pol.listePoints[0].x, pol.listePoints[1].x, pol.listePoints[2].x) + 1)
-        Ymax = Math.ceil(Math.max(pol.listePoints[0].y, pol.listePoints[1].y, pol.listePoints[2].y) + 1)
-
-      }
-      params = {
-        xmin: Xmin,
-        ymin: Ymin,
-        xmax: Xmax,
-        ymax: Ymax,
-        pixelsParCm: ppc,
-        scale: sc,
-      }
-      if (this.sup < 3) g = grille(Xmin, Ymin, Xmax, Ymax, "gray", 0.7);
-      else g = ''
-      if (this.sup == 2) {
-        k = 0.8;
-        carreaux = seyes(Xmin, Ymin, Xmax, Ymax);
-      } else {
-        k = 0.5;
-        carreaux = "";
-      }
-
-      pol.epaisseur = 2
-      texte += `<br>` + mathalea2d(params, pol, polnom, g, carreaux)
-      texte_corr = mathalea2d(params, pol, polnom, polcode, polsom, g, carreaux)
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-    pixelsParCm = 20
-  };
-  this.besoin_formulaire_numerique = [
-    "Type de cahier",
-    3,
-    `1 : Cahier à petits careaux\n 2 : Cahier à gros carreaux (Seyes)\n 3 : Feuille blanche`,
-  ];
-}
 
 /**
  * Représenter une somme de fracions de même dénominateur sur un segment gradué de façon adaptée.
@@ -9720,360 +6985,8 @@ function Ajouter_des_fractions_d_unite() {
   ];
 }
 
-/**
- * Utiliser les notations des segments, droites et demi-droites
- * @Auteur Rémi Angot
- * Référence 6G10
- */
-function Notation_segment_droite_demi_droite() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Notation des droites, segments et demi-droites";
-  this.consigne =
-    "Compléter les programmes de constructions qui ont permis d'obtenir ces figures.";
-  this.nb_questions = 3;
-  this.nb_cols = 3;
-  this.nb_cols_corr = 2;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      pixelsParCm = 40;
-      let p = creerNomDePolygone(3, "PQ");
-      let A = point(0, 0, p[0], "above left");
-      let B = point(1, 1.2, p[1], "above");
-      let C = point(2.2, -0.3, p[2], "above right");
-      let dAB, dAC, dBC, dABCorr, dACCorr, dBCCorr;
-      function creerDroiteDemiSegment(A, B) {
-        let trait, notation;
-        switch (randint(1, 4)) {
-          case 1:
-            trait = droite(A, B);
-            notation = `$(${A.nom}${B.nom})$`;
-            break;
-          case 2:
-            trait = demiDroite(A, B);
-            notation = `$[${A.nom}${B.nom})$`;
-            break;
-          case 3:
-            trait = demiDroite(B, A);
-            notation = `$[${B.nom}${A.nom})$`;
-            break;
-          case 4:
-            trait = segment(A, B);
-            notation = `$[${A.nom}${B.nom}]$`;
-            break;
-        }
-        return [trait, notation];
-      }
-      [dAB, dABCorr] = creerDroiteDemiSegment(A, B);
-      [dAC, dACCorr] = creerDroiteDemiSegment(A, C);
-      [dBC, dBCCorr] = creerDroiteDemiSegment(B, C);
-      let labels = labelPoint(A, B, C);
-
-      texte = `Placer 3 points $${p[0]}$, $${p[1]}$ et $${p[2]}$ non alignés puis tracer... <br><br>`;
-      texte += mathalea2d(
-        { xmin: -1, ymin: -1, xmax: 3, ymax: 2.5, pixelsParCm: 40, scale: 1 },
-        dAB,
-        dBC,
-        dAC,
-        labels
-      );
-      texte_corr = `...tracer ${dABCorr}, ${dBCCorr}, ${dACCorr}.`;
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  //this.besoin_formulaire_numerique = ['Niveau de difficulté',3];
-}
 
 
-
-/**
- * Vocabulaire des triangles
- * 6G20-2
- * @author Sébastien Lozano
- */
-function Vocabulaire_des_triangles_6e() {
-  this.classe = 6;
-  Vocabulaire_des_triangles.call(this);
-}
-
-/**
- * Vocabulaire des triangles
- * 5G20-1
- * @author Sébastien Lozano
- */
-function Vocabulaire_des_triangles_5e() {
-  this.classe = 5;
-  Vocabulaire_des_triangles.call(this);
-}
-// Référence 6G21
-function Construire_un_triangle_6e() {
-  Construire_un_triangle.call(this)
-  this.classe = 6
-}
-function Construire_un_triangle_avec_cible_6e() {
-  Construire_un_triangle_avec_cible.call(this)
-  this.classe = 6
-}
-function Construire_un_triangle_avec_cible_5e() {
-  Construire_un_triangle_avec_cible.call(this)
-  this.classe = 5
-}
-/**
- * Publié le 30/08/202
- * @Auteur Jean-Claude Lhote
- * Réfrence 6G21 et ... (exercice en 5e ? pas encore fait)
- */
-function Construire_un_triangle() {
-  "use strict"
-  Exercice.call(this)
-  this.titre = "Construire un triangle aux instruments";
-  this.nb_questions = 2;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = false;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []
-    this.liste_corrections = []
-    let type_de_questions_disponibles, A, B, C, CC, lAB, lBC, lAC, cA, cB, T, TT, dBC, dAB, objets_enonce, objets_correction, params_enonce, params_correction, nom, sommets
-    if (this.classe == 6) type_de_questions_disponibles = [1, 2]
-    else type_de_questions_disponibles = [1, 2, 3, 4, 5, 6]
-    let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions)
-    for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-      objets_enonce = []
-      objets_correction = []
-      texte = `Le triangle ci-dessous a été réalisé à main levée.<br>Construire ce triangle avec les instruments de géométrie en respectant les mesures indiquées.<br>`
-      texte_corr = `Voici la construction que tu devais réaliser.<br>`
-      nom = creerNomDePolygone(3, "PQ")
-      sommets = []
-      for (let i = 0; i < 3; i++) sommets.push(nom[i])
-      sommets = shuffle(sommets)
-      A = point(0, 0, sommets[0],'left')
-      switch (liste_type_de_questions[i]) {
-        case 1:
-          lAC = randint(35, 45)
-          lBC = calcul(randint(35, 45, lAC) / 10)
-          lAB = calcul(randint(46, 60) / 10)
-          lAC = calcul(lAC / 10)
-          B = pointAdistance(A, lAB, randint(-45, 45), sommets[1])
-          B.positionLabel='right'
-          cA = cercle(A, lAC)
-          cB = cercle(B, lBC)
-          C = pointIntersectionCC(cA, cB, sommets[2], 1)
-          C.positionLabel='above'
-          CC = point(C.x + randint(-5, 5, 0) / 10, C.y + randint(-5, 5, 0) / 10, sommets[2])
-
-          objets_enonce.push(afficheLongueurSegment(B, A), afficheLongueurSegment(C, B), afficheLongueurSegment(A, C))
-          objets_correction.push(traceCompas(A, C, 30, 'gray', 1, 2), traceCompas(B, C, 30, 'gray', 1, 2), afficheLongueurSegment(B, A), afficheLongueurSegment(C, B), afficheLongueurSegment(A, C))
-          texte_corr += `Pour cette construction, nous avons utilisé le compas et la règle graduée.<br>`
-          break;
-
-        case 2:
-          lAC = randint(70, 80) / 10
-          lAB = calcul(randint(46, 60) / 10)
-          B = pointAdistance(A, lAB, randint(-45, 45), sommets[1])
-          cA = cercle(A, lAC)
-          dAB = droite(A, B)
-          dBC = droiteParPointEtPerpendiculaire(B, dAB)
-          C = pointIntersectionLC(dBC, cA, sommets[2], 1)
-          CC = point(C.x + randint(-5, 5, 0) / 10, C.y + randint(-5, 5, 0) / 10, sommets[2])
-
-          objets_enonce.push(afficheLongueurSegment(B, A), afficheLongueurSegment(C, A), codageAngleDroit(A, B, C))
-          objets_correction.push(traceCompas(A, C, 30, 'gray', 1, 2), codageAngleDroit(A, B, C), afficheLongueurSegment(B, A), afficheLongueurSegment(C, A))
-          texte_corr += `Pour cette construction, nous avons utilisé la règle graduée, l'équerre et le compas.<br>`
-          break
-
-        case 3:
-
-          break
-      }
-      T = polygoneAvecNom(A, B, C)
-      TT = polygoneAvecNom(A, B, CC)
-      objets_enonce.push(TT[0], TT[1])
-      objets_correction.push(T[0], T[1])
-      params_enonce = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 1), ymin: Math.min(A.y - 1, B.y - 1, C.y - 1), xmax: Math.max(A.x + 1, B.x + 1, C.x + 1), ymax: Math.max(A.y + 1, B.y + 1, C.y + 1), pixelsParCm: 30, scale: 1, mainlevee: true, amplitude: 1 }
-      params_correction = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 2), ymin: Math.min(A.y - 1, B.y - 1, C.y - 2), xmax: Math.max(A.x + 1, B.x + 1, C.x + 2), ymax: Math.max(A.y + 1, B.y + 1, C.y + 2), pixelsParCm: 30, scale: 1 }
-      texte += mathalea2d(params_enonce, objets_enonce)
-      texte_corr += mathalea2d(params_correction, objets_correction)
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  //	this.besoin_formulaire_numerique = ['Type de questions', 3, `1 : Perpendiculaires\n 2 : Parallèles\n 3 : Mélange`]
-  this.besoin_formulaire2_numerique = [
-    "Type de cahier",
-    3,
-    `1 : Cahier à petits careaux\n 2 : Cahier à gros carreaux (Seyes)\n 3 : Feuille blanche`,
-  ];
-}
-/**
- * publié le 1/12/2020
- * @Auteur Jean-Claude Lhote
- * Réfrence 6G21-1 et ... (exercice en 5e ? pas encore fait)
- */
-function Construire_un_triangle_avec_cible() {
-  "use strict"
-  Exercice.call(this)
-  this.titre = "Construire un triangle avec cible auto-corrective";
-  this.nb_questions = 4;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = false;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []
-    this.liste_corrections = []
-    let celluleAleaRonde = function (rang) {
-      let lettre = lettre_depuis_chiffre(randint(1, 8))
-      let chiffre = Number(randint(1, rang)).toString()
-      return lettre + chiffre
-    }
-
-    let type_de_questions_disponibles, cible, cellule, result, A, B, C, CC, lAB, lBC, lAC, cA, cB, T, TT, dBC,dAC, dAB, objets_enonceml, objets_enonce, objets_correction, params_enonceml, params_enonce, params_correction, nom, sommets
-    if (this.classe == 6) type_de_questions_disponibles = [1, 2]
-    else type_de_questions_disponibles = [1, 2, 3,4]
-    let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions)
-    for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-      objets_enonce = []
-      objets_enonceml = []
-      objets_correction = []
-      texte = `Le triangle ci-dessous a été réalisé à main levée.<br>Construire ce triangle avec les instruments de géométrie en respectant les mesures indiquées.<br>`
-      texte_corr = `Voici la construction que tu devais réaliser.<br>`
-      nom = creerNomDePolygone(3, "PQ")
-      sommets = []
-      for (let i = 0; i < 3; i++) sommets.push(nom[i])
-      sommets = shuffle(sommets)
-      A = point(0, 0, sommets[0],'left')
-      switch (liste_type_de_questions[i]) {
-        case 1:
-          lAC = randint(35, 45)
-          lBC = calcul(randint(35, 45, lAC) / 10)
-          lAB = calcul(randint(46, 60) / 10)
-          lAC = calcul(lAC / 10)
-          B = pointAdistance(A, lAB, randint(-45, 45), sommets[1])
-          B.positionLabel='right'
-          cA = cercle(A, lAC)
-          cB = cercle(B, lBC)
-          C = pointIntersectionCC(cA, cB, sommets[2], 1)
-          C.positionLabel='above'
-          CC = point(C.x + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, C.y + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, sommets[2])
-          cellule = celluleAleaRonde(5)
-          result = dansLaCibleRonde(C.x, C.y, 5, 0.3, cellule)
-          cible = cibleRonde({ x: result[0], y: result[1], rang: 5, taille: 0.3 })
-          objets_enonce.push(cible, segmentAvecExtremites(A, B), labelPoint(A, B))
-          objets_enonceml.push(afficheLongueurSegment(B, A), afficheLongueurSegment(C, B, 'black', 1), afficheLongueurSegment(A, C, 'black', 1))
-          objets_correction.push(cible, traceCompas(A, C, 30, 'gray', 1, 2), traceCompas(B, C, 30, 'gray', 1, 2), afficheLongueurSegment(B, A), afficheLongueurSegment(C, B), afficheLongueurSegment(A, C))
-          texte_corr += `Pour cette construction, nous avons utilisé le compas et la règle graduée.<br>`
-          texte_corr += `Le point ${sommets[2]} se trouve dans le secteur ${cellule}.<br>`
-          break;
-
-        case 2:
-          lAC = randint(70, 80) / 10
-          lAB = calcul(randint(46, 60) / 10)
-          B = pointAdistance(A, lAB, randint(-45, 45), sommets[1])
-          B.positionLabel='right'
-          cA = cercle(A, lAC)
-          dAB = droite(A, B)
-          dBC = droiteParPointEtPerpendiculaire(B, dAB)
-          C = pointIntersectionLC(dBC, cA, sommets[2], 1)
-          C.positionLabel='above'
-          CC = point(C.x + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, C.y + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, sommets[2])
-          cellule = celluleAleaRonde(5)
-          result = dansLaCibleRonde(C.x, C.y, 5, 0.3, cellule)
-          cible = cibleRonde({ x: result[0], y: result[1], rang: 5, taille: 0.3 })
-          objets_enonce.push(cible, segmentAvecExtremites(A, B), labelPoint(A, B))
-          objets_enonceml.push(afficheLongueurSegment(B, A), afficheLongueurSegment(A, C, 'black', 1), codageAngleDroit(A, B, CC))
-          objets_correction.push(cible, traceCompas(A, C, 30, 'gray', 1, 2), codageAngleDroit(A, B, C), afficheLongueurSegment(B, A), afficheLongueurSegment(C, A))
-          texte_corr += `Pour cette construction, nous avons utilisé la règle graduée, l'équerre et le compas.<br>`
-          texte_corr += `Le point ${sommets[2]} se trouve dans le secteur ${cellule}.<br>`
-  
-          break
-
-        case 3:
-          lAB = calcul(randint(46, 60) / 10)
-          B = pointAdistance(A, lAB, randint(-45, 45), sommets[1])
-          B.positionLabel='right'
-          dAB = droite(A, B)
-          dAC=rotation(dAB,A,randint(8,14)*5)
-          dBC=rotation(dAB,B,-randint(8,12)*5)
-          C=pointIntersectionDD(dAC,dBC,sommets[2])
-          C.positionLabel='above'
-          CC = point(C.x + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, C.y + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, sommets[2])
-          cellule = celluleAleaRonde(5)
-          result = dansLaCibleRonde(C.x, C.y, 5, 0.3, cellule)
-          cible = cibleRonde({ x: result[0], y: result[1], rang: 5, taille: 0.3 })
-          objets_enonce.push(cible, segmentAvecExtremites(A, B), labelPoint(A, B))
-          objets_enonceml.push(codeAngle(B, A,CC,1.1),texteParPoint(nombre_avec_espace(Math.round(angle(B,A,C)))+`°`,similitude(B,A,angle(B,A,C)/2,1/lAB+0.1)), codeAngle(A, B,CC,1.1),texteParPoint(nombre_avec_espace(Math.round(angle(A,B,C)))+`°`,similitude(A,B,-angle(A,B,C)/2,1/lAB+0.1)))
-          objets_correction.push(cible,afficheLongueurSegment(B, A),afficheMesureAngle(B, A,C,'black',1), afficheMesureAngle(A, B,C,'black',1))
-          texte_corr += `Pour cette construction, nous avons utilisé le rapporteur.<br>`
-          texte_corr += `Le point ${sommets[2]} se trouve dans le secteur ${cellule}.<br>`
-  
-           
-    
-
-          break
-          case 4:
-            lAB = calcul(randint(46, 60) / 10)
-            lAC = randint(40, 60) / 10
-            B = pointAdistance(A, lAB, randint(-45, 45), sommets[1],'right')
-            C=similitude(B,A,randint(8,24)*5,lAC/lAB,sommets[2],'above')
-            CC = point(C.x + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, C.y + randint(-5, 5, [-2, -1, 0, 1, 2]) / 10, sommets[2])
-            cellule = celluleAleaRonde(5)
-            result = dansLaCibleRonde(C.x, C.y, 5, 0.3, cellule)
-            cible = cibleRonde({ x: result[0], y: result[1], rang: 5, taille: 0.3 })
-            objets_enonce.push(cible, segmentAvecExtremites(A, B), labelPoint(A, B))
-            objets_enonceml.push(codeAngle(B, A,CC,1.1),afficheLongueurSegment(B, A),texteParPoint(nombre_avec_espace(Math.round(angle(B,A,C)))+`°`,similitude(B,A,angle(B,A,C)/2,1/lAB+0.1)),afficheLongueurSegment(A, C, 'black', 1) )
-            objets_correction.push(cible,afficheLongueurSegment(B, A),afficheMesureAngle(B, A,C,'black',1), afficheLongueurSegment(A, C, 'black', 1))
-            texte_corr += `Pour cette construction, nous avons utilisé le rapporteur et la règle graduée.<br>`
-            texte_corr += `Le point ${sommets[2]} se trouve dans le secteur ${cellule}.<br>`
-             break  
-          
-      }
-      T = polygoneAvecNom(A, B, C)
-      TT = polygoneAvecNom(A, B, CC)
-      objets_enonceml.push(TT[0], TT[1])
-      objets_correction.push(T[0], T[1])
-      params_enonceml = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 3), ymin: Math.min(A.y - 1, B.y - 1, C.y - 3), xmax: Math.max(A.x + 1, B.x + 1, C.x + 3), ymax: Math.max(A.y + 1, B.y + 1, C.y + 3), pixelsParCm: 30, scale: 1, mainlevee: true, amplitude: 1 }
-      params_enonce = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 3), ymin: Math.min(A.y - 1, B.y - 1, C.y - 3), xmax: Math.max(A.x + 1, B.x + 1, C.x + 3), ymax: Math.max(A.y + 1, B.y + 1, C.y + 3), pixelsParCm: 30, scale: 1, mainlevee: false, amplitude: 1 }
-      params_correction = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 3), ymin: Math.min(A.y - 1, B.y - 1, C.y - 3), xmax: Math.max(A.x + 1, B.x + 1, C.x + 3), ymax: Math.max(A.y + 1, B.y + 1, C.y + 3), pixelsParCm: 30, scale: 1 }
-      texte += mathalea2d(params_enonceml, objets_enonceml) + mathalea2d(params_enonce, objets_enonce)
-      texte_corr += mathalea2d(params_correction, objets_correction)
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  //	this.besoin_formulaire_numerique = ['Type de questions', 3, `1 : Perpendiculaires\n 2 : Parallèles\n 3 : Mélange`]
-  this.besoin_formulaire2_numerique = [
-    "Type de cahier",
-    3,
-    `1 : Cahier à petits careaux\n 2 : Cahier à gros carreaux (Seyes)\n 3 : Feuille blanche`,
-  ];
-}
 /**
  * Non Publié : base servant à faire des tutoriels vidéos
  * @Auteur Jean-Claude Lhote
@@ -10546,784 +7459,6 @@ function Test_main_levee() {
   this.besoin_formulaire_case_a_cocher = ["main levee"];
 }
 
-/**
- * Exercice du test de positionnement 5e
- * Variantes à venir...
- * @Auteur Jean-Claude Lhote
- * Référence 6G10-2
- */
-function Utiliser_le_codage_pour_decrire_6e() {
-  Utiliser_le_codage_pour_decrire.call(this)
-  this.classe = 6
-}
-// 5G30
-function Utiliser_le_codage_pour_decrire_5e() {
-  Utiliser_le_codage_pour_decrire.call(this)
-  this.classe = 5
-}
-function Utiliser_le_codage_pour_decrire() {
-  "use strict"
-  Exercice.call(this);
-  this.titre = "Utiliser le codage pour décrire une figure";
-  this.nb_questions = 1;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 1;
-  this.sup2 = 1;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let type_de_questions_disponibles;
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let Xmin, Xmax, Ymin, Ymax, ppc = 20, sc, nom, sommets = [], params_enonce, params_correction, objets_enonce, objets_correction
-    let A, B, C, D, E, F, s1, s2, s3, s4, s5, s6, s7, s8, medAC, medBC, dBD, dBC, dAC, dAF
-    if (this.classe == 6) type_de_questions_disponibles = [1, 2, 3]
-    else type_de_questions_disponibles = [1, 2, 3, 4]
-    let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions)
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      objets_enonce = []
-      objets_correction = []
-      params_enonce = {}
-      params_correction = {}
-      nom = creerNomDePolygone(6, "PQ")
-      for (let i = 0; i < 6; i++)
-        sommets.push(nom[i])
-      sommets = shuffle(sommets)
-
-      A = point(0, 0, sommets[0], 'left')
-      switch (liste_type_de_questions[i]) {
-        case 1:
-          C = pointAdistance(A, randint(5, 7), randint(-45, 45), sommets[2], 'right')
-          s2 = segment(A, C)
-          B = similitude(C, A, -85, randint(5, 7) / 10, sommets[1], 'below')
-          s1 = segment(A, B)
-          E = pointSurSegment(A, C, longueur(A, C) / 2.2, sommets[4], 'below')
-          medBC = mediatrice(C, B)
-          medAC = mediatrice(A, C)
-          dBC = droite(C, B)
-          dBD = rotation(dBC, B, randint(-40, -20))
-          dAC = droite(A, C)
-          dAF = rotation(dAC, A, randint(30, 40))
-          D = pointIntersectionDD(dBD, medBC, sommets[3], 'below')
-          D.x += randint(-2, 2, 0) / 5
-          F = pointIntersectionDD(dAF, medAC, sommets[5], 'above')
-          F.x += randint(-2, 2, 0) / 5
-          s5 = segment(B, D)
-          s6 = segment(C, D)
-          s3 = segment(A, F)
-          s7 = segment(C, F)
-          s8 = segment(E, F)
-          s4 = segment(B, C)
-          params_enonce = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 1, D.x - 1, E.x - 1, F.x - 1), ymin: Math.min(A.y - 1, B.y - 1, C.y - 1, D.y - 1, E.y - 1, F.y - 1), xmax: Math.max(A.x + 1, B.x + 1, C.x + 1, D.x + 1, E.x + 1, F.x + 1), ymax: Math.max(A.y + 1, B.y + 1, C.y + 1, D.y + 1, E.y + 1, F.y + 1.5), pixelsParCm: 30, scale: 1, mainlevee: true, amplitude: 1 }
-          objets_enonce.push(s1, s2, s4, s8, s7, s3, s6, s5, codageAngleDroit(B, A, C), codeSegments('//', 'black', A, F, F, C), codeSegments('|||', 'black', A, E, E, C), codeSegments('O', 'black', B, D, D, C), labelPoint(A, B, C, D, E, F), codageAngleDroit(A, E, F))
-          texte = `<br>À l'aide du schéma ci-dessous, déterminer :<br>`
-          texte += `- deux segments de même longueur ;<br>`
-          texte += `- le milieu d'un segment ;<br>`
-          texte += `- un triangle rectangle ;<br>`
-          texte += `- un triangle isocèle ;<br>`
-          texte_corr = `- Deux segments de même mesure : [$${sommets[0] + sommets[4]}$] et $[${sommets[4] + sommets[2]}]$ ou $[${sommets[0] + sommets[5]}]$ et $[${sommets[5] + sommets[2]}]$`
-          texte_corr += ` ou $[${sommets[1] + sommets[3]}]$ et $[${sommets[3] + sommets[2]}]$.<br>`
-          texte_corr += `- $${sommets[4]}$ est le milieu du segment $[${sommets[0] + sommets[2]}]$.<br>`
-          texte_corr += `- $${sommets[0] + sommets[1] + sommets[2]}$ est un triangle rectangle en $${sommets[0]}$, $${sommets[0] + sommets[4] + sommets[5]}$ est un triangle rectangle en $${sommets[4]}$ et $${sommets[2] + sommets[4] + sommets[5]}$ est un triangle rectangle en $${sommets[4]}$.<br>`
-          texte_corr += `- $${sommets[0] + sommets[5] + sommets[2]}$ est un triangle isocèle en $${sommets[5]}$ et $${sommets[1] + sommets[3] + sommets[2]}$ est un triangle isocèle en $${sommets[3]}$.<br>`
-          break
-        case 2:
-          B = pointAdistance(A, randint(5, 7), randint(-45, 45), sommets[1], 'above')
-          C = similitude(A, B, randint(85, 90), 0.95, sommets[2], 'below')
-          D = similitude(B, A, randint(-93, -87), 1, sommets[3], 'below')
-          F = similitude(B, C, -55, 0.8, sommets[5], 'right')
-          E = similitude(C, D, 57, randint(85, 115) / 100, sommets[4], 'right')
-          s1 = segment(D, E)
-          s2 = segment(C, E)
-          s4 = segment(C, F)
-          s5 = segment(B, F)
-          s6 = polygone(A, B, C, D)
-          params_correction = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 1, D.x - 1, E.x - 1, F.x - 1), ymin: Math.min(A.y - 1, B.y - 1, C.y - 1, D.y - 1, E.y - 1, F.y - 1), xmax: Math.max(A.x + 1, B.x + 1, C.x + 1, D.x + 1, E.x + 1, F.x + 1), ymax: Math.max(A.y + 1, B.y + 1, C.y + 1, D.y + 1, E.y + 1, F.y + 1), pixelsParCm: 30, scale: 1, mainlevee: true, amplitude: 1 }
-          objets_correction.push(labelPoint(A, B, C, D, E, F), s1, s2, s4, s5, s6)
-          objets_correction.push(codageAngleDroit(D, A, B), codageAngleDroit(A, B, C), codageAngleDroit(B, C, D), codageAngleDroit(C, D, A))
-          objets_correction.push(codeSegments('||', 'black', D, E, C, E), codeSegments('O', 'black', A, B, B, C, C, D, D, A), codeSegments('|||', 'black', F, C, B, F))
-          texte = `$${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$ est un carré et $${sommets[3] + sommets[2] + sommets[4]}$ est un triangle équilatéral ($${sommets[4]}$ est à l'intérieur du carré $${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$).<br>`
-          texte += ` $${sommets[1] + sommets[2] + sommets[5]}$ est un triangle isocèle en $${sommets[5]}$ ($${sommets[5]}$ est à l'extérieur du carré $${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$).<br>`
-          texte += `Représenter cette configuration par un schéma à main levée et ajouter les codages nécessaires.`
-          texte_corr = `Voilà ci-dessous un schéma qui pourrait convenir à la situation.<br>`
-          break
-        case 3:
-          B = pointAdistance(A, randint(5, 7), randint(-45, 45), sommets[1], 'above')
-          C = similitude(A, B, randint(85, 90), 0.5, sommets[2], 'below')
-          D = similitude(B, A, randint(-93, -87), 0.53, sommets[3], 'below')
-          s1 = segment(D, B)
-          s2 = segment(A, C)
-          E = pointIntersectionDD(droite(A, C), droite(D, B), sommets[4], 'above')
-          F = affiniteOrtho(E, droite(B, C), -1.1, sommets[5], 'right')
-          s3 = polygone(A, B, C, D)
-          s4 = segment(B, F)
-          s5 = segment(C, F)
-          params_correction = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 1, D.x - 1, E.x - 1, F.x - 1), ymin: Math.min(A.y - 1, B.y - 1, C.y - 1, D.y - 1, E.y - 1, F.y - 1), xmax: Math.max(A.x + 1, B.x + 1, C.x + 1, D.x + 1, E.x + 1, F.x + 1), ymax: Math.max(A.y + 1, B.y + 1, C.y + 1, D.y + 1, E.y + 1, F.y + 1), pixelsParCm: 30, scale: 1, mainlevee: true, amplitude: 1 }
-          objets_correction.push(labelPoint(A, B, C, D, E, F), s1, s2, s3, s4, s5)
-          objets_correction.push(codageAngleDroit(D, A, B), codageAngleDroit(A, B, C), codageAngleDroit(B, C, D), codageAngleDroit(C, D, A))
-          objets_correction.push(codeSegments('||', 'black', D, E, E, B, A, E, E, C, F, C, B, F), codeSegments('O', 'black', A, B, D, C), codeSegments('/', 'black', A, D, B, C))
-          texte = `$${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$ est un rectangle. Ses diagonales se coupent en $${sommets[4]}$.<br>`
-          texte += `$${sommets[4] + sommets[1] + sommets[5] + sommets[2]}$ est un losange.<br>`
-          texte += `Représenter cette configuration par un schéma à main levée et ajouter les codages nécssaires.`
-          texte_corr = `Voilà ci-dessous un schéma qui pourrait convenir à la situation.<br>`
-          break
-        case 4:
-          B = pointAdistance(A, randint(6, 7), randint(-30, 30), sommets[1], 'above right')
-          F = similitude(A, B, randint(-70, -50), randint(80, 90) / 100, sommets[5], 'left')
-          D = similitude(B, A, angleOriente(A, B, F) + randint(3, 5), randint(15, 20) / 10, sommets[3], 'below')
-          C = translation2Points(point(B.x + 1, B.y + 1), A, D, sommets[2], 'below right')
-          E = pointIntersectionDD(droite(A, C), droite(D, B), sommets[4], 'above right')
-          s3 = polygone(A, B, C, D)
-          s4 = segment(B, F)
-          s5 = segment(A, F)
-          s1 = segment(B, D)
-          s2 = segment(A, C)
-          params_enonce = { xmin: Math.min(A.x - 1, B.x - 1, C.x - 1, D.x - 1, E.x - 1, F.x - 1), ymin: Math.min(A.y - 1, B.y - 1, C.y - 1, D.y - 1, E.y - 1, F.y - 1), xmax: Math.max(A.x + 1, B.x + 1, C.x + 1, D.x + 1, E.x + 1, F.x + 1), ymax: Math.max(A.y + 1, B.y + 1, C.y + 1, D.y + 1, E.y + 1, F.y + 1), pixelsParCm: 30, scale: 1, mainlevee: true, amplitude: 0.8 }
-          objets_enonce.push(labelPoint(A, B, C, D, E, F), s1, s2, s3, s4, s5)
-          objets_enonce.push(codeAngle(D, A, B, 2, '|', 'red', 2), codeAngle(B, C, D, 2, '|', 'red', 2), codeAngle(A, B, F, 2, '|', 'red', 2))
-          objets_enonce.push(codeAngle(A, B, C, 2, '||', 'blue', 2), codeAngle(A, D, C, 2, '||', 'blue', 2))
-          objets_enonce.push(codeAngle(B, A, F, 2, '///', 'green', 3), codeAngle(B, F, A, 2, '///', 'green', 3))
-
-          objets_enonce.push(codeSegments('||', 'black', B, E, E, D), codeSegments('O', 'black', A, E, E, C))
-          texte = `$${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$ est un rectangle. Ses diagonales se coupent en $${sommets[4]}$.<br>`
-          texte += `$${sommets[4] + sommets[1] + sommets[5] + sommets[2]}$ est un losange.<br>`
-          texte = `<br>À l'aide du schéma ci-dessous, déterminer :<br>`
-          texte += `- la nature du triangle $${sommets[0] + sommets[1] + sommets[5]}$ ;<br>`
-          texte += `- la nature du quadrilatère $${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$ ;<br>`
-          texte += `- la nature de l'angle $\\widehat{${sommets[5] + sommets[1] + sommets[2]}}$ ;<br>`
-          texte_corr = `Le triangle $${sommets[0] + sommets[1] + sommets[5]}$ a deux angles de même mesure, c'est donc un triangle isocèle en $${sommets[1]}$.<br>`
-          texte_corr += `Le quadrilatère  $${sommets[0] + sommets[1] + sommets[2] + sommets[3]}$ a des diagonales qui se coupent en leur milieu, c'est donc un parallélogramme.<br>`
-          texte_corr += `Dans un parallélogramme, les angles consécutifs sont supplémentaires (leur somme vaut 180°).<br>`
-          texte_corr += ` D'après le codage, l'angle $\\widehat{${sommets[2] + sommets[1] + sommets[5]}}$ est la somme de deux angles supplémentaires. C'est donc un angle plat.<br>`
-
-
-          break
-      }
-      if (objets_enonce.length > 0) texte += mathalea2d(params_enonce, objets_enonce)
-      if (objets_correction.length > 0) texte_corr += mathalea2d(params_correction, objets_correction)
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-}
-/**
- * Ref 6G14
- * @Auteur Jean-Claude Lhote
- * publié le 22/11/2020
- */
-function Proprietes_paralleles_perpendiculaires() {
-  "use strict";
-  Exercice.call(this);
-  this.titre = "Utiliser les propriétés des droites perpendiculaires";
-  this.nb_questions = 3;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 4;
-  this.sup2 = 1;
-  this.correction_detaillee_disponible = true;
-  this.correction_detaillee = false;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let type_de_questions_disponibles, questions_par_niveau = [];
-    questions_par_niveau.push(range(3))
-    questions_par_niveau.push(rangeMinMax(9, 15))
-    questions_par_niveau.push(rangeMinMax(19, 31, 20))
-
-    if (this.sup < 4) type_de_questions_disponibles = questions_par_niveau[parseInt(this.sup) - 1]
-    else type_de_questions_disponibles = questions_par_niveau[0].concat(questions_par_niveau[1].concat(questions_par_niveau[2]))
-
-    let liste_type_de_questions = combinaison_listes(
-      type_de_questions_disponibles,
-      this.nb_questions
-    );
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let droites = [], code, raisonnement, numDroites = [], phrases = [], textetemp
-    let d = [], P = [], objets = [], num1, num2, couleurd = [], droiteP, PP, Inter
-    let droitecolor = function (num) {
-      let couleurs
-      sortie_html ? couleurs = ['red', 'blue', 'green', 'black', 'magenta', 'orange'] : couleurs = ['black', 'black', 'black', 'black', 'black', 'black'];
-      return couleurs[num]
-    }
-    if (sortie_html) {
-      num1 = `<tspan dy="5" style="font-size:70%">`
-      num2 = `</tspan><tspan dy="-5">)</tspan>`
-    }
-    else {
-      num1 = `_`
-      num2 = `)`
-    }
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      texte = ""
-      texte_corr = ""
-      phrases.length = 0;
-      droites.length = 0;
-      objets.length = 0;
-      d.length = 0;
-      P.length = 0;
-      couleurd.length = 0
-      numDroites = shuffle([1, 2, 3, 4, 5]);
-      raisonnement = liste_type_de_questions[i]
-
-      switch (raisonnement) {
-        case 0: // si 1//2 et 2//3 alors 1//3
-          code = [[1, 2, 1], [2, 3, 1]];
-          break;
-        case 1: // si 1//2 et 2T3 alors 1T3
-          code = [[1, 2, 1], [2, 3, -1]]
-          break;
-        case 2: // si 1T2 et 2T3 alors 1//3
-          code = [[1, 2, -1], [2, 3, -1]]
-          break;
-        case 3: // si 1T2 et 2//3 alors 1T3
-          code = [[1, 2, -1], [2, 3, 1]]
-          break;
-        case 8: // Si 1//2 et 2//3 et 3//4 alors 1//4
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, 1]];
-          break;
-        case 9: // Si 1//2 et 2//3 et 3T4 alors 1T4
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, -1]];
-          break;
-        case 10: // Si 1//2 et 2T3 et 3//4 alors 1T4
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, 1]];
-          break;
-        case 11: // Si 1//2 et 2T3 et 3T4 alors 1//4
-          code = [[1, 2, 1], [2, 3, -1], [3, 4, -1]];
-          break;
-        case 12: // Si 1T2 et 2//3 et 3//4 alors 1T4
-          code = [[1, 2, -1], [2, 3, 1], [3, 4, 1]];
-          break;
-        case 13: // Si 1T2 et 2//3 et 3T4 alors 1//4
-          code = [[1, 2, -1], [2, 3, 1], [3, 4, -1]];
-          break;
-        case 14: // Si 1T2 et 2T3 et 3//4 alors 1//4
-          code = [[1, 2, -1], [2, 3, -1], [3, 4, 1]];
-          break;
-        case 15: // Si 1T2 et 2T3 et 3T4 alors 1T4
-          code = [[1, 2, -1], [2, 3, -1], [3, 4, -1]];
-          break;
-        case 16: // Si 1//2 et 2//3 et 3//4 et 4//5 alors 1//5
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, 1], [4, 5, 1]];
-          break;
-        case 17: // Si 1//2 et 2//3 et 3T4 et 4//5 alors 1T5
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, -1], [4, 5, 1]];
-          break;
-        case 18: // Si 1//2 et 2T3 et 3//4 et 4//5 alors 1T5
-          code = [[1, 2, 1], [2, 3, -1], [3, 4, 1], [4, 5, 1]];
-          break;
-        case 19: // Si 1//2 et 2T3 et 3T4 et 4//5 alors 1//5
-          code = [[1, 2, 1], [2, 3, -1], [3, 4, -1], [4, 5, 1]];
-          break;
-        case 20: // Si 1T2 et 2//3 et 3//4 et 4//5 alors 1T5
-          code = [[1, 2, -1], [2, 3, 1], [3, 4, 1], [4, 5, 1]];
-          break;
-        case 21: // Si 1T2 et 2//3 et 3T4 et 4//5 alors 1//5
-          code = [[1, 2, -1], [2, 3, 1], [3, 4, -1], [4, 5, 1]];
-          break;
-        case 22: // Si 1T2 et 2T3 et 3//4 et 4//5 alors 1//5
-          code = [[1, 2, -1], [2, 3, -1], [3, 4, 1], [4, 5, 1]];
-          break;
-        case 23: // Si 1T2 et 2T3 et 3T4 et 4//5 alors 1T5
-          code = [[1, 2, -1], [2, 3, -1], [3, 4, -1], [4, 5, 1]];
-          break;
-        case 24: // Si 1//2 et 2//3 et 3//4 et 4T5 alors 1T5
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, 1], [4, 5, -1]];
-          break;
-        case 25: // Si 1//2 et 2//3 et 3T4 et 4T5 alors 1//5
-          code = [[1, 2, 1], [2, 3, 1], [3, 4, -1], [4, 5, -1]];
-          break;
-        case 26: // Si 1//2 et 2T3 et 3//4 et 4T5 alors 1//5
-          code = [[1, 2, 1], [2, 3, -1], [3, 4, 1], [4, 5, -1]];
-          break;
-        case 27: // Si 1//2 et 2T3 et 3T4 et 4T5 alors 1T5
-          code = [[1, 2, 1], [2, 3, -1], [3, 4, -1], [4, 5, -1]];
-          break;
-        case 28: // Si 1T2 et 2//3 et 3//4 et 4T5 alors 1//5
-          code = [[1, 2, -1], [2, 3, 1], [3, 4, 1], [4, 5, -1]];
-          break;
-        case 29: // Si 1T2 et 2//3 et 3T4 et 4T5 alors 1T5
-          code = [[1, 2, -1], [2, 3, 1], [3, 4, -1], [4, 5, -1]];
-          break;
-        case 30: // Si 1T2 et 2T3 et 3//4 et 4T5 alors 1T5
-          code = [[1, 2, -1], [2, 3, -1], [3, 4, 1], [4, 5, -1]];
-          break;
-        case 31: // Si 1T2 et 2T3 et 3T4 et 4T5 alors 1//5
-          code = [[1, 2, -1], [2, 3, -1], [3, 4, -1], [4, 5, -1]];
-          break;
-
-      }
-
-      // enoncé mélangé
-      texte += `On sait que `
-      couleurd.push(randint(0, 5))
-      for (let j = 0; j < code.length; j++) {
-        textetemp = `$(d_${numDroites[code[j][0] - 1]})`;
-        if (code[j][2] == 1) {
-          textetemp += `//`
-          couleurd.push(couleurd[j])
-        }
-        else {
-          textetemp += `\\perp`
-          couleurd.push((couleurd[j] + 1) % 6)
-        }
-        textetemp += `(d_${numDroites[code[j][1] - 1]})$`
-        phrases.push(textetemp)
-      }
-      //phrases=shuffle(phrases)
-      for (let j = 0; j < code.length - 1; j++) {
-        texte += phrases[j]
-        if (j != code.length - 2) texte += `, `
-        else texte += ` et `
-      }
-      texte += phrases[code.length - 1]
-      texte += `.<br>Que peut-on dire de $(d_${numDroites[code[0][0] - 1]})$ et $(d_${numDroites[code[code.length - 1][1] - 1]})$ ?`
-
-      //construction de la figure
-
-      P.push(point(0, 0))
-      droiteP = droiteParPointEtPente(P[0], randint(-1, 1, 0) / 10, `(d${num1}${numDroites[code[0][0] - 1]}${num2}`, droitecolor(couleurd[0]))
-      droiteP.epaisseur = 2
-      droite.pointilles = false
-      d.push(droiteP)
-      objets.push(d[0])
-      for (let x = 0; x < code.length; x++) {
-        if (code[x][2] == 1) {
-          P.push(point((x + 1) * 2, (x + 1) * 2))
-          droiteP = droiteParPointEtParallele(P[x + 1], d[x], `(d${num1}${numDroites[code[x][1] - 1]}${num2}`, droitecolor(couleurd[x + 1]))
-          droiteP.epaisseur = 2
-          droiteP.pointilles = d[x].pointilles
-          d.push(droiteP)
-        }
-        else {
-          P.push(point((x + 1) * 2, (x + 1) * 2))
-          droiteP = droiteParPointEtPerpendiculaire(P[x + 1], d[x], `(d${num1}${numDroites[code[x][1] - 1]}${num2}`, droitecolor(couleurd[x + 1]))
-          droiteP.epaisseur = 2
-          droiteP.pointilles = x % 3 + 1
-          Inter = pointIntersectionDD(d[x], droiteP)
-          PP = rotation(P[x + 1], Inter, 90)
-          d.push(droiteP)
-          objets.push(codageAngleDroit(PP, Inter, P[x + 1], 'black', 0.6))
-        }
-        objets.push(d[x + 1])
-      }
-      for (let i = 0; i < code.length; i++) { // on ajoute les angles droits
-
-      }
-      // correction raisonnement ordonné
-      fenetreMathalea2d = [-2, -2, 15, 10]
-      texte_corr = `À partir de l\'énoncé, on peut réaliser le schéma suivant (il en existe une infinité)<br> Les droites données parallèles dans l'énoncé sont de même couleur/style.<br>`
-      texte_corr += mathalea2d({ xmin: -2, xmax: 15, ymin: -2, ymax: 10, pixelsParCm: 20, scale: 0.3, mainlevee: false, amplitude: 0.3 }, objets) + `<br>`
-      for (let j = 0; j < code.length - 1; j++) {
-        if (this.correction_detaillee) texte_corr += `On sait que : `
-        else texte_corr += `Comme `
-        texte_corr += `$(d_${numDroites[code[j][0] - 1]})`;
-        if (code[j][2] == 1) texte_corr += `//`
-        else texte_corr += `\\perp`
-        texte_corr += `(d_${numDroites[code[j][1] - 1]})$ et `
-        texte_corr += `$(d_${numDroites[code[j + 1][0] - 1]})`;
-        if (code[j + 1][2] == 1) texte_corr += `//`
-        else texte_corr += `\\perp`
-        texte_corr += `(d_${numDroites[code[j + 1][1] - 1]})$`
-        // quelle propriété ?
-        if (code[j][2] * code[j + 1][2] == -1) { // Une parallèle et une perpendiculaire
-          if (this.correction_detaillee) texte_corr += `.<br> Or «Si deux droites sont parallèles alors toute droite perpendiculaire à l'une est aussi perpendiculaire à l'autre».<br>Donc`
-          else texte_corr += `, on en déduit que `
-          texte_corr += ` $(d_${numDroites[code[0][0] - 1]})\\perp(d_${numDroites[code[j + 1][1] - 1]})$.<br>`
-          code[j + 1][0] = code[j][0]
-          code[j + 1][2] = -1
-        }
-        else if (code[j][2] > 0) { // deux parallèles
-          if (this.correction_detaillee) texte_corr += `.<br> Or «Si deux droites sont parallèles à une même droite alors elles sont parallèles entre elles».<br>Donc`
-          else texte_corr += `, on en déduit que `
-          texte_corr += ` $(d_${numDroites[code[0][0] - 1]})//(d_${numDroites[code[j + 1][1] - 1]})$.<br>`
-          code[j + 1][0] = code[j][0]
-          code[j + 1][2] = 1
-
-        }
-        else { //deux perpendiculaires
-          if (this.correction_detaillee) texte_corr += `.<br> Or «Si deux droites sont perpendiculaires à une même droite alors elles sont parallèles entre elles».<br>Donc`
-          else texte_corr += `, on en déduit que `
-          texte_corr += ` $(d_${numDroites[code[0][0] - 1]})//(d_${numDroites[code[j + 1][1] - 1]})$.<br>`
-          code[j + 1][0] = code[j][0]
-          code[j + 1][2] = 1
-
-        }
-
-
-      }
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte + "<br>");
-        this.liste_corrections.push(texte_corr + "<br>");
-        i++;
-      }
-      cpt++;
-    }
-
-    liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = ['Nombre d\'étapes de raisonnement :', 4, `1 : 1 étape\n 2 : 2 étapes\n 3 : 3 étapes\n4 : Mélange aléatoire`]
-  // this.besoin_formulaire2_case_a_cocher = [
-  //  "Avec figure ? ",false];
-}
-
-/**
- * Fonction générale pour exercices de constructions de parallèles et perpendiculaires
- * références 6G11, 6G12 et 6G12-1
- * @Auteur Jean-Claude Lhote
- */
-function Parallele_et_Perpendiculaires() {
-  "use strict";
-  Exercice.call(this);
-  this.titre = "Tracer des parallèles et des perpendiculaires";
-  this.nb_questions = 1;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 1;
-  this.sup2 = 1;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let type_de_questions_disponibles;
-    type_de_questions_disponibles = [parseInt(this.sup)]; // Le choix 1 ou 2 ou 3 : 1=perpendiculaires, 2=parallèles, 3=des perpendiculaires et des paralèlles
-    let liste_type_de_questions = combinaison_listes(
-      type_de_questions_disponibles,
-      this.nb_questions
-    );
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let Xmin, Xmax, Ymin, Ymax, ppc = 20, sc;
-
-    let A,
-      B,
-      C,
-      D,
-      xE,
-      E,
-      F,
-      CC,
-      DD,
-      EE,
-      d,
-      s1,
-      s2,
-      enonce,
-      correction,
-      dB,
-      dC,
-      dD,
-      dE,
-      g,
-      lC,
-      lD,
-      lE,
-      cB,
-      cC,
-      cD,
-      cE,
-      cF,
-      cG,
-      FF,
-      BB,
-      carreaux,
-      k,
-      objets_enonce = [],
-      objets_correction = [],
-
-      p;
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      objets_enonce.splice(0)
-      objets_correction.splice(0)
-      if (this.sup2 == 2)
-        k = 0.8;
-      else
-        k = 0.5;
-      switch (liste_type_de_questions[i]) {
-        case 1:
-          A = point(0, 0, "A", "above left");
-          B = point(10, randint(-4, 4, [-1, 0, 1]), "B", "above right");
-          d = droite(A, B);
-          d.isVisible = true;
-          C = point(randint(2, 3), randint(3, 4), "C", "above left");
-          D = point(randint(7, 8), randint(-7, -6), "D");
-          dB = droiteParPointEtPerpendiculaire(B, d);
-          xE = 11;
-          E = pointSurDroite(dB, 11, "E", "left");
-          while (!Number.isInteger(E.y)) {
-            xE++;
-            E = pointSurDroite(dB, xE, "E", "left");
-          }
-          F = point(E.x, B.y);
-          s1 = segment(B, F, "red");
-          s1.epaisseur = 2;
-          s1.pointilles = true;
-          s2 = segment(F, E, "blue");
-          s2.epaisseur = 2;
-          s2.pointilles = true;
-          dC = droiteParPointEtPerpendiculaire(C, d);
-          dD = droiteParPointEtPerpendiculaire(D, d);
-          BB = rotation(A, B, 90);
-          CC = pointIntersectionDD(dC, d, "M", "below right");
-          DD = pointIntersectionDD(dD, d, "N", "above left");
-          lC = arrondi(longueur(CC, A) * k, 1);
-          lD = arrondi(longueur(DD, A) * k, 1);
-          cB = codageAngleDroit(A, B, BB);
-          cC = codageAngleDroit(C, CC, B);
-          cD = codageAngleDroit(D, DD, B);
-          objets_correction.push(s1,
-            s2,
-            dC,
-            dD,
-            dB,
-            cB,
-            cC,
-            cD,
-            d,
-            tracePoint(A, B, C, D, E, CC, DD),
-            labelPoint(A, B, C, D, E, CC, DD),
-            afficheCoteSegment(
-              segment(A, CC),
-              `${tex_nombre(lC)} cm`,
-              0.5,
-              "red",
-              1,
-              0.5,
-              "red"
-            ),
-            afficheCoteSegment(
-              segment(A, DD),
-              `${tex_nombre(lD)} cm`,
-              -0.5,
-              "red",
-              1,
-              -0.5,
-              "red"
-            )
-          );
-          objets_enonce.push(
-            tracePoint(A, B, C, D),
-            labelPoint(A, B, C, D),
-            d,
-          );
-          if (sortie_html) enonce = num_alpha(0) + ` Reproduire la figure ci-dessous.<br>`;
-          else enonce = num_alpha(0) + ` Utiliser un crayon à papier afin de pouvoir gommer si besoin.<br>`;
-          enonce +=
-            num_alpha(1) +
-            ` Tracer la droite perpendiculaire à $(AB)$ passant par $B$.<br>`;
-          enonce +=
-            num_alpha(2) +
-            ` Tracer la droite perpendiculaire à $(AB)$ passant par $C$ et nommer $M$ le point d'intersection de cette droite avec la droite $(AB)$.<br>`;
-          enonce +=
-            num_alpha(3) +
-            ` Tracer la droite perpendiculaire à $(AB)$ passant par $D$ et nommer $N$ le point d'intersection de cette droite avec la droite $(AB)$.<br>`;
-          enonce +=
-            num_alpha(4) +
-            ` Mesurer ensuite les distances $AM$ et $AN$.<br> Pour l'auto-correction comparer ces mesures avec celles données dans la correction<br>`;
-          correction = `<br>$AM \\approx ${tex_nombre(
-            lC
-          )}$ cm et $AN \\approx ${tex_nombre(lD)}$ cm.<br>`;
-          correction += `Pour la perpendiculaire en $B$, contrôle la position du point $E$.<br>`;
-          Xmin = Math.floor(Math.min(A.x, B.x, C.x, D.x, E.x, CC.x, DD.x) - 1)
-          Xmax = Math.ceil(Math.max(A.x, B.x, C.x, D.x, E.x, CC.x, DD.x) + 1)
-          Ymin = Math.floor(Math.min(A.y, B.y, C.y, D.y, E.y, CC.y, DD.y) - 1)
-          Ymax = Math.ceil(Math.max(A.y, B.y, C.y, D.y, E.y, CC.y, DD.y) + 1)
-          break;
-        case 2:
-          A = point(2, 0, "A", 'below left');
-          B = point(12, randint(-4, 4, 0), "B");
-          d = droite(A, B);
-          d.isVisible = true;
-          C = point(0, randint(3, 4), "C", 'above');
-          D = point(randint(7, 8), randint(-7, -6), "D", 'below right');
-          E = point(randint(4, 5), randint(5, 6), "E", 'below right');
-          F = point(2, -3, "F", "left");
-
-          dE = droiteParPointEtParallele(E, d);
-          dC = droiteParPointEtParallele(C, d);
-          dD = droiteParPointEtParallele(D, d);
-          p = droite(A, F);
-          p.isVisible = true;
-          CC = pointIntersectionDD(dC, p, 'M', 'above left');
-          DD = pointIntersectionDD(dD, p, 'N', 'above left');
-          EE = pointIntersectionDD(dE, p, 'O', 'above left');
-          lC = arrondi(longueur(CC, A) * k, 1);
-          lD = arrondi(longueur(DD, A) * k, 1);
-          lE = arrondi(longueur(EE, A) * k, 1);
-          objets_correction.push(dC, dD, dE, d, p, tracePoint(A, B, C, D, E, F), labelPoint(A, B, C, D, E, F, CC, DD, EE), afficheCoteSegment(segment(A, CC), `${tex_nombre(lC)} cm`, .2, 'red', 1, 0.5, 'red'), afficheCoteSegment(segment(DD, A), `${tex_nombre(lD)} cm`, -0.2, 'green', 1, -0.5, 'green'), afficheCoteSegment(segment(A, EE), `${tex_nombre(lE)} cm`, -0.2, 'blue', 1, -0.5, 'blue'))
-          objets_enonce.push(tracePoint(A, B, C, D, E, F), labelPoint(A, B, C, D, E, F), d, p);
-
-          if (sortie_html) enonce = num_alpha(0) + ` Reproduire la figure ci-dessous.<br>`;
-          else enonce = num_alpha(0) + ` Utiliser un crayon à papier afin de pouvoir gommer si besoin.<br>`;
-          enonce += num_alpha(1) + ` Tracer la droite parallèle à $(AB)$ passant par $C$ et nommer $M$, le point d'intersection de cette droite avec la droite $(AF)$.<br>`;
-          enonce += num_alpha(2) + ` Tracer la droite parallèle à $(AB)$ passant par $D$ et nommer $N$, le point d'intersection de cette droite avec la droite $(AF)$.<br>`;
-          enonce += num_alpha(3) + ` Tracer la droite parallèle à $(AB)$ passant par $E$ et nommer $O$, le point d'intersection de cette droite avec la droite $(AF)$.<br>`;
-          enonce += num_alpha(4) + ` Mesurer les distances $AM$, $AN$ et $AO$. Pour l'auto-correction, comparer ces mesures avec celles données par  l'ordinateur dans la correction.<br>`;
-
-          correction = `<br>$AM \\approx ${tex_nombre(
-            lC
-          )}$ cm, $AN \\approx ${tex_nombre(
-            lD
-          )}$ cm et $AO \\approx${tex_nombre(
-            lE
-          )}$ cm.<br>`;
-          Xmin = Math.floor(Math.min(A.x, B.x, C.x, D.x, E.x, F.x, EE.x, CC.x, DD.x) - 1)
-          Xmax = Math.ceil(Math.max(A.x, B.x, C.x, D.x, E.x, F.x, EE.x, CC.x, DD.x) + 1)
-          Ymin = Math.floor(Math.min(A.y, B.y, C.y, D.y, E.y, F.y, EE.y, CC.y, DD.y) - 1)
-          Ymax = Math.ceil(Math.max(A.y, B.y, C.y, D.y, E.y, F.y, EE.y, CC.y, DD.y) + 1)
-          break;
-        case 3:
-          A = point(0, 0, "A", "above left");
-          B = point(10, randint(-4, 4, [-1, 0, 1]), "B", "above right");
-          d = droite(A, B);
-          d.isVisible = true;
-          C = point(randint(2, 3), randint(3, 4), "C", "above left");
-          D = point(randint(7, 8), randint(-7, -6), "D");
-          dB = droiteParPointEtPerpendiculaire(B, d);
-          xE = 11;
-          E = pointSurDroite(dB, 11, "E", "left");
-          while (!Number.isInteger(E.y)) {
-            xE++;
-            E = pointSurDroite(dB, xE, "E", "left");
-          }
-          F = point(E.x, B.y);
-          dE = droiteParPointEtParallele(E, d);
-          dD = droiteParPointEtParallele(D, d);
-          dC = droiteParPointEtPerpendiculaire(C, d)
-          BB = rotation(A, B, 90);
-          CC = pointIntersectionDD(dC, d, "M", "below right");
-          DD = pointIntersectionDD(dD, dB, "N", "above left");
-          EE = pointIntersectionDD(dC, dE, 'O', 'above left');
-          FF = pointIntersectionDD(dD, dC)
-
-          lC = arrondi(longueur(CC, A) * k, 1);
-          lD = arrondi(longueur(DD, A) * k, 1);
-          lE = arrondi(longueur(EE, A) * k, 1);
-          cB = codageAngleDroit(A, B, BB);
-          cC = codageAngleDroit(C, CC, B);
-          cD = codageAngleDroit(D, DD, B, 'red');
-          cE = codageAngleDroit(B, E, EE, 'red')
-          cF = codageAngleDroit(C, EE, E, 'red')
-          cG = codageAngleDroit(C, FF, D, 'red')
-
-          objets_correction.push(dC, dD, dB, dE, cB, cC, cD, cE, cF, cG, d, tracePoint(A, B, C, D, E, CC, DD, EE), labelPoint(A, B, C, D, E, CC, DD, EE), afficheCoteSegment(
-            segment(A, CC),
-            `${tex_nombre(lC)} cm`,
-            0.5,
-            "red",
-            1,
-            0.5,
-            "red"
-          ),
-            afficheCoteSegment(
-              segment(A, DD),
-              `${tex_nombre(lD)} cm`,
-              0,
-              "blue",
-              1,
-              -0.5,
-              "blue"
-            ),
-            afficheCoteSegment(
-              segment(A, EE),
-              `${tex_nombre(lE)} cm`,
-              0,
-              "green",
-              1,
-              -0.5,
-              "green"
-            ));
-          objets_enonce.push(tracePoint(A, B, C, D, E), labelPoint(A, B, C, D, E), d);
-          if (sortie_html) enonce = num_alpha(0) + ` Reproduire la figure ci-dessous.<br>`;
-          else enonce = num_alpha(0) + ` Utiliser un crayon à papier afin de pouvoir gommer si besoin.<br>`;
-          enonce += num_alpha(1) + ` Tracer la droite perpendiculaire à $(AB)$ passant par $B$.<br>`;
-          enonce += num_alpha(2) + ` Tracer la droite perpendiculaire à $(AB)$ passant par $C$ et nomme $M$, le point d'intersection de cette droite avec la droite $(AB)$.<br>`
-          enonce += num_alpha(3) + ` Tracer la droite parallèle à $(AB)$ passant par $D$ et nomme $N$, le point d'intersection de cette droite avec la droite $(BE)$.<br>`;
-          enonce += num_alpha(4) + ` Tracer la droite parallèle à $(AB)$ passant par $E$ et nomme $O$, le point d'intersection de cette droite avec la droite $(CM)$.<br>`
-          enonce += num_alpha(5) + ` Mesurer les distances $AM$, $AN$ et $AO$. Pour l'auto-correction, comparer ces mesures avec celles données par  l'ordinateur dans la correction.<br>`;
-
-          correction = `<br>$AM \\approx ${tex_nombre(
-            lC
-          )}$ cm, $AN \\approx ${tex_nombre(
-            lD
-          )}$ cm et $AO \\approx${tex_nombre(
-            lE
-          )}$ cm.<br>`;
-          correction += `Les angle droits en rouge se justifient par la propriété :<br> ${texte_en_couleur(`Si deux droites sont parallèles, alors toute droite perpendiculaire à l'une est aussi perpendiculaire à l'autre`, 'red')}.<br>`
-          correction += `Vérifier les angles droits à l'équerre.<br>`
-          Xmin = Math.floor(Math.min(A.x, B.x, C.x, D.x, E.x, F.x, EE.x, CC.x, DD.x) - 1)
-          Xmax = Math.ceil(Math.max(A.x, B.x, C.x, D.x, E.x, F.x, EE.x, CC.x, DD.x) + 1)
-          Ymin = Math.floor(Math.min(A.y, B.y, C.y, D.y, E.y, F.y, EE.y, CC.y, DD.y) - 1)
-          Ymax = Math.ceil(Math.max(A.y, B.y, C.y, D.y, E.y, F.y, EE.y, CC.y, DD.y) + 1)
-
-          break
-      }
-      if (this.sup2 < 3) g = grille(Xmin, Ymin, Xmax, Ymax, "gray", 0.7);
-      else g = "";
-      if (this.sup2 == 2) {
-        sc = 0.8;
-        carreaux = seyes(Xmin, Ymin, Xmax, Ymax);
-      } else {
-        sc = 0.5;
-        carreaux = "";
-      }
-      objets_enonce.push(g, carreaux)
-      objets_correction.push(g, carreaux)
-
-      enonce += mathalea2d(
-        {
-          xmin: Xmin,
-          ymin: Ymin,
-          xmax: Xmax,
-          ymax: Ymax,
-          pixelsParCm: ppc,
-          scale: sc,
-        },
-        objets_enonce
-      );
-      correction += mathalea2d(
-        {
-          xmin: Xmin,
-          ymin: Ymin,
-          xmax: Xmax,
-          ymax: Ymax,
-          pixelsParCm: ppc,
-          scale: sc,
-        },
-        objets_correction
-      );
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(enonce + "<br>");
-        this.liste_corrections.push(correction + "<br>");
-        i++;
-      }
-      cpt++;
-    }
-
-    liste_de_question_to_contenu(this);
-  };
-  //	this.besoin_formulaire_numerique = ['Type de questions', 3, `1 : Perpendiculaires\n 2 : Parallèles\n 3 : Mélange`]
-  this.besoin_formulaire2_numerique = [
-    "Type de cahier",
-    3,
-    `1 : Cahier à petits careaux\n 2 : Cahier à gros carreaux (Seyes)\n 3 : Feuille blanche`,
-  ];
-}
 
 
 /**
@@ -11854,935 +7989,6 @@ function Organiser_donnees_depuis_texte() {
   this.besoin_formulaire2_numerique = ['Nombre de fruits différents', 8];
   this.besoin_formulaire_case_a_cocher = ['Pour utiliser des nombres décimaux et des masses', false];
 }
-
-/**
- * Un nombre à 2 chiffres (non multiple de 10) + 9
- * @Auteur Rémi Angot
- * Référence 6G23-2
- */
-function Tracer_triangle_2_angles() {
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.titre = "Tracer un triangle dont on connait une longueur et 2 angles";
-  this.consigne = "";
-  this.nb_questions = 3;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      let c = calcul(randint(30, 70) / 10);
-      let angle1 = randint(20, 70);
-      let angle2 = randint(20, 70);
-      let p = creerNomDePolygone(3);
-      texte = `Tracer un triangle $${p}$ tel que $${p[0] + p[1]}=${tex_nombre(
-        c
-      )}$ cm, $\\widehat{${p[1] + p[0] + p[2]
-        }}=${angle1}\\degree$ et $\\widehat{${p[0] + p[1] + p[2]
-        }}=${angle2}\\degree$.`;
-      texte += `<br> Mesurer $${p[0] + p[2]}$ et $${p[1] + p[2]}$.`;
-      let A0 = point(0, 0, p[0], "left");
-      let B0 = point(c, 0, p[1], "right");
-      let s0 = segmentAvecExtremites(A0, B0);
-      let t1 = afficheLongueurSegment(B0, A0);
-
-      let A1 = point(B0.x + 4, 0, p[0], "left");
-      let B1 = point(A1.x + c, 0, p[1], "right");
-      let s1 = segment(A1, B1);
-      s1.styleExtremites = "-|";
-      let c1 = rotation(B1, A1, angle1);
-      let C1 = pointSurSegment(c1, A1, -3);
-      let s2 = segment(A1, C1);
-      let t2 = afficheMesureAngle(B1, A1, C1);
-
-      let A2 = point(B1.x + 4, 0, p[0], "left");
-      let B2 = point(A2.x + c, 0, p[1], "right");
-      let s3 = segment(A2, B2);
-      let c2 = rotation(B2, A2, angle1);
-      let C2 = pointSurSegment(c2, A2, -3);
-      let s4 = segment(A2, C2);
-      let c3 = rotation(A2, B2, -angle2);
-      let C3 = pointSurSegment(c3, B2, -3);
-      let t3 = afficheMesureAngle(A2, B2, C3);
-      let s5 = segment(B2, C3);
-      let d1 = droite(A2, C2);
-      d1.isVisible = false;
-      let d2 = droite(B2, C3);
-      d2.isVisible = false;
-      let C = pointIntersectionDD(d1, d2, p[2]);
-      let l = labelPoint(A0, B0, A1, B1, A2, B2, C);
-
-      if (sortie_html) {
-        texte_corr = mathalea2d(
-          { xmin: -1, xmax: 3 * c + 10, ymin: -1, ymax: 10 },
-          s0,
-          s1,
-          s2,
-          s3,
-          s4,
-          s5,
-          t1,
-          t2,
-          t3,
-          l
-        );
-        texte_corr += '<br><br>'
-      } else {
-        texte_corr = "";
-      }
-      texte_corr += `$${p[0] + p[2]}\\approx${tex_nombre(
-        longueur(A2, C, 1)
-      )}$ cm et $${p[1] + p[2]}\\approx${tex_nombre(longueur(B2, C, 1))}$ cm.`;
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  };
-  //this.besoin_formulaire_numerique = ['Niveau de difficulté',3];
-}
-
-/**
- * fonction servant à compléter des solides, inspirée des fonctions de 6G42 et 6G43
- * référence : 6G41
- * @Auteur Mireille Gain, s'inspirant fortement de Jean-Claude Lhote
- */
-function Representer_un_solide() {
-  "use strict";
-  Exercice.call(this);// Héritage de la classe Exercice ()
-  this.titre = "Compléter une représentation en perspective cavalière";
-  this.nb_questions = 1;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 1;
-  this.sup2 = 1;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let type_de_questions_disponibles;
-
-    if (this.sup == 3) type_de_questions_disponibles = [1, 2]
-    else if (this.sup == 5) type_de_questions_disponibles = [1, 2, 4]
-    else if (this.sup == 7) type_de_questions_disponibles = [1, 2, 4, 6]
-    else type_de_questions_disponibles = [parseInt(this.sup)];
-
-    let liste_type_de_questions = combinaison_listes(
-      type_de_questions_disponibles,
-      this.nb_questions
-    );// Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let Xmin, Xmax, Ymin, Ymax, ppc, sc;
-
-    if (this.classe == 6) type_de_questions_disponibles = [1, 2];
-    // sixième : cube et pavé droit
-    else if (this.classe == 5) type_de_questions_disponibles = [1, 2, 4];
-    // cinquième : on ajoute le prisme
-    else if (this.classe == 4) type_de_questions_disponibles = [1, 2, 4, 6]
-    // Quatrième : on ajoute la pyramide
-
-
-    if (this.sup2 == 1) sc = 0.5;
-    else sc = 0.8;
-
-    let A, B, C, D, E, F, G, H, I,
-      AB, BC, CD, DA, EF, FG, GH, HE, AE, BF, CG, DH, IA, IB, IE, IF, BD, FH,
-      coeffpersp,
-      enonce,
-      correction,
-      carreaux, g,
-      objets_enonce = [],
-      objets_correction = [], p,
-      matrace = tracePoint(I)
-      ;
-
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;) {
-      let nom = creerNomDePolygone(8, "PQ"),
-        anglepersp = choice([30, 45, -30, -45, 150, 135, -150, -135])
-      if (anglepersp % 10 == 0) coeffpersp = 0.6
-      else coeffpersp = 0.4
-      objets_correction = []
-      objets_enonce = []
-
-      switch (liste_type_de_questions[i]) {
-        case 1: //cube
-          enonce = `$${nom}$ est un cube.<br>`
-          if (sortie_html) enonce += `Reproduire et compléter la figure ci-dessous, en repassant de la même couleur les segments parallèles et de même longueur.<br>`;
-          correction = `Figure complétée :<br>`
-          break;
-
-        case 2: //pavé droit
-          enonce = `$${nom}$ est un pavé droit.<br>`
-          if (sortie_html) enonce += `Reproduire et compléter la figure ci-dessous, en repassant de la même couleur les segments parallèles et de même longueur.<br>`;
-          correction = `Figure complétée :<br>`
-          break;
-
-        case 4: //prisme
-          enonce = `On considère un prisme à base triangulaire.<br>Reproduire et compléter la figure ci-dessous, en repassant de la même couleur les segments parallèles et de même longueur.<br>`;
-          correction = `Figure complétée :<br>`
-          break;
-
-        case 6: //pyramide
-          enonce = `On considère une pyramide à base rectangulaire.<br>Reproduire et compléter la figure ci-dessous, en repassant de la même couleur les segments parallèles et de même longueur.<br>`;
-          correction = `Figure complétée :<br>`
-          break;
-      }
-
-      switch (liste_type_de_questions[i] % 2) {
-        case 1:
-          A = point(6, 0, nom[0], "left");
-          B = point(11, 0, nom[1], "right");
-          C = point(11, 5, nom[2], "right");
-          D = point(6, 5, nom[3], "left");
-          E = similitude(B, A, anglepersp, coeffpersp, nom[4], 'left')
-          E.x = Math.round(E.x)
-          E.y = Math.round(E.y)
-          break;
-
-        case 0:
-          A = point(5, 0, nom[0], "left");
-          B = point(9 + randint(1, 3), 0, nom[1], "right");
-          C = point(B.x, randint(3, 7), nom[2], "right");
-          D = point(A.x, C.y, nom[3], "left");
-          E = similitude(B, A, anglepersp, coeffpersp * randint(5, 12) / 10, nom[4], 'left')
-          E.x = Math.round(E.x)
-          E.y = Math.round(E.y)
-          break;
-      }
-
-      F = translation2Points(E, A, B, nom[5], 'right')
-      G = translation2Points(F, B, C, nom[6], 'right')
-      H = translation2Points(G, C, D, nom[7], 'left')
-      I = milieu(D, G)
-      matrace.taille = 4
-      matrace.opacite = 0.9
-      matrace.epaisseur = 5
-      matrace.style = 'x'
-      AB = segment(A, B)
-      BC = segment(B, C)
-      CD = segment(C, D)
-      DA = segment(D, A)
-      EF = segment(E, F)
-      FG = segment(F, G)
-      GH = segment(G, H)
-      HE = segment(H, E)
-      AE = segment(A, E)
-      BF = segment(B, F)
-      CG = segment(C, G)
-      DH = segment(D, H)
-      IA = segment(A, I)
-      IB = segment(B, I)
-      IE = segment(E, I)
-      IF = segment(F, I)
-      BD = segment(B, D)
-      FH = segment(F, H)
-      AB.epaisseur = 2
-      BC.epaisseur = 2
-      CD.epaisseur = 2
-      DA.epaisseur = 2
-      EF.epaisseur = 2
-      FG.epaisseur = 2
-      GH.epaisseur = 2
-      HE.epaisseur = 2
-      AE.epaisseur = 2
-      BF.epaisseur = 2
-      CG.epaisseur = 2
-      DH.epaisseur = 2
-      IA.epaisseur = 1
-      IB.epaisseur = 1
-      IE.epaisseur = 1
-      IF.epaisseur = 1
-      BD.epaisseur = 2
-      FH.epaisseur = 2
-      AB.color = 'black'
-      BC.color = 'black'
-      CD.color = 'black'
-      DA.color = 'black'
-      EF.color = 'black'
-      FG.color = 'black'
-      GH.color = 'black'
-      HE.color = 'black'
-      AE.color = 'black'
-      BF.color = 'black'
-      CG.color = 'black'
-      IA.color = 'black'
-
-      if (G.y < C.y && G.x < C.x) {
-        CG.pointilles = true
-        GH.pointilles = true
-        FG.pointilles = true
-        IF.pointilles = true
-        FH.pointilles = true
-        CG.color = 'gray'
-        GH.color = 'gray'
-        FG.color = 'gray'
-        CG.opacite = 0.7
-        GH.opacite = 0.7
-        FG.opacite = 0.7
-      }
-      else if (E.y > A.y && E.x > A.x) {
-        AE.pointilles = true
-        EF.pointilles = true
-        HE.pointilles = true
-        IE.pointilles = true
-        FH.pointilles = true
-        AE.color = 'gray'
-        EF.color = 'gray'
-        HE.color = 'gray'
-        AE.opacite = 0.7
-        EF.opacite = 0.7
-        HE.opacite = 0.7
-      }
-      else if (F.x < B.x && F.y > B.y) {
-        BF.pointilles = true
-        FG.pointilles = true
-        EF.pointilles = true
-        IF.pointilles = true
-        FH.pointilles = true
-        BF.color = 'gray'
-        FG.color = 'gray'
-        EF.color = 'gray'
-        BF.opacite = 0.7
-        FG.opacite = 0.7
-        EF.opacite = 0.7
-      }
-      else if (H.x > D.x && H.y < D.y) {
-        DH.pointilles = true
-        GH.pointilles = true
-        HE.pointilles = true
-        IE.pointilles = true
-        FH.pointilles = true
-        DH.color = 'gray'
-        GH.color = 'gray'
-        HE.color = 'gray'
-        DH.opacite = 0.7
-        GH.opacite = 0.7
-        HE.opacite = 0.7
-      }
-      Xmin = Math.min(A.x, E.x) - 1;
-      Ymin = Math.min(A.y, E.y) - 1;
-      Xmax = Math.max(B.x, F.x) + 2;
-      Ymax = Math.max(D.y, H.y) + 1;
-      ppc = 20;
-
-      if (this.sup2 < 3) { g = grille(Xmin, Ymin, Xmax, Ymax, "gray", 0.7); }
-      else g = "";
-      if (this.sup2 == 2) { carreaux = seyes(Xmin, Ymin, Xmax, Ymax); sc = 0.8 }
-      else { carreaux = ""; sc = 0.5 }
-
-      let params = {
-        xmin: Xmin,
-        ymin: Ymin,
-        xmax: Xmax,
-        ymax: Ymax,
-        pixelsParCm: ppc,
-        scale: sc,
-      }
-
-      if (liste_type_de_questions[i] == 1) {
-        objets_enonce.push(AB, BC, CD, DA, AE, labelPoint(A, B, C, D, E),
-          g,
-          carreaux
-        );
-      }
-
-      if (liste_type_de_questions[i] == 2) {
-        objets_enonce.push(AB, BC, CD, DA, AE, labelPoint(A, B, C, D, E),
-          g,
-          carreaux
-        );
-      }
-
-      if (liste_type_de_questions[i] == 4) {
-        objets_enonce.push(AB, DA, BD, AE,
-          g,
-          carreaux
-        );
-      }
-
-      if (liste_type_de_questions[i] == 6) {
-        objets_enonce.push(AB, BF, tracePoint(I, .5, 'red'), labelPoint(I),
-          g,
-          carreaux
-        );
-      }
-
-      enonce += mathalea2d(params, objets_enonce);
-      if (liste_type_de_questions[i] == 1) {
-        AB.color = 'green'
-        BC.color = 'red'
-        CD.color = 'green'
-        DA.color = 'red'
-        EF.color = 'green'
-        FG.color = 'red'
-        GH.color = 'green'
-        HE.color = 'red'
-        AE.color = 'blue'
-        BF.color = 'blue'
-        CG.color = 'blue'
-        DH.color = 'blue'
-        objets_correction.push(AB, BC, CD, DA, EF, FG, GH, HE, AE, BF, CG, DH, labelPoint(A, B, C, D, E, F, G, H),
-          g,
-          carreaux
-        );
-      }
-
-      if (liste_type_de_questions[i] == 2) {
-        AB.color = 'green'
-        BC.color = 'red'
-        CD.color = 'green'
-        DA.color = 'red'
-        EF.color = 'green'
-        FG.color = 'red'
-        GH.color = 'green'
-        HE.color = 'red'
-        AE.color = 'blue'
-        BF.color = 'blue'
-        CG.color = 'blue'
-        DH.color = 'blue'
-        objets_correction.push(AB, BC, CD, DA, EF, FG, GH, HE, AE, BF, CG, DH, labelPoint(A, B, C, D, E, F, G, H),
-          g,
-          carreaux
-        );
-      }
-
-      if (liste_type_de_questions[i] == 4) {
-        AB.color = 'green'
-        BC.color = 'red'
-        CD.color = 'green'
-        DA.color = 'red'
-        EF.color = 'green'
-        FG.color = 'red'
-        GH.color = 'green'
-        HE.color = 'red'
-        AE.color = 'blue'
-        BF.color = 'blue'
-        CG.color = 'blue'
-        DH.color = 'blue'
-        objets_correction.push(AB, DA, BD, EF, HE, AE, BF, DH, FH,
-          g,
-          carreaux
-        );
-      }
-
-      if (liste_type_de_questions[i] == 6) {
-        AB.color = 'green'
-        BC.color = 'red'
-        CD.color = 'green'
-        DA.color = 'red'
-        EF.color = 'green'
-        FG.color = 'red'
-        GH.color = 'green'
-        HE.color = 'red'
-        AE.color = 'blue'
-        BF.color = 'blue'
-        CG.color = 'blue'
-        DH.color = 'blue'
-        objets_correction.push(AB, EF, AE, BF, IA, IB, IE, IF, tracePoint(I),
-          g,
-          carreaux
-        );
-      }
-
-      correction += mathalea2d(params, objets_correction);
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(enonce + "<br>");
-        this.liste_corrections.push(correction + "<br>");
-        i++;
-      }
-      cpt++;
-    }
-
-    liste_de_question_to_contenu(this);
-  };
-  if (this.classe == 6) this.besoin_formulaire_numerique = ['Type de solides', 3, `1 : Cubes\n 2 : Pavés droits\n 3 : Mélange cubes et pavés`]
-  if (this.classe == 5) this.besoin_formulaire_numerique = ['Type de solides', 5, `1 : Cubes\n 2 : Pavés droits\n 3 : Mélange cubes et pavés\n 4 : Prismes\n 5 : Mélange cubes, pavés, prismes`]
-  if (this.classe == 4) this.besoin_formulaire_numerique = ['Type de solides', 7, `1 : Cubes\n 2 : Pavés droits\n 3 : Mélange cubes et pavés\n 4 : Prismes\n 5 : Mélange cubes, pavés, prismes\n 6 : Pyramides\n 7 : Mélange cubes, pavés, prismes, pyramides`]
-
-  this.besoin_formulaire2_numerique = [
-    "Type de cahier",
-    3,
-    `1 : Cahier à petits careaux\n 2 : Cahier à gros carreaux (Seyes)\n 3 : Feuille blanche`,
-  ];
-}
-
-
-// Référence 6G41
-function Representer_un_solide_6e() {
-  this.sup = 1;
-  this.classe = 6;
-  Representer_un_solide.call(this);
-}
-
-// 5G51
-function Representer_un_solide_5e() {
-  this.sup = 4;
-  this.classe = 5;
-  Representer_un_solide.call(this);
-}
-
-/**
- * fonction servant à plusieurs exercice autour du cube et du pavé droit
- * références : 6G42 et 6G43
- * @Auteur Jean-Claude Lhote
- */
-function Solide_6e() {
-  "use strict";
-  Exercice.call(this);
-  this.titre = "Connaître les propriétés du cube et du pavé droit";
-  this.nb_questions = 1;
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  this.sup = 1;
-  this.sup2 = 1;
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    let type_de_questions_disponibles;
-    if (this.sup == 3) type_de_questions_disponibles = [1, 2]
-    else type_de_questions_disponibles = [parseInt(this.sup)];
-
-    if (this.titre == "Utiliser le vocabulaire associé au pavé droit")
-      for (let n = 0; n < type_de_questions_disponibles.length; n++)
-        type_de_questions_disponibles[n] += 2
-    let liste_type_de_questions = combinaison_listes(
-      type_de_questions_disponibles,
-      this.nb_questions
-    );
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-    let Xmin, Xmax, Ymin, Ymax, ppc, sc;
-
-    if (this.sup2 == 1) sc = 0.5;
-    else sc = 0.8;
-
-    let A, B, C, D, E, F, G, H,
-      AB, BC, CD, DA, EF, FG, GH, HE, AE, BF, CG, DH,
-      lAB, lBC, lAE, anglepersp, coeffpersp,
-      codesseg = [],
-      enonce,
-      correction,
-      carreaux, g,
-      objets_enonce = [],
-      objets_correction = [],
-      k,
-      p;
-    for (
-      let i = 0, texte, texte_corr, cpt = 0;
-      i < this.nb_questions && cpt < 50;
-
-    ) {
-      let nom = creerNomDePolygone(8, "PQ"),
-        anglepersp = choice([30, 45, -30, -45, 150, 135, -150, -135])
-      if (anglepersp % 10 == 0) coeffpersp = 0.6
-      else coeffpersp = 0.4
-      objets_correction = []
-      objets_enonce = []
-      switch (liste_type_de_questions[i]) {
-        case 1: //cube
-          enonce = `${nom} est un cube.<br>`
-          if (sortie_html) enonce += ` Reproduire la figure ci-dessous sur le cahier.<br>`;
-          enonce += ` Repasse tous les segments de même longueur dans une même couleur.<br>`;
-          correction = `Le cube ${nom}.<br>`
-          break;
-
-        case 2:
-          enonce = `${nom} est un pavé droit.<br>`
-          if (sortie_html) enonce += ` Reproduire la figure ci-dessous sur le cahier.<br>`;
-          enonce += ` Repasse tous les segments de même longueur dans une même couleur.<br>`;
-          correction = `Le pavé droit ${nom}.<br>`
-          break;
-
-        case 3:
-          enonce = `${nom} est un cube.<br>`
-          break;
-
-        case 4:
-          enonce = `${nom} est un pavé droit.<br>`
-          break;
-      }
-      let aretes_paralleles = [[[0, 1], [3, 2], [4, 5], [7, 6]], [[0, 3], [1, 2], [4, 7], [5, 6]], [[0, 4], [1, 5], [2, 6], [3, 7]]]
-      let faces_paralleles = [[[0, 1, 2, 3], [4, 5, 6, 7]], [[0, 4, 7, 3], [1, 5, 6, 2]], [[0, 1, 5, 4], [3, 2, 6, 7]]]
-      let aretes_perp = [[[0, 1], [0, 4], [0, 3], [1, 5], [1, 2]], [[0, 4], [0, 1], [0, 3], [4, 5], [4, 7]], [[0, 3], [0, 1], [0, 4], [3, 2], [3, 7]], [[1, 2], [1, 0], [1, 5], [2, 3], [2, 6]], [[1, 5], [1, 0], [1, 2],
-      [5, 4], [5, 6]], [[5, 4], [5, 1], [5, 6], [4, 0], [4, 7]], [[5, 6], [5, 1], [5, 4], [6, 2], [6, 7]]
-        , [[6, 2], [6, 5], [6, 7], [2, 1], [2, 3]], [[2, 3], [2, 1], [2, 6], [3, 0], [3, 7]], [[3, 7], [3, 2], [3, 0], [7, 4], [7, 6]], [[7, 4], [4, 0], [4, 5], [7, 3], [7, 6]], [[7, 6], [6, 2], [6, 5], [7, 3], [7, 4]]]
-      let faces_perp = [[[0, 1, 2, 3], [1, 5, 6, 2], [2, 6, 7, 3], [3, 7, 4, 0], [0, 1, 5, 4]], [[1, 5, 6, 2], [0, 1, 2, 3], [2, 6, 7, 3], [5, 6, 7, 4], [1, 5, 4, 0]], [[0, 1, 5, 4], [1, 5, 6, 2], [4, 5, 6, 7], [0, 4, 7, 3], [0, 1, 2, 3]],
-      [[4, 5, 6, 7], [0, 1, 5, 4], [1, 5, 6, 2], [2, 6, 7, 3], [0, 4, 7, 3]], [[0, 4, 7, 3], [0, 1, 2, 3], [0, 1, 5, 4], [4, 5, 6, 7], [3, 2, 6, 7]], [[3, 2, 6, 7], [0, 1, 2, 3], [1, 5, 6, 2], [4, 5, 6, 7], [0, 4, 7, 3]]]
-      let k, l, s
-
-      switch (randint(1, 4)) {
-
-        case 1: // citer les arêtes parallèles à une arrête donnée
-          [k, l, s] = [randint(0, 2), randint(0, 3), randint(0, 1)]
-          enonce += `Citer toutes les arêtes parallèles à [$${nom[aretes_paralleles[k][l][s]] + nom[aretes_paralleles[k][l][(s + 1) % 2]]}$].<br>`
-          correction = `Les arêtes parallèles à [$${nom[aretes_paralleles[k][l][s]] + nom[aretes_paralleles[k][l][(s + 1) % 2]]}$] sont [$${nom[aretes_paralleles[k][(l + 1) % 4][s]] + nom[aretes_paralleles[k][(l + 1) % 4][(s + 1) % 2]]}$], [$${nom[aretes_paralleles[k][(l + 2) % 4][s]] + nom[aretes_paralleles[k][(l + 2) % 4][(s + 1) % 2]]}$] et [$${nom[aretes_paralleles[k][(l + 3) % 4][s]] + nom[aretes_paralleles[k][(l + 3) % 4][(s + 1) % 2]]}$].<br>`
-          break;
-
-        case 2: // citer la face parallèle à une face donnée
-          [k, l, s] = [randint(0, 2), randint(0, 1), randint(0, 3)]
-          enonce += `Quelle est la face parallèle à $${nom[faces_paralleles[k][l][s]] + nom[faces_paralleles[k][l][(s + 1) % 4]] + nom[faces_paralleles[k][l][(s + 2) % 4]] + nom[faces_paralleles[k][l][(s + 3) % 4]]}$ ?<br>`
-          correction = `La face parallèle à $${nom[faces_paralleles[k][l][s]] + nom[faces_paralleles[k][l][(s + 1) % 4]] + nom[faces_paralleles[k][l][(s + 2) % 4]] + nom[faces_paralleles[k][l][(s + 3) % 4]]}$ est la face $${nom[faces_paralleles[k][(l + 1) % 2][s]] + nom[faces_paralleles[k][(l + 1) % 2][(s + 1) % 4]] + nom[faces_paralleles[k][(l + 1) % 2][(s + 2) % 4]] + nom[faces_paralleles[k][(l + 1) % 2][(s + 3) % 4]]}$.<br>`
-          break;
-
-        case 3: // citer les arêtes perpendiculaires à une arête donnée
-          [k, l, s] = [randint(0, 11), 0, randint(0, 1)]
-          enonce += `Quelles sont les arêtes peprendiculaires à l'arête [$${nom[aretes_perp[k][l][s]] + nom[aretes_perp[k][l][(s + 1) % 2]]}$] ?<br>`
-          correction = `Les arêtes perpendiculaires à l'arête [$${nom[aretes_perp[k][l][s]] + nom[aretes_perp[k][l][(s + 1) % 2]]}$] sont [$${nom[aretes_perp[k][1][s]] + nom[aretes_perp[k][1][(s + 1) % 2]]}$], [$${nom[aretes_perp[k][2][s]] + nom[aretes_perp[k][2][(s + 1) % 2]]}$], [$${nom[aretes_perp[k][3][s]] + nom[aretes_perp[k][3][(s + 1) % 2]]}$] et [$${nom[aretes_perp[k][4][s]] + nom[aretes_perp[k][4][(s + 1) % 2]]}$].`
-          break;
-
-        case 4: // citer les faces perpendiculaires à une face donnée
-          [k, l, s] = [randint(0, 5), 0, randint(0, 3)]
-          enonce += `Quelles sont les faces perpendiculaires à la face $${nom[faces_perp[k][l][s]] + nom[faces_perp[k][l][(s + 1) % 4]] + nom[faces_perp[k][l][(s + 2) % 4]] + nom[faces_perp[k][l][(s + 3) % 4]]}$ ?<br>`
-          correction = `Les faces perpendiculaires à la face $${nom[faces_perp[k][l][s]] + nom[faces_perp[k][l][(s + 1) % 4]] + nom[faces_perp[k][l][(s + 2) % 4]] + nom[faces_perp[k][l][(s + 3) % 4]]}$ `
-          correction += `sont les faces $${nom[faces_perp[k][l + 1][s]] + nom[faces_perp[k][l + 1][(s + 1) % 4]] + nom[faces_perp[k][l + 1][(s + 2) % 4]] + nom[faces_perp[k][l + 1][(s + 3) % 4]]}$, `
-          correction += `$${nom[faces_perp[k][l + 2][s]] + nom[faces_perp[k][l + 2][(s + 1) % 4]] + nom[faces_perp[k][l + 2][(s + 2) % 4]] + nom[faces_perp[k][l + 2][(s + 3) % 4]]}$, `
-          correction += `$${nom[faces_perp[k][l + 3][s]] + nom[faces_perp[k][l + 3][(s + 1) % 4]] + nom[faces_perp[k][l + 3][(s + 2) % 4]] + nom[faces_perp[k][l + 3][(s + 3) % 4]]}$ et `
-          correction += `$${nom[faces_perp[k][l + 4][s]] + nom[faces_perp[k][l + 4][(s + 1) % 4]] + nom[faces_perp[k][l + 4][(s + 2) % 4]] + nom[faces_perp[k][l + 4][(s + 3) % 4]]}$.`
-          break;
-      }
-
-      switch (liste_type_de_questions[i] % 2) {
-        case 1:
-          A = point(6, 0, nom[0], "left");
-          B = point(11, 0, nom[1], "right");
-          C = point(11, 5, nom[2], "right");
-          D = point(6, 5, nom[3], "left");
-          p = polygone(A, B, C, D)
-          E = similitude(B, A, anglepersp, coeffpersp, nom[4], 'left')
-          E.x = Math.round(E.x)
-          E.y = Math.round(E.y)
-          break;
-
-        case 0:
-          A = point(5, 0, nom[0], "left");
-          B = point(9 + randint(1, 3), 0, nom[1], "right");
-          C = point(B.x, randint(3, 7), nom[2], "right");
-          D = point(A.x, C.y, nom[3], "left");
-          p = polygone(A, B, C, D)
-          E = similitude(B, A, anglepersp, coeffpersp * randint(5, 12) / 10, nom[4], 'left')
-          E.x = Math.round(E.x)
-          E.y = Math.round(E.y)
-          break;
-      }
-
-      p = polygone(A, B, C, D)
-      F = translation2Points(E, A, B, nom[5], 'right')
-      G = translation2Points(F, B, C, nom[6], 'right')
-      H = translation2Points(G, C, D, nom[7], 'left')
-      AB = segment(A, B)
-      BC = segment(B, C)
-      CD = segment(C, D)
-      DA = segment(D, A)
-      EF = segment(E, F)
-      FG = segment(F, G)
-      GH = segment(G, H)
-      HE = segment(H, E)
-      AE = segment(A, E)
-      BF = segment(B, F)
-      CG = segment(C, G)
-      DH = segment(D, H)
-      AB.epaisseur = 2
-      BC.epaisseur = 2
-      CD.epaisseur = 2
-      DA.epaisseur = 2
-      EF.epaisseur = 2
-      FG.epaisseur = 2
-      GH.epaisseur = 2
-      HE.epaisseur = 2
-      AE.epaisseur = 2
-      BF.epaisseur = 2
-      CG.epaisseur = 2
-      DH.epaisseur = 2
-      AB.color = 'black'
-      BC.color = 'black'
-      CD.color = 'black'
-      DA.color = 'black'
-      EF.color = 'black'
-      FG.color = 'black'
-      GH.color = 'black'
-      HE.color = 'black'
-      AE.color = 'black'
-      BF.color = 'black'
-      CG.color = 'black'
-      if (G.y < C.y && G.x < C.x) {
-        CG.pointilles = true
-        CG.color = 'gray'
-        CG.opacite = 0.7
-        GH.pointilles = true
-        GH.color = 'gray'
-        GH.opacite = 0.7
-        FG.pointilles = true
-        FG.color = 'gray'
-        FG.opacite = 0.7
-      }
-      else if (E.y > A.y && E.x > A.x) {
-        AE.pointilles = true
-        EF.pointilles = true
-        HE.pointilles = true
-        AE.color = 'gray'
-        EF.color = 'gray'
-        HE.color = 'gray'
-        AE.opacite = 0.7
-        EF.opacite = 0.7
-        HE.opacite = 0.7
-      }
-      else if (F.x < B.x && F.y > B.y) {
-        BF.pointilles = true
-        FG.pointilles = true
-        EF.pointilles = true
-        BF.color = 'gray'
-        FG.color = 'gray'
-        EF.color = 'gray'
-        BF.opacite = 0.7
-        FG.opacite = 0.7
-        EF.opacite = 0.7
-      }
-      else if (H.x > D.x && H.y < D.y) {
-        DH.pointilles = true
-        GH.pointilles = true
-        HE.pointilles = true
-        DH.color = 'gray'
-        GH.color = 'gray'
-        HE.color = 'gray'
-        DH.opacite = 0.7
-        GH.opacite = 0.7
-        HE.opacite = 0.7
-      }
-      Xmin = Math.min(A.x, E.x) - 1;
-      Ymin = Math.min(A.y, E.y) - 1;
-      Xmax = Math.max(B.x, F.x) + 2;
-      Ymax = Math.max(D.y, H.y) + 1;
-      ppc = 20;
-
-      if (this.sup2 < 3) g = grille(Xmin, Ymin, Xmax, Ymax, "gray", 0.7);
-      else g = "";
-      if (this.sup2 == 2) {
-        carreaux = seyes(Xmin, Ymin, Xmax, Ymax);
-        sc = 0.8
-      }
-      else {
-        carreaux = "";
-        sc = 0.5
-      }
-      objets_enonce.push(AB, BC, CD, DA, EF, FG, GH, HE, AE, BF, CG, DH, labelPoint(A, B, C, D, E, F, G, H), p,
-        g,
-        carreaux
-      );
-
-      let params = {
-        xmin: Xmin,
-        ymin: Ymin,
-        xmax: Xmax,
-        ymax: Ymax,
-        pixelsParCm: ppc,
-        scale: sc,
-      }
-
-      enonce += mathalea2d(params, objets_enonce);
-      if (liste_type_de_questions[i] == 1) {
-        codesseg = [codeSegments('||', 'green', [A, B, C, D, A, E, F, G, H, E]), codeSegments('||', 'green', B, F, C, G, D, H)]
-        AB.color = 'green'
-        BC.color = 'green'
-        CD.color = 'green'
-        DA.color = 'green'
-        EF.color = 'green'
-        FG.color = 'green'
-        GH.color = 'green'
-        HE.color = 'green'
-        AE.color = 'green'
-        BF.color = 'green'
-        CG.color = 'green'
-        DH.color = 'green'
-      }
-      else {
-        codesseg = [codeSegments('||', 'green', A, B, C, D, E, F, G, H), codeSegments('O', 'red', A, E, B, F, C, G, D, H), codeSegments('×', 'blue', D, A, B, C, F, G, H, E)]
-        AB.color = 'green'
-        BC.color = 'blue'
-        CD.color = 'green'
-        DA.color = 'blue'
-        EF.color = 'green'
-        FG.color = 'blue'
-        GH.color = 'green'
-        HE.color = 'blue'
-        AE.color = 'red'
-        BF.color = 'red'
-        CG.color = 'red'
-        DH.color = 'red'
-      }
-
-      objets_correction.push(AB, BC, CD, DA, EF, FG, GH, HE, AE, BF, CG, DH, labelPoint(A, B, C, D, E, F, G, H),
-        g,
-        carreaux
-      );
-
-      if (liste_type_de_questions[i] < 3) correction += mathalea2d(params, objets_correction, codesseg);
-
-      if (this.liste_questions.indexOf(texte) == -1) {
-        // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(enonce + "<br>");
-        this.liste_corrections.push(correction + "<br>");
-        i++;
-      }
-      cpt++;
-    }
-
-    liste_de_question_to_contenu(this);
-  };
-  this.besoin_formulaire_numerique = ['Type de solides', 3, `1 : Cubes\n 2 : Pavés droits\n 3 : Mélange`]
-  this.besoin_formulaire2_numerique = [
-    "Type de cahier",
-    3,
-    `1 : Cahier à petits careaux\n 2 : Cahier à gros carreaux (Seyes)\n 3 : Feuille blanche`,
-  ];
-}
-
-/** 
- * * Calculer le produit de deux décimaux à partir d'un produit de deux entiers
- * * 6C30-2
- * @author Sébastien Lozano
- */
-
-function Produit_de_decimaux_a_partir_d_un_produit_connu() {
-  'use strict';
-  Exercice.call(this); // Héritage de la classe Exercice()
-  this.beta = false;
-  this.sup = 1;
-  if (this.beta) {
-    this.nb_questions = 3;
-  } else {
-    this.nb_questions = 3;
-  };
-
-  this.titre = "Calculer le produit de deux décimaux connaissant le produit de deux entiers";
-  this.consigne = ``;
-
-  this.nb_cols = 1;
-  this.nb_cols_corr = 1;
-  //this.nb_questions_modifiable = false;
-  sortie_html ? this.spacing = 2.5 : this.spacing = 1.5;
-  sortie_html ? this.spacing_corr = 2.5 : this.spacing_corr = 1.5;
-
-  let type_de_questions_disponibles;
-
-  this.nouvelle_version = function (numero_de_l_exercice) {
-    if (this.beta) {
-      type_de_questions_disponibles = [0, 1, 2];
-    } else {
-      //type_de_questions_disponibles = shuffle([choice([1,3]),choice([2,4]),0]);
-      type_de_questions_disponibles = shuffle([0, 1, 2]);
-
-    };
-
-    this.liste_questions = []; // Liste de questions
-    this.liste_corrections = []; // Liste de questions corrigées
-
-    //let liste_type_de_questions  = combinaison_listes(type_de_questions_disponibles,this.nb_questions) // Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-    let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions) // Tous les types de questions sont posées --> à remettre comme ci dessus		
-
-    for (let i = 0, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-
-      // pour les situations, autant de situations que de cas dans le switch !
-      let situations = [
-        {//case 0 --> (d1u1xp1)xd2u2
-          d1: randint(1, 9),
-          u1: randint(1, 9),
-          d2: randint(1, 9),
-          u2: randint(1, 9),
-          p1: randint(-3, 3, [0]),
-          p2: randint(-3, 3, [0]),
-        },
-      ];
-      let enonces = [];
-      //for (let k=0;k<3;k++) {
-      enonces.push({
-        enonce: `
-            Sachant que $${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)} = ${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)))}$,
-            calculer $${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (10 ** situations[0].p1)))}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)}$.
-					`,
-        question: ``,
-        correction: `
-					$${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (10 ** situations[0].p1)))}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)} = ${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${tex_nombrec(10 ** situations[0].p1)} \\times ${calcul(situations[0].d2 * 10 + situations[0].u2)} = ${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)}\\times ${tex_nombrec(10 ** situations[0].p1)} =  ${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)))}\\times ${tex_nombrec(10 ** situations[0].p1)} = ${tex_nombrec(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)) * calcul(10 ** situations[0].p1))}$
-					`
-      });
-      enonces.push({
-        enonce: `
-            Sachant que $${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)} = ${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)))}$,
-            calculer $${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1)))}\\times ${tex_nombre(calcul((situations[0].d2 * 10 + situations[0].u2) * (10 ** situations[0].p2)))}$.
-					`,
-        question: ``,
-        correction: `
-					$${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1)))}\\times ${tex_nombre(calcul((situations[0].d2 * 10 + situations[0].u2) * (10 ** situations[0].p2)))} = ${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)}\\times ${tex_nombrec(10 ** situations[0].p2)} = ${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)))}\\times ${tex_nombrec(10 ** situations[0].p2)} = ${tex_nombrec(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)) * calcul(10 ** situations[0].p2))}$
-					`
-      });
-      enonces.push({
-        enonce: `
-            Sachant que $${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)} = ${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)))}$,
-            calculer $${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (10 ** situations[0].p1)))}\\times ${tex_nombre(calcul((situations[0].d2 * 10 + situations[0].u2) * (10 ** situations[0].p2)))}$.
-					`,
-        question: ``,
-        correction: `
-					$${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (10 ** situations[0].p1)))}\\times ${tex_nombre(calcul((situations[0].d2 * 10 + situations[0].u2) * (10 ** situations[0].p2)))} = ${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${tex_nombrec(10 ** situations[0].p1)} \\times ${calcul(situations[0].d2 * 10 + situations[0].u2)}\\times ${tex_nombrec(10 ** situations[0].p2)} = ${calcul(situations[0].d1 * 10 + situations[0].u1)}\\times ${calcul(situations[0].d2 * 10 + situations[0].u2)}\\times ${tex_nombrec(10 ** situations[0].p1)}\\times ${tex_nombrec(10 ** situations[0].p2)} = ${tex_nombre(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)))}\\times ${tex_nombrec(10 ** situations[0].p1)}\\times ${tex_nombrec(10 ** situations[0].p2)} = ${tex_nombrec(calcul((situations[0].d1 * 10 + situations[0].u1) * (situations[0].d2 * 10 + situations[0].u2)) * calcul(10 ** situations[0].p1) * calcul(10 ** situations[0].p2))}$
-					`
-      });
-
-      //};
-
-      // autant de case que d'elements dans le tableau des situations
-      switch (liste_type_de_questions[i]) {
-        case 0:
-          texte = `${enonces[0].enonce}`;
-          if (this.beta) {
-            texte += `<br>`;
-            texte += `<br> =====CORRECTION======<br>${enonces[0].correction}`;
-            texte += `             `
-            texte_corr = ``;
-          } else {
-            texte_corr = `${enonces[0].correction}`;
-          };
-          break;
-        case 1:
-          texte = `${enonces[1].enonce}`;
-          if (this.beta) {
-            texte += `<br>`;
-            texte += `<br> =====CORRECTION======<br>${enonces[1].correction}`;
-            texte_corr = ``;
-          } else {
-            texte_corr = `${enonces[1].correction}`;
-          };
-          break;
-        case 2:
-          texte = `${enonces[2].enonce}`;
-          if (this.beta) {
-            texte += `<br>`;
-            texte += `<br> =====CORRECTION======<br>${enonces[2].correction}`;
-            texte_corr = ``;
-          } else {
-            texte_corr = `${enonces[2].correction}`;
-          };
-          break;
-      };
-      if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en crée une autre
-        this.liste_questions.push(texte);
-        this.liste_corrections.push(texte_corr);
-        i++;
-      }
-      cpt++;
-    }
-    liste_de_question_to_contenu(this);
-  }
-};
-
 
 /** 
  * * Encadrer_un_decimal_par_deux_entiers_consecutifs
@@ -13365,23 +8571,14 @@ function Arrondir_une_valeur() {
   this.besoin_formulaire_numerique = ['Type de nombre', 2, `1 : Nombre décimal\n 2 : Fraction`];
   this.besoin_formulaire2_case_a_cocher = ["Affichage de la valeur donnée à la calculatrice", false];
 }
-
 function Arrondir_une_valeur_6e() {
   Arrondir_une_valeur.call(this);
   this.sup = 1;
 }
 
-function Arrondir_une_valeur_5e() {
-  Arrondir_une_valeur.call(this);
-  this.sup = 2;
-  this.besoin_formulaire_numerique = ['Type de nombre', 2, `1 : Nombre décimal\n 2 : Fraction`];
-}
 
-function Arrondir_une_valeur_4e() {
-  this.sup = 3;
-  Arrondir_une_valeur.call(this);
-  this.besoin_formulaire_numerique = ['Type de nombre', 3, `1 : Nombre décimal\n 2 : Fraction\n 3 : Racine carrée`];
-}
+
+
 /** 
  * * Donner le chiffre des ... le nombre de ...
  * * 6N10-3
