+++ include/mathalea_exercices_4e.js
@@ -1,4 +1,15 @@
 /**
+ * 
+ * référence ?
+ */
+
+function Arrondir_une_valeur_4e() {
+  this.sup = 3;
+  Arrondir_une_valeur.call(this);
+  this.besoin_formulaire_numerique = ['Type de nombre', 3, `1 : Nombre décimal\n 2 : Fraction\n 3 : Racine carrée`];
+}
+
+/**
  * * Comparer des puissances de 10.
  *
  * Paramétrages possibles :
